### 2025-01-10 #7653 @editorialbot
[PRE REVIEW]: PyBOP: A Python package for battery model optimisation and parameterisation
**Submitting author:** @bradyplanden (<a href="http://orcid.org/0000-0002-1082-9125">Brady Planden</a>)
**Repository:** https://github.com/pybop-team/PyBOP
**Branch with paper.md** (empty if default branch): joss
**Version:** v24.12
