### 2025-01-11 718b3aa
feat: adds mean absolute error cost function
