### 2025-01-12 6ceb8ae
feat: simulated annealing optimiser
