### 2025-01-13 6bf717f
tests: adds simulated annealing integration tests
