### 2025-01-13 0446720
refactor: adds error_measure mixin, MSE cost function
