+++ pybop/costs/fitting_costs.py
@@ -14,12 +14,7 @@ class FittingCost(BaseCost):
     for evaluating model predictions against a set of data. The cost function
     quantifies the goodness-of-fit between the model predictions and the
     observed data, with a lower cost value indicating a better fit.
-
-    Inherits all parameters and attributes from ``BaseCost``.
     """
-
-    def __init__(self, problem):
-        super().__init__(problem)
 
     def compute(
         self,
@@ -65,13 +60,62 @@ class FittingCost(BaseCost):
             The residual difference between the model prediction and the target.
         dy : np.ndarray, optional
             The corresponding gradient with respect to the parameters for each signal.
+
+        Returns
+        -------
+        tuple or float
+            If dy is not None, returns a tuple containing the cost (float) and the
+            gradient (np.ndarray), otherwise returns only the computed cost (float).
         """
         raise NotImplementedError
 
 
-class RootMeanSquaredError(FittingCost):
+class ErrorMeasureMixin:
     """
-    Root mean square error cost function.
+    Mixin class providing reshaping to the error measure method.
+    """
+
+    def _format_output(
+        self, e: np.ndarray, dy: Optional[np.ndarray] = None
+    ) -> Union[float, tuple[float, np.ndarray]]:
+        """
+        Format output based on number of outputs and gradient presence.
+        """
+        if dy is not None:
+            return (
+                (e.item(), dy.flatten())
+                if self.n_outputs == 1
+                else (e.sum(), dy.sum(1))
+            )
+        return e.item() if self.n_outputs == 1 else np.sum(e)
+
+
+class MeanSquaredError(FittingCost, ErrorMeasureMixin):
+    """
+    Mean square error (MSE) cost function.
+
+    Computes the mean square error between model predictions and the target
+    data, providing a measure of the differences between predicted values and
+    observed values.
+    """
+
+    def _error_measure(
+        self,
+        r: np.ndarray,
+        dy: Optional[np.ndarray] = None,
+    ) -> Union[float, tuple[float, np.ndarray]]:
+        e = np.mean(np.abs(r) ** 2, axis=1)
+
+        if dy is not None:
+            de = 2 * np.mean((r * dy.T), axis=2)
+            return self._format_output(e, de)
+
+        return self._format_output(e)
+
+
+class RootMeanSquaredError(FittingCost, ErrorMeasureMixin):
+    """
+    Root mean square error (RMSE) cost function.
 
     Computes the root mean square error between model predictions and the target
     data, providing a measure of the differences between predicted values and
@@ -83,33 +127,13 @@ class RootMeanSquaredError(FittingCost):
         r: np.ndarray,
         dy: Optional[np.ndarray] = None,
     ) -> Union[float, tuple[float, np.ndarray]]:
-        """
-        Computes the cost function for the given predictions.
-
-        Parameters
-        ----------
-        r : np.ndarray
-            The residual difference between the model prediction and the target.
-        dy : np.ndarray, optional
-            The corresponding gradient with respect to the parameters for each signal.
-
-        Returns
-        -------
-        tuple or float
-            If dy is not None, returns a tuple containing the cost (float) and the
-            gradient (np.ndarray), otherwise returns only the computed cost (float).
-        """
         e = np.sqrt(np.mean(np.abs(r) ** 2, axis=1))
 
         if dy is not None:
             de = np.mean((r * dy.T), axis=2) / (e + np.finfo(float).eps)
-            return (
-                (e.item(), de.flatten())
-                if self.n_outputs == 1
-                else (e.sum(), de.sum(1))
-            )
+            return self._format_output(e, de)
 
-        return e.item() if self.n_outputs == 1 else np.sum(e)
+        return self._format_output(e)
 
 
[…]
