### 2025-01-13 9132c5a
fix: remove mixin and correct cost output shape calculation, update tests.
