+++ pybop/optimisers/_cuckoo.py
@@ -197,3 +197,13 @@ class CuckooSearchImpl(PopulationBasedOptimiser):
         Returns the name of the optimiser.
         """
         return "Cuckoo Search"
+
+    @property
+    def pa(self):
+        return self._pa
+
+    @pa.setter
+    def pa(self, pa):
+        if not isinstance(pa, (int, float)) or not 0 < pa <= 1:
+            raise ValueError("pa must be a numeric value between 0 and 1.")
+        self._pa = float(pa)
