### 2025-01-13 2952474
tests: adds MAE, MSE costs to eis integration.
