+++ pybop/optimisers/_cuckoo.py
@@ -204,6 +204,7 @@ class CuckooSearchImpl(PopulationBasedOptimiser):
 
     @pa.setter
     def pa(self, pa):
+        """Setter for abandonment rate"""
         if not isinstance(pa, (int, float)) or not 0 < pa <= 1:
             raise ValueError("pa must be a numeric value between 0 and 1.")
         self._pa = float(pa)
