### 2025-01-14 b050e07
Add dataset.get_subset
