### 2025-01-14 b550e3e
Add Dataset.get_subset (#624)
* Add dataset.get_subset

* Update domain setting
