### 2025-01-14 1461084
fix: correct optimisation bound kwarg, gradients for MSE, MAE, update test params
