### 2025-01-15 423113c
fix: missed TypeError match
