### 2025-01-15 66081a7
fix: restores gradient calculation for MSE, MAE
