+++ pybop/costs/fitting_costs.py
@@ -91,7 +91,7 @@ class MeanSquaredError(FittingCost):
         e = np.mean(np.abs(r) ** 2)
 
         if dy is not None:
-            de = 2 * np.sum((r * dy.T), axis=self.numpy_axis)
+            de = 2 * np.mean((r * dy.T), axis=self.numpy_axis)
             return e, de
 
         return e
@@ -138,7 +138,7 @@ class MeanAbsoluteError(FittingCost):
 
         if dy is not None:
             sign_r = np.sign(r)
-            de = np.sum(sign_r * dy.T, axis=self.numpy_axis)
+            de = np.mean(sign_r * dy.T, axis=self.numpy_axis)
             return e, de
 
         return e
