### 2025-01-15 13cfff9
Initial cost copy implementation
