### 2025-01-16 7c3b135
added examples for costs parameterized by fitting functions
