### 2025-01-16 fbe288e
added minimal testing file for EP-BOLFI integration
