### 2025-01-17 7059714
Apply suggestions from code review
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
