### 2025-01-17 5f49e88
refactor: updates OptimisationResult properties to return complete list, adds `x_best` method for best run parameter values
