### 2025-01-17 9d37972
fix: update multi_start_optimisation example
