### 2025-01-17 f210039
Fix most of the bugs created by resolving merge conflicts
