### 2025-01-17 600c01b
Add some checks in existing tests
