### 2025-01-17 4a44271
CHANGELOG updated, PyBOP-high-level.svg updated
