### 2025-01-21 8824173
adds properties for `_best` attributes
