### 2025-01-21 032f95e
adds assert for fisher property
