### 2025-01-23 3111a49
DigiBatt is added as the funding partner in the PyBOP high level image (#627)
* DigiBatt is added as the funding partner in the PyBOP high level image

* Changelog updated

* CHANGELOG updated, PyBOP-high-level.svg updated

* Changelog updated

* image update
