### 2025-01-23 0e2802a
refactor: clean-up sa cooling attrs, update hypers in integration test
