### 2025-01-23 bbcdbd4
tests: adds simulated annealing to optimisation unit tests
