### 2025-01-23 bee0fde
docs: update simulated annealing docstring
