+++ pybop/optimisers/pints_optimisers.py
@@ -20,14 +20,14 @@ class GradientDescent(BasePintsOptimiser):
     Implements a simple gradient descent optimisation algorithm.
 
     This class extends the gradient descent optimiser from the PINTS library, designed
-    to minimize a scalar function of one or more variables.
+    to minimise a scalar function of one or more variables.
 
     Note that this optimiser does not support boundary constraints.
 
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -100,7 +100,7 @@ class AdamW(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -170,7 +170,7 @@ class IRPropMin(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -240,7 +240,7 @@ class PSO(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -310,7 +310,7 @@ class SNES(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -380,7 +380,7 @@ class XNES(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -452,7 +452,7 @@ class NelderMead(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -522,7 +522,7 @@ class CMAES(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -597,7 +597,7 @@ class CuckooSearch(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -669,7 +669,7 @@ class RandomSearch(BasePintsOptimiser):
     Parameters
     ----------
     cost : callable
-        The cost function to be minimized.
+        The cost function to be minimised.
     max_iterations : int, optional
         Maximum number of iterations for the optimisation.
     min_iterations : int, optional (default=2)
@@ -730,11 +730,19 @@ class SimulatedAnnealing(BasePintsOptimiser):
     """
     Adapter for Simulated Annealing optimiser in PyBOP.
 
+    Simulated Annealing is a probabilistic optimisation algorithm inspired by the annealing
[…]
