### 2025-01-23 b878933
refactor: sa properties and add coverage
