### 2025-01-23 ed6f8b2
tests: fix missing sa unit test addition, small fixes to sa
