+++ pybop/optimisers/_simulated_annealing.py
@@ -92,7 +92,7 @@ class SimulatedAnnealingImpl(PintsOptimiser):
         accordingly.
         """
         if not self._ready_for_tell:
-            raise RuntimeError("ask() not called before tell()")
+            raise RuntimeError("ask() must be called before tell().")
         self._ready_for_tell = False
 
         # Unpack reply
