### 2025-01-23 f9090ac
docs: update README w/ new optimisers
