### 2025-01-23 107707b
tests: increase coverage
