### 2025-01-23 5cc5c4e
examples: small updates to sa example
