### 2025-01-23 d472553
Merge MultiOptimisationResult with OptimisationResullt (#601)
* Merge Multi into OptimisationResullt

* Update multi_start_optimisation.py

* Add test

* Add fisher as property

* Include n_evaluations

* refactor: updates OptimisationResult properties to return complete list, adds `x_best` method for best run parameter values

* fix: update multi_start_optimisation example

* adds properties for `_best` attributes

* adds assert for fisher property

* Update append to extend

* Update CHANGELOG.md

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github
[…]
