### 2025-01-24 c4f653d
docs: add recent cost additions
