### 2025-01-24 9df5000
Adds MAE and MSE cost classes (#621)
* feat: adds mean absolute error cost function

* refactor: adds error_measure mixin, MSE cost function

* fix: remove mixin and correct cost output shape calculation, update tests.

* tests: adds MAE, MSE costs to eis integration.

* fix: correct optimisation bound kwarg, gradients for MSE, MAE, update test params

* fix: remove redundant scipy bounds condition

* fix: missed TypeError match

* fix: restores gradient calculation for MSE, MAE
