+++ tests/integration/test_eis_parameterisation.py
@@ -57,6 +57,8 @@ class TestEISParameterisation:
         params=[
             pybop.GaussianLogLikelihood,
             pybop.SumSquaredError,
+            pybop.MeanAbsoluteError,
+            pybop.MeanSquaredError,
             pybop.Minkowski,
             pybop.LogPosterior,
         ]
