### 2025-01-24 f39ea84
refactor: updates default ECM OCV application
