+++ tests/unit/test_parameter_sets.py
@@ -26,7 +26,6 @@ class TestParameterSets:
             "Cell-jig heat transfer coefficient [W/K]": 10,
             "Jig thermal mass [J/K]": 500,
             "Jig-air heat transfer coefficient [W/K]": 10,
-            "Open-circuit voltage [V]": "default",
             "R0 [Ohm]": 0.001,
             "Element-1 initial overpotential [V]": 0,
             "Element-2 initial overpotential [V]": 0,
