### 2025-01-24 8c2e22f
Move ParameterValues construction to init
