+++ pybop/parameters/parameter_set.py
@@ -30,68 +30,71 @@ class ParameterSet:
     json_path : str, optional
         Path to a JSON file containing parameter data. If provided, parameters will be imported from this file during
         initialisation.
+    formation_concentrations : bool, optional
+            If True, re-calculates the initial concentrations of lithium in the active material (default: False).
     """
 
     def __init__(
         self,
         parameter_set: Union[str, dict, ParameterValues] = None,
         json_path: Optional[str] = None,
-        formation_concentrations: bool = False,
+        formation_concentrations: Optional[bool] = False,
     ):
         if parameter_set is not None and json_path is not None:
             raise ValueError(
                 "ParameterSet needs either a parameter_set or json_path as an input, not both."
             )
-        self.json_path = json_path
-        self.params = None
+        self._json_path = None
+        self.parameter_values = None
         self.chemistry = None
         self.formation_concentrations = formation_concentrations
 
-        if isinstance(parameter_set, str):
-            # Use class method
-            self.params = self.pybamm(parameter_set)
-        elif isinstance(parameter_set, dict):
-            # Keep as dictionary to allow "default" as value
-            self.params = parameter_set
-        elif isinstance(parameter_set, ParameterValues):
-            self.params = parameter_set
+        if json_path is not None:
+            self.import_parameters(json_path)
+        else:
+            self.parameter_values = self.convert_to_parameter_values(parameter_set)
 
-        if self.params is not None:
-            self.chemistry = self.params.get("chemistry", None)
+        if self.parameter_values is not None:
+            self.chemistry = self.parameter_values.get("chemistry", None)
 
-    def __call__(self, formation_concentrations: Optional[bool] = False):
+            if self.formation_concentrations:
+                set_formation_concentrations(self.parameter_values)
+
+    def convert_to_parameter_values(self, parameter_set):
         """
-        Obtain a copy of the parameter set as a PyBaMM ParameterValues object.
+        Converts a parameter set to a PyBaMM ParameterValues object.
+        """
+        if parameter_set is None:
+            return None
+        elif isinstance(parameter_set, str):
+            # Use class method
+            return self.pybamm(parameter_set)
+        elif isinstance(parameter_set, dict):
+            return ParameterValues(parameter_set)
+        elif isinstance(parameter_set, ParameterValues):
+            return parameter_set
+        else:
+            return parameter_set.parameter_values
 
-        Parameters
-        ----------
-        set_formation_concentrations : bool, optional
-            If True, re-calculates the initial concentrations of lithium in the active material (default: False).
+    def __call__(self):
+        """
+        Return the parameter set as a pybamm.ParameterValues object.
 
         Returns
         -------
         pybamm.ParameterValues
-            A PyBaMM parameter set corresponding to the parameter set stored in self.params.
+            The parameter set for a PyBaMM model.
         """
-        self.formation_concentrations = formation_concentrations
-
-        if self.params is None:
-            return None
-        elif isinstance(self.params, dict):
-            parameter_set = ParameterValues(self.params).copy()
-        else:
-            parameter_set = self.params.copy()
-
         if self.formation_concentrations:
-            set_formation_concentrations(parameter_set)
+            set_formation_concentrations(self.parameter_values)
 
-        return parameter_set
+        return self.parameter_values
 
     def __setitem__(self, key, value):
-        self.params[key] = value
+        self.parameter_values[key] = value
 
     def __getitem__(self, key):
-        return self.params[key]
+        return self.parameter_values[key]
 
     @staticmethod
     def evaluate_symbol(symbol: Union[Symbol, Number], params: dict):
@@ -124,7 +127,7 @@ class ParameterSet:
         """
         A list of parameter names
         """
-        return list(self.params.keys())
[…]
