+++ examples/scripts/comparison_examples/spm_scipymin.py
@@ -4,7 +4,7 @@ import pybamm
 import pybop
 
 # Define model
-parameter_set = pybop.ParameterSet.pybamm("Chen2020")
+parameter_set = pybop.ParameterSet("Chen2020")
 solver = pybamm.IDAKLUSolver()
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
 
