### 2025-01-24 5935386
FIM added for PyBOP cost functions
