+++ pybop/costs/_likelihoods.py
@@ -21,6 +21,33 @@ class BaseLikelihood(BaseCost):
         self.n_data = problem.n_data
         self.minimising = False
 
+    def observed_fisher(self, y: dict, dy: Optional[np.ndarray] = None) -> np.ndarray:
+        """
+        Compute the observed Fisher Information Matrix (FIM) for the given data.
+
+        The FIM is computed as the average of the squared gradients with respect to
+        the model parameters.
+
+        Parameters
+        ----------
+        y : dict
+            The model outputs for which to compute the FIM.
+        dy : np.ndarray, optional
+            The sensitivities (gradients) of the model outputs with respect to parameters.
+
+        Returns
+        -------
+        np.ndarray
+            The observed Fisher Information Matrix.
+        """
+        if dy is None:
+            _, dy = self.compute(y, dy=True)  # Compute gradients if not provided
+
+        # Calculate squared gradients element-wise
+        fisher_info = np.sum(np.square(dy), axis=0) / self.n_data
+
+        return fisher_info
+
 
 class BaseMetaLikelihood(BaseLikelihood):
     """
