### 2025-01-24 6bfdaab
Remove ParameterSet call
