### 2025-01-24 29b33df
Align ParameterSet with pybamm.ParameterValues (#593)
* Align ParameterSet with ParameterValues

* Update single_pulse_circuit_model.ipynb

* Increase coverage

* Add two more tests

* Apply suggestions from code review

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

* refactor: updates default ECM OCV application

* Move ParameterValues construction to init

* Deprecate ParameterSet.pybamm

* Update test_parameter_sets.py

* Add ParameterSet.to_pybamm

* Remove ParameterSet call

* Increase coverage

---------

Co-authored-by: Brady Planden <55357039+BradyPlan
[…]
