+++ examples/scripts/comparison_examples/maximum_a_posteriori.py
@@ -6,7 +6,7 @@ import pybop
 sigma = 0.002
 
 # Construct and update initial parameter values
-parameter_set = pybop.ParameterSet.pybamm("Chen2020")
+parameter_set = pybop.ParameterSet("Chen2020")
 parameter_set.update(
     {
         "Negative electrode active material volume fraction": 0.43,
