### 2025-01-24 8696cea
tests: updates integration tests, splits JaxCosts and loosen asserts as currently experimentally supported
