### 2025-01-27 51fc450
test: refactor pytest marking
