+++ tests/unit/test_models.py
@@ -15,6 +15,8 @@ class TestModels:
     A class to test the models.
     """
 
+    pytestmark = pytest.mark.unit
+
     @pytest.mark.parametrize(
         "model_class, expected_name, options",
         [
@@ -41,7 +43,6 @@ class TestModels:
             (pybop.empirical.Thevenin, "Equivalent Circuit Thevenin Model", None),
         ],
     )
-    @pytest.mark.unit
     def test_model_classes(self, model_class, expected_name, options):
         parameter_set = pybop.ParameterSet({"Nominal cell capacity [A.h]": 5.12})
         model = model_class(options=options, parameter_set=parameter_set)
@@ -66,7 +67,6 @@ class TestModels:
         model = request.param
         return model.copy()
 
-    @pytest.mark.unit
     def test_non_default_solver(self):
         solver = pybamm.CasadiSolver(
             mode="fast",
@@ -78,7 +78,6 @@ class TestModels:
         assert model.solver.atol == 1e-6
         assert model.solver.rtol == 1e-6
 
-    @pytest.mark.unit
     def test_predict_without_pybamm(self, model):
         model.pybamm_model = None
 
@@ -96,7 +95,6 @@ class TestModels:
             assert new_model.pybamm_model is not None
             assert new_model.parameter_set is not None
 
-    @pytest.mark.unit
     def test_predict_with_inputs(self, model):
         # Define inputs
         t_eval = np.linspace(0, 10, 100)
@@ -132,7 +130,6 @@ class TestModels:
         ):
             model.predict(inputs=inputs)
 
-    @pytest.mark.unit
     def test_predict_without_allow_infeasible_solutions(self, model):
         if isinstance(model, (pybop.lithium_ion.SPM, pybop.lithium_ion.SPMe)):
             model.allow_infeasible_solutions = False
@@ -147,7 +144,6 @@ class TestModels:
             ):
                 model.predict(t_eval=t_eval, inputs=inputs)
 
-    @pytest.mark.unit
     def test_build(self, model):
         if isinstance(model, pybop.lithium_ion.SPMe):
             model.build(initial_state={"Initial SoC": 1.0})
@@ -164,7 +160,6 @@ class TestModels:
             model.build()
             assert model.built_model is not None
 
-    @pytest.mark.unit
     def test_rebuild(self, model):
         model.build()
         initial_built_model = model._built_model
@@ -198,7 +193,6 @@ class TestModels:
                 initial_built_model, attribute
             )
 
-    @pytest.mark.unit
     def test_parameter_set_definition(self):
         # Test initilisation with different types of parameter set
         param_dict = {"Nominal cell capacity [A.h]": 5}
@@ -216,7 +210,6 @@ class TestModels:
         model = pybop.BaseModel(parameter_set=pybop_parameter_set)
         assert model.parameter_set == parameter_set
 
-    @pytest.mark.unit
     def test_rebuild_geometric_parameters(self):
         parameter_set = pybop.ParameterSet.pybamm("Chen2020")
         parameters = pybop.Parameters(
@@ -288,7 +281,6 @@ class TestModels:
     @pytest.mark.parametrize(
         "model_cls", [StandaloneDecay, pybop.ExponentialDecayModel]
     )
-    @pytest.mark.unit
     def test_reinit(self, model_cls):
         k = 0.1
         y0 = 1
@@ -316,7 +308,6 @@ class TestModels:
     @pytest.mark.parametrize(
         "model_cls", [StandaloneDecay, pybop.ExponentialDecayModel]
     )
-    @pytest.mark.unit
     def test_simulate(self, model_cls):
         k = 0.1
         y0 = 1
@@ -332,7 +323,6 @@ class TestModels:
[…]
