### 2025-01-27 844e780
Adds simulated annealing optimiser (#619)
* feat: simulated annealing optimiser

* tests: adds simulated annealing integration tests

* refactor: clean-up sa cooling attrs, update hypers in integration test

* tests: adds simulated annealing to optimisation unit tests

* docs: update simulated annealing docstring

* refactor: sa properties and add coverage

* tests: fix missing sa unit test addition, small fixes to sa

* docs: update README w/ new optimisers

* tests: increase coverage

* examples: small updates to sa example

* docs: add recent cost additions

* tests: updates to transformati
[…]
