+++ tests/integration/test_spm_parameterisations.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pybamm
 import pytest
 
 import pybop
@@ -29,20 +28,21 @@ class Test_SPM_Parameterisation:
                 "Positive electrode active material volume fraction": x[1],
             }
         )
-        solver = pybamm.IDAKLUSolver()
-        return pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
+        return pybop.lithium_ion.SPM(parameter_set=parameter_set)
 
     @pytest.fixture
     def parameters(self):
         return pybop.Parameters(
             pybop.Parameter(
                 "Negative electrode active material volume fraction",
-                prior=pybop.Uniform(0.4, 0.75),
-                bounds=[0.375, 0.775],
+                prior=pybop.Uniform(0.3, 0.9),
+                initial_value=pybop.Uniform(0.4, 0.75).rvs()[0],
+                bounds=[0.3, 0.8],
             ),
             pybop.Parameter(
                 "Positive electrode active material volume fraction",
-                prior=pybop.Uniform(0.4, 0.75),
+                prior=pybop.Uniform(0.3, 0.9),
+                initial_value=pybop.Uniform(0.4, 0.75).rvs()[0],
                 # no bounds
             ),
         )
@@ -62,11 +62,9 @@ class Test_SPM_Parameterisation:
             pybop.SumofPower,
             pybop.Minkowski,
             pybop.LogPosterior,
-            pybop.JaxSumSquaredError,
-            pybop.JaxLogNormalLikelihood,
         ]
     )
-    def cost(self, request):
+    def cost_cls(self, request):
         return request.param
 
     def noise(self, sigma, values):
@@ -75,6 +73,7 @@ class Test_SPM_Parameterisation:
     @pytest.fixture(
         params=[
             pybop.SciPyDifferentialEvolution,
+            pybop.SimulatedAnnealing,
             pybop.CuckooSearch,
             pybop.NelderMead,
             pybop.IRPropMin,
@@ -88,7 +87,7 @@ class Test_SPM_Parameterisation:
         return request.param
 
     @pytest.fixture
-    def optim(self, optimiser, model, parameters, cost, init_soc):
+    def optim(self, optimiser, model, parameters, cost_cls, init_soc):
         # Form dataset
         solution = self.get_data(model, init_soc)
         dataset = pybop.Dataset(
@@ -104,43 +103,42 @@ class Test_SPM_Parameterisation:
         problem = pybop.FittingProblem(model, parameters, dataset)
 
         # Construct the cost
-        if cost in [
-            pybop.GaussianLogLikelihoodKnownSigma,
-            pybop.JaxLogNormalLikelihood,
-        ]:
-            cost = cost(problem, sigma0=self.sigma0)
-        elif cost is pybop.GaussianLogLikelihood:
-            cost = cost(problem, sigma0=self.sigma0 * 4)  # Initial sigma0 guess
-        elif cost is pybop.LogPosterior:
-            cost = cost(
+        if cost_cls is pybop.GaussianLogLikelihoodKnownSigma:
+            cost = cost_cls(problem, sigma0=self.sigma0)
+        elif cost_cls is pybop.GaussianLogLikelihood:
+            cost = cost_cls(problem, sigma0=self.sigma0 * 4)  # Initial sigma0 guess
+        elif cost_cls is pybop.LogPosterior:
+            cost = cost_cls(
                 pybop.GaussianLogLikelihoodKnownSigma(problem, sigma0=self.sigma0)
             )
-        elif cost in [pybop.SumofPower, pybop.Minkowski]:
-            cost = cost(problem, p=2.5)
+        elif cost_cls in [pybop.SumofPower, pybop.Minkowski]:
+            cost = cost_cls(problem, p=2.5)
[…]
