### 2025-01-27 056ee6b
feat: adds iRProp+ optimiser
