+++ pybop/optimisers/_irpropplus.py
@@ -1,0 +1,200 @@
+import numpy as np
+from pints import Boundaries, RectangularBoundaries
+from pints import Optimiser as PintsOptimiser
+
+
+class IRPropPlusImpl(PintsOptimiser):
+    """
+    The iRprop+ algorithm adjusts step sizes based on the sign of the gradient
+    in each dimension, increasing step sizes when the sign remains consistent
+    and decreasing them when the sign changes.
+
+    References:
+    - [1] Igel and Hüsken (2003): Empirical Evaluation of Improved Rprop Algorithms.
+    - [2] Riedmiller and Braun (1993): A Direct Adaptive Method for Faster Backpropagation.
+    - [3] Igel and Husk (2003): Improving the Rprop Learning Algorithm.
+
+    Parameters
+    ----------
+    x0 : array-like
+        Initial starting point for the optimisation.
+    sigma0 : float or array-like, optional
+        Initial step size(s). If a scalar is provided, it is applied to all dimensions.
+        Default is 0.1.
+    boundaries : pints.Boundaries, optional
+        Boundary constraints for the optimisation. If None, no boundaries are applied.
+
+    Attributes
+    ----------
+    eta_min : float
+        Factor by which the step size is reduced when the gradient sign changes.
+        Default is 0.5.
+    eta_max : float
+        Factor by which the step size is increased when the gradient sign remains consistent.
+        Default is 1.2.
+    step_min : float
+        Minimum allowable step size. Default is 1e-3 * min(sigma0).
+    step_max : float or None
+        Maximum allowable step size. Default is None (unlimited).
+    """
+
+    def __init__(self, x0, sigma0=0.1, boundaries=None):
+        super().__init__(x0, sigma0, boundaries)
+
+        # Validate inputs
+        if len(self._sigma0) != len(x0):
+            raise ValueError(
+                "Length of `sigma0` must match the dimensionality of `x0`."
+            )
+        if boundaries is not None and not isinstance(boundaries, Boundaries):
+            raise ValueError(
+                "`boundaries` must be a `pints.Boundaries` object or None."
+            )
+
+        # Set hypers
+        self.eta_min = 0.5
+        self.eta_max = 1.2
+        self.step_min = 1e-4 * np.min(self._sigma0)
+        self.step_max = None
+
+        # Store the previous update for backtracking
+        self._update_previous = np.zeros_like(x0, dtype=float)
+
+        # Internal states
+        self._x_current = np.array(x0, dtype=float)
+        self._gradient_previous = None
+        self._step_sizes = np.copy(self._sigma0)
+        self._f_current = np.inf
+        self._x_best = np.array(x0, dtype=float)
+        self._f_best = np.inf
+        self._running = False
+        self._ready_for_tell = False
+
+        # Boundaries
+        self._boundaries = boundaries
+        if isinstance(boundaries, RectangularBoundaries):
+            self._lower = boundaries.lower()
+            self._upper = boundaries.upper()
+        else:
+            self._lower = self._upper = None
+
+        # Set proposals
+        self._proposed = np.copy(self._x_current)
+        self._proposed.setflags(write=False)
+
+    def ask(self):
+        """
+        Proposes the next point for evaluation.
+
+        Returns
+        -------
+        list
+            A list containing the proposed point.
+        """
+        if not self._running:
+            if self.step_min is not None and self.step_max is not None:
+                if self.step_min >= self.step_max:
+                    raise ValueError(
+                        f"Minimum step size ({self.step_min}) must be less than "
+                        f"maximum step size ({self.step_max})."
+                    )
+            self._running = True
+
+        self._ready_for_tell = True
+        return [self._proposed]
+
+    def tell(self, reply):
+        """
+        Updates the optimiser with the function value and gradient at the proposed point.
+
+        Parameters
+        ----------
[…]
