### 2025-01-27 3a2a44e
feat: adds global parameter sensitivity analysis, updates BaseCost to accept a list of proposals
