+++ pybop/costs/base_cost.py
@@ -1,6 +1,8 @@
 from typing import Optional, Union
 
 import numpy as np
+from SALib.analyze import sobol
+from SALib.sample.sobol import sample
 
 from pybop import BaseProblem
 from pybop._utils import add_spaces
@@ -61,75 +63,107 @@ class BaseCost:
 
     def __call__(
         self,
-        inputs: Union[Inputs, list],
+        inputs: Union[Inputs, list, np.ndarray],
         calculate_grad: bool = False,
         apply_transform: bool = False,
         for_optimiser: bool = False,
-    ) -> Union[float, tuple[float, np.ndarray]]:
+    ) -> Union[float, list, tuple[float, np.ndarray], list[tuple[float, np.ndarray]]]:
         """
-        This method calls the forward model via problem.evaluate(inputs),
-        and computes the cost for the given output by calling self.compute().
+        Compute cost and optional gradient for given input parameters.
 
         Parameters
         ----------
-        inputs : Inputs or list-like
-            The input parameters for which the cost and optionally the gradient
-            will be computed.
-        calculate_grad : bool, optional, default=False
+        inputs : Union[Inputs, list, np.ndarray]
+            Input parameters for cost computation. Supports single or multiple inputs.
+        calculate_grad : bool, default=False
             If True, both the cost and gradient will be computed. Otherwise, only the
             cost is computed.
-        apply_transform : bool, optional, default=False
+        apply_transform : bool, default=False
             If True, applies a transformation to the inputs before evaluating the model.
-        for_optimiser : bool, optional, default=False
-            If True, returns the cost value if self.minimising=True and the negative of
-            the cost value if self.minimising=False (i.e. the cost is being maximised).
+        for_optimiser : bool, default=False
+            If True, adjusts output sign based on minimisation/maximisation setting.
 
         Returns
         -------
-        float or tuple
-            - If `calculate_grad` is False, returns the computed cost (float).
-            - If `calculate_grad` is True, returns a tuple containing the cost (float)
-              and the gradient (np.ndarray).
-
-        Raises
-        ------
-        ValueError
-            If an error occurs during the calculation of the cost.
+        Union[float, list, tuple[float, np.ndarray], list[tuple[float, np.ndarray]]]
+            - Single input, no gradient: float
+            - Multiple inputs, no gradient: list[float]
+            - Single input with gradient: tuple[float, np.ndarray]
+            - Multiple inputs with gradient: list[tuple[float, np.ndarray]]
         """
-        # Note, we use the transformation and parameter properties here to enable
-        # differing attributes within the `LogPosterior` class
+        # Convert dict to list for sequential computations
+        if isinstance(inputs, dict):
+            inputs = list(inputs.values())
+
+        input_list = np.atleast_2d(inputs)
+        minimising = self.minimising or not for_optimiser
+        sign = 1 if minimising else -1
+
+        results = []
+        for input_val in input_list:
+            result = self._evaluate_single_input(
+                input_val,
+                calculate_grad=calculate_grad,
+                apply_transform=apply_transform,
+                sign=sign,
+            )
+            results.append(result)
+
+        return results[0] if len(input_list) == 1 else results
+
+    def _evaluate_single_input(
+        self,
+        input_value: Union[Inputs, np.ndarray],
+        calculate_grad: bool,
+        apply_transform: bool,
+        sign: int,
+    ) -> Union[float, tuple[float, np.ndarray]]:
+        """
+        Evaluate cost and optional gradient for a single input.
+        """
+        # Setup input transformation
[…]
