+++ pyproject.toml
@@ -26,10 +26,11 @@ classifiers = [
 ]
 requires-python = ">=3.9, <3.13"
 dependencies = [
-  "pybamm[jax]>=24.9",
-  "numpy>=1.16, <2.0",
-  "scipy>=1.3",
-  "pints>=0.5",
+    "pybamm[jax]>=24.9",
+    "numpy>=1.16, <2.0",
+    "scipy>=1.3",
+    "pints>=0.5",
+    "SALib>=1.5",
 ]
 
 [project.optional-dependencies]
