### 2025-01-27 66cc648
fix: missed example syntax, adds changelog entry
