+++ examples/scripts/comparison_examples/spm_CMAES.py
@@ -56,7 +56,7 @@ print("True parameters:", parameters.true_value())
 pybop.plot.dataset(dataset)
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.quick(problem, problem_inputs=results.x_best, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
