### 2025-01-27 672cdc2
Pytest for FIM calculation added
