### 2025-01-27 320cafb
fix pre-commit warnings & tests
