### 2025-01-27 155d4ca
Enable domain input
