+++ pybop/observers/unscented_kalman.py
@@ -32,8 +32,10 @@ class UnscentedKalmanFilterObserver(Observer):
         Dataset object containing the data to fit the model to.
     check_model : bool, optional
         Flag to indicate if the model should be checked (default: True).
-    signal: str
-        The signal to observe.
+    signal: list[str], optional
+        A list of variables to observe (default: ["Voltage [V]"]).
+    domain : str, optional
+        The name of the domain (default: "Time [s]").
     initial_state : dict, optional
         A valid initial state, e.g. the initial open-circuit voltage (default: None).
     """
@@ -50,6 +52,7 @@ class UnscentedKalmanFilterObserver(Observer):
         dataset: Optional[Dataset] = None,
         check_model: bool = True,
         signal: Optional[list[str]] = None,
+        domain: Optional[str] = None,
         additional_variables: Optional[list[str]] = None,
         initial_state: Optional[float] = None,
     ) -> None:
@@ -66,7 +69,13 @@ class UnscentedKalmanFilterObserver(Observer):
             )
 
         super().__init__(
-            parameters, model, check_model, signal, additional_variables, initial_state
+            parameters,
+            model,
+            check_model,
+            signal,
+            domain,
+            additional_variables,
+            initial_state,
         )
         if dataset is not None:
             # Check that the dataset contains necessary variables
