### 2025-01-27 65dd066
Apply suggestions from code review
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
