+++ tests/unit/test_likelihoods.py
@@ -217,24 +217,19 @@ class TestLikelihoods:
     def test_observed_fisher(self, one_signal_problem):
         likelihood = pybop.GaussianLogLikelihoodKnownSigma(one_signal_problem, sigma0=0.1)
         
-        # Create mock data and sensitivities
-        y = {"Voltage [V]": np.array([3.7, 3.6, 3.5])}
-        dy = np.array([[0.1, 0.2, 0.3]])
+        # Get the actual voltage data from the dataset
+        y = {"Voltage [V]": one_signal_problem.dataset["Voltage [V]"].data}
+
+        # Now let's test with a mock dy
+        n_data_points = len(y["Voltage [V]"])
+        mock_dy = np.random.rand(1, n_data_points)  # 1 parameter
         
-        # Compute observed Fisher Information Matrix
-        fim = likelihood.observed_fisher(y, dy)
+        # Compute observed Fisher Information Matrix with provided dy
+        fim_with_dy = likelihood.observed_fisher(y, mock_dy)
         
-        # Check that FIM is a 1x1 array (since we have one parameter)
-        assert fim.shape == (1, 1)
-        
-        # Check that FIM is positive (for a single parameter, it should be)
-        assert fim[0, 0] > 0
+        # Check that FIM is a 1x1 array
+        assert fim_with_dy.shape == (n_data_points,)
         
         # Check that FIM is computed correctly
-        expected_fim = np.sum(np.square(dy)) / likelihood.n_data
-        np.testing.assert_allclose(fim, expected_fim, rtol=1e-7)
-        
-        # Test without providing dy
-        fim_without_dy = likelihood.observed_fisher(y)
-        assert fim_without_dy.shape == (1, 1)
-        assert fim_without_dy[0, 0] > 0
+        expected_fim = np.sum(np.square(mock_dy), axis=0) / n_data_points
+        np.testing.assert_allclose(fim_with_dy, expected_fim, rtol=1e-7)
