### 2025-01-27 13004db
Changelog updated, example file updated to show FIM calculation
