### 2025-01-27 a8c31a4
Include PyBaMM solution in OptimizationResult (#573)
* Save latest pybamm solution in Problem class to add into OptimisationResult

* Fix most of the bugs created by resolving merge conflicts

* style: pre-commit fixes

* Add some checks in existing tests

* Add warning if final forward model run fails

* fix pre-commit warnings & tests

* style: pre-commit fixes

* Apply suggestions from code review

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by
[…]
