### 2025-01-27 062c7bf
fix: loosens rhat assertion w/ update configuration
