### 2025-01-28 02a2139
Updates to test suite (#641)
* test: refactor pytest marking

* fix: loosens rhat assertion w/ update configuration
