+++ tests/unit/test_optimisation.py
@@ -15,6 +15,8 @@ class TestOptimisation:
     A class to test the optimisation class.
     """
 
+    pytestmark = pytest.mark.unit
+
     @pytest.fixture
     def dataset(self):
         return pybop.Dataset(
@@ -92,7 +94,6 @@ class TestOptimisation:
             (pybop.SimulatedAnnealing, "Simulated Annealing", False),
         ],
     )
-    @pytest.mark.unit
     def test_optimiser_classes(
         self, two_param_cost, optimiser, expected_name, sensitivities
     ):
@@ -117,7 +118,6 @@ class TestOptimisation:
             if issubclass(optimiser, pybop.BasePintsOptimiser):
                 assert optim._boundaries is None
 
-    @pytest.mark.unit
     def test_no_optimisation_parameters(self, model, dataset):
         problem = pybop.FittingProblem(
             model=model, parameters=pybop.Parameters(), dataset=dataset
@@ -143,7 +143,6 @@ class TestOptimisation:
             pybop.SimulatedAnnealing,
         ],
     )
-    @pytest.mark.unit
     def test_optimiser_kwargs(self, cost, optimiser):
         def assert_log_update(optim):
             x_search = 0.7
@@ -376,13 +375,11 @@ class TestOptimisation:
             assert optim.x0 == x0_new
             assert optim.x0 != x0
 
-    @pytest.mark.unit
     def test_cuckoo_no_bounds(self, cost):
         optim = pybop.CuckooSearch(cost=cost, bounds=None, max_iterations=1)
         optim.run()
         assert optim.optimiser._boundaries is None
 
-    @pytest.mark.unit
     def test_randomsearch_bounds(self, two_param_cost):
         # Test clip_candidates with bound
         bounds = {"upper": [0.62, 0.57], "lower": [0.58, 0.53]}
@@ -402,7 +399,6 @@ class TestOptimisation:
         clipped_candidates = optimiser.optimiser.clip_candidates(candidates)
         assert np.allclose(clipped_candidates, candidates)
 
-    @pytest.mark.unit
     def test_randomsearch_ask_without_bounds(self, two_param_cost):
         # Initialize optimiser without boundaries
         optim = pybop.RandomSearch(
@@ -422,12 +418,10 @@ class TestOptimisation:
         assert np.all(candidates >= optim.optimiser._x0 - 6 * optim.optimiser._sigma0)
         assert np.all(candidates <= optim.optimiser._x0 + 6 * optim.optimiser._sigma0)
 
-    @pytest.mark.unit
     def test_adamw_impl_bounds(self):
         with pytest.warns(UserWarning, match="Boundaries ignored by AdamW"):
             pybop.AdamWImpl(x0=[0.1], boundaries=[0.0, 0.2])
 
-    @pytest.mark.unit
     def test_scipy_minimize_with_jac(self, cost):
         # Check a method that uses gradient information
         optim = pybop.SciPyMinimize(
@@ -442,7 +436,6 @@ class TestOptimisation:
         ):
             optim = pybop.SciPyMinimize(cost=cost, jac="Invalid string")
 
-    @pytest.mark.unit
     def test_scipy_minimize_invalid_x0(self, cost):
         # Check a starting point that returns an infinite cost
         invalid_x0 = np.array([1.1])
@@ -455,7 +448,6 @@ class TestOptimisation:
         optim.run()
         assert abs(optim._cost0) != np.inf
 
-    @pytest.mark.unit
     def test_single_parameter(self, cost):
         # Test catch for optimisers that can only run with multiple parameters
         with pytest.raises(
@@ -464,7 +456,6 @@ class TestOptimisation:
         ):
             pybop.CMAES(cost=cost)
 
-    @pytest.mark.unit
     def test_invalid_cost(self):
[…]
