+++ tests/unit/test_problem.py
@@ -10,6 +10,8 @@ class TestProblem:
     A class to test the problem class.
     """
 
+    pytestmark = pytest.mark.unit
+
     @pytest.fixture
     def model(self):
         return pybop.lithium_ion.SPM()
@@ -65,7 +67,6 @@ class TestProblem:
     def signal(self):
         return "Voltage [V]"
 
-    @pytest.mark.unit
     def test_base_problem(self, parameters, model, dataset):
         # Construct Problem
         problem = pybop.BaseProblem(parameters, model=model)
@@ -104,7 +105,6 @@ class TestProblem:
                 parameters=[parameter_list[0], "Invalid string"]
             )
 
-    @pytest.mark.unit
     def test_fitting_problem(self, parameters, dataset, model, signal):
         with pytest.warns(UserWarning) as record:
             problem = pybop.FittingProblem(
@@ -184,7 +184,6 @@ class TestProblem:
         with pytest.raises(ValueError):
             pybop.FittingProblem(model, parameters, bad_dataset, signal=two_signals)
 
-    @pytest.mark.unit
     def test_fitting_problem_eis(self, parameters):
         model = pybop.lithium_ion.SPM(eis=True)
 
@@ -212,7 +211,6 @@ class TestProblem:
         out = problem.evaluate(inputs=[0.0, 0.0])
         assert not np.isfinite(out["Impedance"])
 
-    @pytest.mark.unit
     def test_multi_fitting_problem(self, model, parameters, dataset, signal):
         problem_1 = pybop.FittingProblem(model, parameters, dataset, signal=signal)
 
@@ -255,7 +253,6 @@ class TestProblem:
         )
         assert len(combined_problem.solution) == 2
 
-    @pytest.mark.unit
     def test_design_problem(self, parameters, experiment, model):
         with pytest.warns(UserWarning) as record:
             problem = pybop.DesignProblem(
@@ -300,7 +297,6 @@ class TestProblem:
         problem = pybop.DesignProblem(model, pybop.Parameters(), experiment)
         assert problem.initial_state == {"Initial SoC": 0.8}
 
-    @pytest.mark.unit
     def test_problem_construct_with_model_predict(
         self, parameters, model, dataset, signal
     ):
