### 2025-01-28 f656517
fixed transformations and cost calls in EP-BOLFI implementation
