### 2025-01-28 9434289
refactor: add property for pybamm solution
