### 2025-01-28 8484bd9
made DataSet instances copyable
