### 2025-01-28 2bf16b1
Create ocp_average
