### 2025-01-28 6c7c989
made base_echem children instances copyable
