### 2025-01-28 be2f192
fix: add missed test
