+++ pybop/__init__.py
@@ -196,7 +196,6 @@ from ._classification import classify_using_hessian
 # Applications
 #
 from .applications.ocp_average import ocp_average
-from .applications.gitt_ocv_extract import gitt_ocv_extract
 
 #
 # Plotting classes
