### 2025-01-28 da6aa4f
made DataSet instances pickleable
