### 2025-01-28 f4d1222
made base_echem children instances pickleable
