### 2025-01-28 70df2bf
tests: add direct tests for OptimisationResult "best" properties
