### 2025-01-28 6a24894
Add stoichiometric_fit
