### 2025-01-28 2cd6d0e
Create test_applications.py
