+++ pybop/plot/problem.py
@@ -24,7 +24,7 @@ def quick(problem, problem_inputs: Inputs = None, show=True, **layout_kwargs):
     **layout_kwargs : optional
             Valid Plotly layout keys and their values,
             e.g. `xaxis_title="Time / s"` or
-            `xaxis={"title": "Time / s", "titlefont_size": 18}`.
+            `xaxis={"title": "Time / s", "title": {font={"size":14}}}`.
 
     Returns
     -------
