+++ pybop/plot/parameters.py
@@ -15,7 +15,7 @@ def parameters(optim, show=True, **layout_kwargs):
     **layout_kwargs : optional
         Valid Plotly layout keys and their values,
         e.g. `xaxis_title="Time [s]"` or
-        `xaxis={"title": "Time [s]", "titlefont_size": 18}`.
+        `xaxis={"title": "Time [s]", "title": {font={"size":14}}}`.
 
     Returns
     -------
