### 2025-01-29 34d12b2
docs: updates to plotting optional args
