+++ pybop/plot/standard_plots.py
@@ -368,7 +368,7 @@ def trajectories(x, y, trace_names=None, show=True, **layout_kwargs):
     **layout_kwargs : optional
             Valid Plotly layout keys and their values,
             e.g. `xaxis_title="Time / s"` or
-            `xaxis={"title": "Time / s", "title": {font={"size":14}}}`.
+            `xaxis={"title": "Time [s]", font={"size":14}}`
 
     Returns
     -------
