+++ pybop/plot/convergence.py
@@ -14,7 +14,7 @@ def convergence(optim, show=True, **layout_kwargs):
     **layout_kwargs : optional
         Valid Plotly layout keys and their values,
         e.g. `xaxis_title="Time [s]"` or
-        `xaxis={"title": "Time [s]", "title": {font={"size":14}}}`.
+        `xaxis={"title": "Time [s]", font={"size":14}}`
 
     Returns
     ---------
