### 2025-01-29 1b93bc5
Fixes for Plotly v6 (#646)
* fix: plotly v6 updates

* docs: updates to plotting optional args
