### 2025-01-29 f9f729b
tests: adds unit tests for sensitivity analysis
