+++ tests/unit/test_cost.py
@@ -400,3 +400,20 @@ class TestCosts:
             match="All problems must be of the same class type.",
         ):
             pybop.WeightedCost(cost1, cost2)
+
+    def test_parameter_sensitivities(self, problem):
+        cost = pybop.MeanAbsoluteError(problem)
+        result = cost.sensitivity_analysis(4)
+
+        # Assertions
+        assert isinstance(result, dict)
+        assert "S1" in result
+        assert "ST" in result
+        assert isinstance(result["S1"], np.ndarray)
+        assert isinstance(result["S2"], np.ndarray)
+        assert isinstance(result["ST"], np.ndarray)
+        assert isinstance(result["S1_conf"], np.ndarray)
+        assert isinstance(result["ST_conf"], np.ndarray)
+        assert isinstance(result["S2_conf"], np.ndarray)
+        assert result["S1"].shape == (1,)
+        assert result["ST"].shape == (1,)
