### 2025-01-29 61ac28c
tests: adds explicit dictionary type cost call
