### 2025-01-30 3273346
docs: update docstrings
