### 2025-01-30 b8d77b9
Update Weppner & Huggins model
