+++ pybop/models/lithium_ion/weppner_huggins.py
@@ -5,6 +5,8 @@ from pybamm import DummySolver, Parameter, ParameterValues, citations
 from pybamm import lithium_ion as pybamm_lithium_ion
 from pybamm import t as pybamm_t
 
+from pybop import ParameterSet
+
 
 class BaseWeppnerHuggins(pybamm_lithium_ion.BaseModel):
     """
@@ -20,25 +22,28 @@ class BaseWeppnerHuggins(pybamm_lithium_ion.BaseModel):
         Valid PyBaMM model option keys and their values, for example:
         parameter_set : pybamm.ParameterValues or dict, optional
             The parameters for the model. If None, default parameters provided by PyBaMM are used.
-        geometry : dict, optional
-            The geometry definitions for the model. If None, default geometry from PyBaMM is used.
-        submesh_types : dict, optional
-            The types of submeshes to use. If None, default submesh types from PyBaMM are used.
-        var_pts : dict, optional
-            The discretization points for each variable in the model. If None, default points from PyBaMM are used.
-        spatial_methods : dict, optional
-            The spatial methods used for discretization. If None, default spatial methods from PyBaMM are used.
-        solver : pybamm.Solver, optional
-            The solver to use for simulating the model. If None, the default solver from PyBaMM is used.
+        options : dict, optional
+            A dictionary of options to customise the behaviour of the PyBaMM model.
     """
 
     def __init__(self, name="Weppner & Huggins model", **model_kwargs):
-        # Model kwargs (build, options) are not implemented, keeping here for consistent interface
-        if model_kwargs is not dict(build=True):
+        unused_keys = []
+        for key in model_kwargs.keys():
+            if key not in ["build", "parameter_set", "options"]:
+                unused_keys.append(key)
+        options = {"working electrode": "positive"}
+        if model_kwargs.get("options", None) is not None:
+            for key, value in model_kwargs["options"].items():
+                if key in ["working electrode"]:
+                    options[key] = value
+                else:
+                    unused_keys.append("options[" + key + "]")
+        if any(unused_keys):
             unused_kwargs_warning = "The input model_kwargs are not currently used by the Weppner & Huggins model."
             warnings.warn(unused_kwargs_warning, UserWarning, stacklevel=2)
 
-        super().__init__({}, name)
+        super().__init__(options=options, name=name, build=True)
+        self._summary_variables = []
 
         citations.register("""
             @article{Weppner1977,
@@ -54,52 +59,40 @@ class BaseWeppnerHuggins(pybamm_lithium_ion.BaseModel):
             }
         """)
 
-        # `self.param` is a class containing all the relevant parameters and functions for
-        # this model. These are purely symbolic at this stage, and will be set by the
-        # `ParameterValues` class when the model is processed.
-        self.options["working electrode"] = "positive"
-        self._summary_variables = []
-
-        t = pybamm_t
         ######################
         # Parameters
         ######################
+        # Parameters are purely symbolic at this stage, and will be set by the
+        # `ParameterValues` class when the model is processed.
 
-        d_s = Parameter("Positive particle diffusivity [m2.s-1]")
+        Q_th = Parameter("Theoretical electrode capacity [A.s]")
 
-        c_s_max = Parameter("Maximum concentration in positive electrode [mol.m-3]")
-
-        i_app = self.param.current_density_with_time
-
-        U = Parameter("Reference OCP [V]")
-
+        U = Parameter("Reference voltage [V]")
         U_prime = Parameter("Derivative of the OCP wrt stoichiometry [V]")
[…]
