### 2025-01-30 e543f05
corrected EP-BOLFI output for plots, added citations
