+++ pybop/experimental/_ep_bolfi.py
@@ -1,7 +1,10 @@
+import copy
 import json
+import time
 
 import ep_bolfi
 import numpy as np
+from pybamm import citations
 
 from pybop import BaseOptimiser
 from pybop.experimental.base_bayes_optimiser import BayesianOptimisationResult
@@ -23,6 +26,36 @@ class EP_BOLFI(BaseOptimiser):
     Only compatible with MultivariateParameters with
     MultivariateGaussian prior and an ExpectationPropagationCost.
     """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        citations.register("""@article{
+            Minka2013,
+            title={{Expectation Propagation for approximate Bayesian inference}},
+            author={Minka, T},
+            journal={Proceedings of the Seventeenth Conference on Uncertainty in Artificial Intelligence (UAI2001)},
+            pages={362-369},
+            year={2013}
+        }""")
+        citations.register("""@article{
+            Gutmann2016,
+            title={{Bayesian optimization for likelihood-free inference of simulator-based statistical models}},
+            author={Gutmann, M and Corander, J},
+            journal={Journal of Machine Learning Research},
+            volume={17},
+            pages={1-47},
+            year={2016}
+        }""")
+        citations.register("""@article{
+            Kuhn2022,
+            title={{Bayesian Parameterization of Continuum Battery Models from Featurized Electrochemical Measurements Considering Noise}},
+            author={Kuhn, Y and Wolf, H and Latz, A and Horstmann, B},
+            journal={Batteries & Supercaps},
+            volume={6},
+            pages={e202200374},
+            year={2023},
+            publisher={Chemistry Europe}
+        }""")
 
     def model_wrapper(self, inputs):
         evaluation = self.cost.problem.model.predict(
@@ -67,7 +100,6 @@ class EP_BOLFI(BaseOptimiser):
         self.r_features = self.unset_options.pop("r_features", None)
         # Read in live feedback options.
         self.show_trials = self.unset_options.pop("show_trials", None)
-        self.verbose = self.unset_options.pop("verbose", None)
         # Read in auxiliary EP-BOLFI settings.
         self.gelman_rubin_threshold = self.unset_options.pop(
             "gelman_rubin_threshold", None
@@ -124,6 +156,7 @@ class EP_BOLFI(BaseOptimiser):
         )
 
     def _run(self):
+        start = time.time()
         # bolfi_posterior is the full GPy object containing the state at
         # the end of the last feature iteration, while the
         # MultivariateGaussian is a slight approximation.
@@ -144,24 +177,50 @@ class EP_BOLFI(BaseOptimiser):
             self.model_resampling_increase,
             self.independent_mcmc_chains,
             self.scramble_ep_feature_order,
+            False,  # normalize_features
             self.show_trials,
             self.verbose,
             self.seed,
         )
+        end = time.time()
         ep_bolfi_result = json.loads(self.optimiser.result_to_json(seed=self.seed))
-        len_log = len(
-            self.optimiser.log_of_tried_parameters[
-                list(self.parameters.param.values())[0]
-            ]
-        )
-        transposed_log = [[] for _ in range(len_log)]
-        for log in self.optimiser.log_of_tried_parameters.values():
-            for j, l in enumerate(log):
-                transposed_log[j].append(l)
+        ep_bolfi_log = json.loads(self.optimiser.log_to_json())
+        x_list = np.array(list(ep_bolfi_log["tried parameters"].values())).T
+        # Collect all features into one cost. Note: they are logarithms,
+        # so this is a multiplicative combination.
+        feature_costs = np.array(list(ep_bolfi_log["discrepancies"].values()))
+        cost_list = copy.deepcopy(feature_costs[0])
[…]
