### 2025-01-30 82372c1
Create SPDiffusion model
