+++ examples/scripts/battery_parameterisation/gitt_pulse.py
@@ -1,13 +1,13 @@
 import numpy as np
 
 import pybop
-from pybop.models.lithium_ion.weppner_huggins import (
+from pybop.models.lithium_ion.basic_SP_diffusion import (
     convert_physical_to_electrode_parameters,
 )
 
 # Define model
 parameter_set = pybop.ParameterSet("Xu2019")
-model = pybop.lithium_ion.SPM(
+model = pybop.lithium_ion.SPMe(
     parameter_set=parameter_set, options={"working electrode": "positive"}
 )
 
@@ -35,49 +35,31 @@ dataset = pybop.Dataset(
         "Voltage [V]": corrupt_values,
     }
 )
-pulse_index = np.flatnonzero(dataset["Current function [A]"])
 
 # Define parameter set
-parameter_set = convert_physical_to_electrode_parameters(parameter_set, "positive")
-parameter_set.update(
-    {
-        "Reference voltage [V]": dataset["Voltage [V]"][pulse_index[0]],
-        "Derivative of the OCP wrt stoichiometry [V]": (
-            (dataset["Voltage [V]"][-1] - dataset["Voltage [V]"][0])
-            / (
-                (
-                    dataset["Discharge capacity [A.h]"][-1]
-                    - dataset["Discharge capacity [A.h]"][0]
-                )
-                / (parameter_set["Theoretical electrode capacity [A.s]"] / 3600)
-            )
-        ),
-    },
+parameter_set = convert_physical_to_electrode_parameters(
+    model._unprocessed_parameter_set, "positive"
 )
 
 # Fitting parameters
 parameters = pybop.Parameters(
     pybop.Parameter(
         "Particle diffusion time scale [s]",
-        prior=pybop.Gaussian(6000, 1000),
-        bounds=[2000, 10000],
+        prior=pybop.Gaussian(2000, 1000),
     ),
     pybop.Parameter(
-        "Reference voltage [V]",
-        initial_value=parameter_set["Reference voltage [V]"],
+        "Series resistance [Ohm]",
+        initial_value=parameter_set["Series resistance [Ohm]"],
     ),
 )
 
 # Define the cost to optimise
-model = pybop.lithium_ion.WeppnerHuggins(parameter_set=parameter_set)
-dataset = dataset.get_subset(pulse_index)
+model = pybop.lithium_ion.SPDiffusion(parameter_set=parameter_set)
 problem = pybop.FittingProblem(model, parameters, dataset)
 cost = pybop.RootMeanSquaredError(problem)
 
-# Build the optimisation problem
+# Build and run the optimisation problem
 optim = pybop.SciPyMinimize(cost=cost)
-
-# Run the optimisation problem
 results = optim.run()
 
 # Plot the timeseries output
