+++ pybop/models/lithium_ion/basic_SP_diffusion.py
@@ -1,0 +1,322 @@
+import warnings
+
+import pybamm
+from pybamm import (
+    Event,
+    FunctionParameter,
+    Parameter,
+    ParameterValues,
+    PrimaryBroadcast,
+    Scalar,
+    SpatialVariable,
+    Variable,
+)
+from pybamm import lithium_ion as pybamm_lithium_ion
+from pybamm import t as pybamm_t
+from pybamm.input.parameters.lithium_ion.Xu2019 import (
+    nmc_ocp_Xu2019,
+)
+
+from pybop import ParameterSet
+
+
+class BaseSPDiffusion(pybamm_lithium_ion.BaseModel):
+    """
+    Diffusion model for a single particle representing a half-cell for GITT.
+
+    This model can be used with PyBOP through the `pybop.SPDiffusion` class.
+
+    Parameters
+    ----------
+    name : str, optional
+        The name of the model.
+    **model_kwargs : optional
+        Valid PyBaMM model option keys and their values, for example:
+        parameter_set : pybamm.ParameterValues or dict, optional
+            The parameters for the model. If None, default parameters provided by PyBaMM are used.
+        geometry : dict, optional
+            The geometry definitions for the model. If None, default geometry from PyBaMM is used.
+        submesh_types : dict, optional
+            The types of submeshes to use. If None, default submesh types from PyBaMM are used.
+        var_pts : dict, optional
+            The discretization points for each variable in the model. If None, default points from PyBaMM are used.
+        spatial_methods : dict, optional
+            The spatial methods used for discretization. If None, default spatial methods from PyBaMM are used.
+        solver : pybamm.Solver, optional
+            The solver to use for simulating the model. If None, the default solver from PyBaMM is used.
+        build : bool, optional
+            If True, the model is built upon creation (default: False).
+        options : dict, optional
+            A dictionary of options to customise the behaviour of the PyBaMM model.
+    """
+
+    def __init__(self, name="Single Particle Diffusion Model", **model_kwargs):
+        unused_keys = []
+        for key in model_kwargs.keys():
+            if key not in ["build", "parameter_set", "options"]:
+                unused_keys.append(key)
+        if model_kwargs.get("build", True) is False:
+            unused_keys.append("build")
+        options = {"working electrode": "positive"}
+        if model_kwargs.get("options", None) is not None:
+            for key, value in model_kwargs["options"].items():
+                if key in ["working electrode"]:
+                    options[key] = value
+                else:
+                    unused_keys.append("options[" + key + "]")
+        if any(unused_keys):
+            unused_kwargs_warning = "The input model_kwargs are not currently used by the SP Diffusion Model."
+            warnings.warn(unused_kwargs_warning, UserWarning, stacklevel=2)
+
+        super().__init__(options=options, name=name, build=True)
+
+        # Unpack model options
+        electrode = self.options["working electrode"]
+        if electrode not in ["positive", "negative"]:
+            raise ValueError(
+                f"Unrecognised electrode type: {electrode}, "
+                'expecting either "positive" or "negative".'
+            )
+
+        pybamm.citations.register("Xu2019")  # for the OCP
+
+        ######################
+        # Variables
+        ######################
+        # Variables that depend on time only are created without a domain
+        Q = Variable("Discharge capacity [A.h]")
+        Qt = Variable("Throughput capacity [A.h]")
+
+        # Variables that vary spatially are created with a domain
[…]
