### 2025-01-30 72005d3
Add SPDiffusion to model tests
