+++ tests/unit/test_models.py
@@ -30,6 +30,7 @@ class TestModels:
                 {"number of MSMR reactions": ("6", "4")},
             ),
             (pybop.lithium_ion.WeppnerHuggins, "Weppner & Huggins Model", None),
+            (pybop.lithium_ion.SPDiffusion, "Single Particle Diffusion Model", None),
             (
                 pybop.lithium_ion.GroupedSPMe,
                 "Grouped Single Particle Model with Electrolyte",
@@ -58,6 +59,7 @@ class TestModels:
             pybop.lithium_ion.MPM(),
             pybop.lithium_ion.MSMR(options={"number of MSMR reactions": ("6", "4")}),
             pybop.lithium_ion.WeppnerHuggins(),
+            pybop.lithium_ion.SPDiffusion(),
             pybop.lithium_ion.GroupedSPMe(),
             pybop.lithium_ion.GroupedSPMe(options={"surface form": "differential"}),
             pybop.empirical.Thevenin(),
@@ -103,7 +105,9 @@ class TestModels:
                 "Negative electrode relative porosity": 0.52,
                 "Positive electrode relative porosity": 0.63,
             }
-        elif isinstance(model, (pybop.lithium_ion.WeppnerHuggins)):
+        elif isinstance(
+            model, (pybop.lithium_ion.WeppnerHuggins, pybop.lithium_ion.SPDiffusion)
+        ):
             inputs = {
                 "Theoretical electrode capacity [A.s]": 5000,
             }
