+++ pybop/models/lithium_ion/basic_SP_diffusion.py
@@ -65,7 +65,7 @@ class BaseSPDiffusion(pybamm_lithium_ion.BaseModel):
                 else:
                     unused_keys.append("options[" + key + "]")
         if any(unused_keys):
-            unused_kwargs_warning = "The input model_kwargs are not currently used by the SP Diffusion Model."
+            unused_kwargs_warning = f"The input model_kwargs {unused_keys} are not currently used by the SP Diffusion Model."
             warnings.warn(unused_kwargs_warning, UserWarning, stacklevel=2)
 
         super().__init__(options=options, name=name, build=True)
