### 2025-01-30 2fc63bd
Update CHANGELOG.md
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
