### 2025-01-30 1c5b0f0
Update pybop/costs/_likelihoods.py
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
