### 2025-01-30 0aa55b3
added predictive posterior visualization, showcases parallel evaluation workaround
