+++ pybop/experimental/test_ep_bolfi.py
@@ -1,7 +1,10 @@
+import copy
+
 import numpy as np
 from _ep_bolfi import EP_BOLFI
 from multivariate_parameters import MultivariateParameters
 from multivariate_priors import MultivariateGaussian
+from plot_bayes import bayes
 from pybamm import print_citations
 
 import pybop
@@ -31,25 +34,27 @@ unknowns = MultivariateParameters(
     prior=MultivariateGaussian([np.log(original_diffusivity)], [[np.log(10)]]),
 )
 
-problem = pybop.FittingProblem(model, unknowns, dataset, signal=["Voltage [V]"])
-cost = pybop.WeightedCost(pybop.SumSquaredError(problem))
-# Override the forced Parameters class in BaseCost instantiation.
-cost.parameters = unknowns
-optim = EP_BOLFI(
-    cost,
-    t_eval=t_eval,
-    ep_iterations=1,
-    ep_dampener=0,
-    bolfi_initial_evidence=15,
-    bolfi_total_evidence=30,
-    bolfi_posterior_samples=20,
-    verbose=True,
-)
+if __name__ == "__main__":
+    problem = pybop.FittingProblem(model, unknowns, dataset, signal=["Voltage [V]"])
+    pickleable_problem = copy.deepcopy(problem)
+    cost = pybop.WeightedCost(pybop.SumSquaredError(problem))
+    # Override the forced Parameters class in BaseCost instantiation.
+    cost.parameters = unknowns
+    optim = EP_BOLFI(
+        cost,
+        t_eval=t_eval,
+        ep_iterations=1,
+        ep_dampener=0,
+        bolfi_initial_evidence=15,
+        bolfi_total_evidence=30,
+        bolfi_posterior_samples=20,
+        verbose=True,
+    )
 
-results = optim.run()
+    results = optim.run()
+    bayes(copy.deepcopy(pickleable_problem), results)
+    pybop.plot.quick(copy.deepcopy(pickleable_problem), problem_inputs=results.x)
+    pybop.plot.convergence(optim, yaxis_type="log")
+    pybop.plot.parameters(optim, yaxis_type="log")
 
-pybop.plot.quick(problem, problem_inputs=results.x)
-pybop.plot.convergence(optim, yaxis_type="log")
-pybop.plot.parameters(optim, yaxis_type="log")
-
-print_citations()
+    print_citations()
