### 2025-01-30 fd2b3f3
Create GITTFit with example
