+++ pybop/applications/gitt_methods.py
@@ -1,5 +1,7 @@
 from typing import Optional
 
+import numpy as np
+
 import pybop
 from pybop import BaseApplication
 
@@ -46,7 +48,9 @@ class GITTPulseFit(BaseApplication):
 
         # Define the cost to optimise
         self.parameter_set = parameter_set
-        self.model = pybop.lithium_ion.SPDiffusion(parameter_set=self.parameter_set)
+        self.model = pybop.lithium_ion.SPDiffusion(
+            parameter_set=self.parameter_set, build=True
+        )
         self.problem = pybop.FittingProblem(self.model, self.parameters, gitt_pulse)
         self.cost = cost(self.problem)
 
@@ -54,3 +58,83 @@ class GITTPulseFit(BaseApplication):
         self.optim = optimiser(cost=self.cost, verbose=verbose)
         self.results = self.optim.run()
         self.parameter_set.update(self.parameters.as_dict(self.results.x))
+
+
+class GITTFit(BaseApplication):
+    """
+    Fit the diffusion timescale of each pulse from a galvanostatic intermittent
+    titration technique (GITT) measurement.
+
+    Parameters
+    ----------
+    gitt_dataset : pybop.Dataset
+        A dataset containing the "Time [s]", "Current [A]" and "Voltage [V]" for one
+        pulse obtained from a GITT measurement.
+    gitt_pulse_index : list[np.ndarray]
+        A nested list of integers representing the indices of each pulse in the dataset.
+    parameter_set : pybop.ParameterSet
+        A parameter set containing values for the parameters of the SPDiffusion model.
+    cost : pybop.BaseCost, optional
+        The cost function to quantify the error (default: pybop.RootMeanSquaredError).
+    optimiser : pybop.BaseOptimiser, optional
+        The optimisation algorithm to use (default: pybop.SciPyMinimize).
+    verbose : bool, optional
+        If True, progress messages are printed (default: False).
+    """
+
+    def __init__(
+        self,
+        gitt_dataset: pybop.Dataset,
+        gitt_pulse_index: list[np.ndarray],
+        parameter_set: pybop.ParameterSet,
+        cost: Optional[pybop.BaseCost] = pybop.RootMeanSquaredError,
+        optimiser: Optional[pybop.BaseOptimiser] = pybop.SciPyMinimize,
+        verbose: bool = False,
+    ):
+        # Preallocate outputs
+        self.pulses = []
+        stoichiometry = []
+        diffusion_time = []
+        series_resistance = []
+
+        for pulse_index in gitt_pulse_index:
+            # Estimate the parameters for this pulse
+            self.pulses.append(
+                pybop.GITTPulseFit(
+                    gitt_pulse=gitt_dataset.get_subset(pulse_index),
+                    parameter_set=parameter_set,
+                    cost=cost,
+                    optimiser=optimiser,
+                    verbose=verbose,
+                )
+            )
+
+            # Log and update the parameter estimates for the next iteration
+            parameter_set.update(
+                {
+                    "Particle diffusion time scale [s]": self.pulses[-1].results.x[0],
+                    "Series resistance [Ohm]": self.pulses[-1].results.x[1],
+                }
+            )
+            diffusion_time.append(parameter_set["Particle diffusion time scale [s]"])
+            series_resistance.append(parameter_set["Series resistance [Ohm]"])
+            stoichiometry.append(parameter_set["Initial stoichiometry"])
+            parameter_set.update(
+                {
+                    "Initial stoichiometry": stoichiometry[-1]
+                    + (
+                        gitt_dataset["Discharge capacity [A.h]"][pulse_index[-1] + 1]
+                        - gitt_dataset["Discharge capacity [A.h]"][pulse_index[0] - 1]
[…]
