+++ examples/scripts/battery_parameterisation/gitt_fitting.py
@@ -1,10 +1,10 @@
 import numpy as np
+from pybamm import Interpolant
 
 import pybop
 from pybop.models.lithium_ion.basic_SP_diffusion import (
     convert_physical_to_electrode_parameters,
 )
-from pybamm import Interpolant
 
 # Define model
 parameter_set = pybop.ParameterSet("Xu2019")
@@ -18,7 +18,8 @@ initial_state = {"Initial SoC": 0.9}
 experiment = pybop.Experiment(
     [
         "Rest for 1 second",
-        ("Discharge at 1C for 10 minutes (10 second period)", "Rest for 20 minutes") * 8,
+        ("Discharge at 1C for 10 minutes (10 second period)", "Rest for 20 minutes")
+        * 8,
     ]
 )
 values = model.predict(initial_state=initial_state, experiment=experiment)
@@ -38,7 +39,11 @@ dataset = pybop.Dataset(
 
 # Determine the indices corresponding to each pulse in the dataset
 nonzero_index = np.concatenate(
-    ([-1], np.flatnonzero(dataset["Current function [A]"]), [len(dataset["Current function [A]"])+1])
+    (
+        [-1],
+        np.flatnonzero(dataset["Current function [A]"]),
+        [len(dataset["Current function [A]"]) + 1],
+    )
 )
 nonzero_gaps = np.extract(
     nonzero_index[1:] - nonzero_index[:-1] != 1,  # check if there is a gap
@@ -46,7 +51,7 @@ nonzero_gaps = np.extract(
 )
 pulse_index = []
 for start, finish in zip(nonzero_gaps[:-1], nonzero_gaps[1:]):
-    pulse_index.append([i for i in nonzero_index if i >= start and i<finish])
+    pulse_index.append([i for i in nonzero_index if i >= start and i < finish])
 
 # Define parameter set
 parameter_set = convert_physical_to_electrode_parameters(
@@ -57,14 +62,17 @@ parameter_set = convert_physical_to_electrode_parameters(
 gitt_fit = pybop.GITTFit(dataset, pulse_index, parameter_set)
 
 # Plot the parameters
-pybop.plot.dataset(gitt_fit.parameter_data, signal=["Particle diffusion time scale [s]"])
+pybop.plot.dataset(
+    gitt_fit.parameter_data, signal=["Particle diffusion time scale [s]"]
+)
 pybop.plot.dataset(gitt_fit.parameter_data, signal=["Series resistance [Ohm]"])
 
 # Replace the diffusivity in the original model
 parameter_set = pybop.ParameterSet("Xu2019")
 
+
 def diffusivity(sto, T):
-    return parameter_set["Positive particle radius [m]"]**2 / Interpolant(
+    return parameter_set["Positive particle radius [m]"] ** 2 / Interpolant(
         gitt_fit.parameter_data["Stoichiometry"],
         gitt_fit.parameter_data["Particle diffusion time scale [s]"],
         sto,
@@ -74,7 +82,9 @@ def diffusivity(sto, T):
 
 # Compare the original and identified model predictions
 parameter_set.update({"Positive particle diffusivity [m2.s-1]": diffusivity})
-values = model.predict(initial_state=initial_state, experiment=experiment, parameter_set=parameter_set)
+values = model.predict(
+    initial_state=initial_state, experiment=experiment, parameter_set=parameter_set
+)
 pybop.plot.trajectories(
     values["Time [s]"].data,
     [dataset["Voltage [V]"], values["Voltage [V]"].data],
