### 2025-01-30 caf028b
Update stoichiometry_fitting.py
