### 2025-01-31 21921e7
added switch to all problem classes to enable parallelization workaround
