### 2025-01-31 ced4cb5
rewrote parallelization in EP-BOLFI example and plotting with global parallelization fix
