+++ pybop/experimental/test_ep_bolfi.py
@@ -1,5 +1,3 @@
-import copy
-
 import numpy as np
 from _ep_bolfi import EP_BOLFI
 from multivariate_parameters import MultivariateParameters
@@ -35,8 +33,9 @@ unknowns = MultivariateParameters(
 )
 
 if __name__ == "__main__":
-    problem = pybop.FittingProblem(model, unknowns, dataset, signal=["Voltage [V]"])
-    pickleable_problem = copy.deepcopy(problem)
+    problem = pybop.FittingProblem(
+        model, unknowns, dataset, signal=["Voltage [V]"], parallelizable=True
+    )
     cost = pybop.WeightedCost(pybop.SumSquaredError(problem))
     # Override the forced Parameters class in BaseCost instantiation.
     cost.parameters = unknowns
@@ -52,8 +51,9 @@ if __name__ == "__main__":
     )
 
     results = optim.run()
-    bayes(copy.deepcopy(pickleable_problem), results)
-    pybop.plot.quick(copy.deepcopy(pickleable_problem), problem_inputs=results.x)
+
+    bayes(problem, results)
+    pybop.plot.quick(problem, problem_inputs=results.x)
     pybop.plot.convergence(optim, yaxis_type="log")
     pybop.plot.parameters(optim, yaxis_type="log")
 
