### 2025-01-31 0095edf
removes redundant checks implemented in parent class, adds coverage
