### 2025-01-31 e184cfe
merged upstream changes to model-parallelization with parallezation fix to problem classes
