+++ tests/integration/test_classification.py
@@ -1,0 +1,173 @@
+import numpy as np
+import pytest
+from pybamm import Parameter
+
+import pybop
+
+
+class TestClassification:
+    """
+    A class to test the classification of different optimisation results.
+    """
+
+    pytestmark = pytest.mark.integration
+
+    @pytest.fixture(
+        params=[
+            np.asarray([0.05, 0.05]),
+            np.asarray([0.1, 0.05]),
+            np.asarray([0.05, 0.01]),
+        ]
+    )
+    def parameters(self, request):
+        ground_truth = request.param
+        return pybop.Parameters(
+            pybop.Parameter(
+                "R0 [Ohm]",
+                prior=pybop.Gaussian(0.05, 0.01),
+                bounds=[0.02, 0.08],
+                true_value=ground_truth[0],
+            ),
+            pybop.Parameter(
+                "R1 [Ohm]",
+                prior=pybop.Gaussian(0.05, 0.01),
+                bounds=[0.02, 0.08],
+                true_value=ground_truth[1],
+            ),
+        )
+
+    @pytest.fixture
+    def parameter_set(self):
+        parameter_set = pybop.ParameterSet(
+            json_path="examples/parameters/initial_ecm_parameters.json"
+        )
+        parameter_set.update({"C1 [F]": 1000})
+        return parameter_set
+
+    @pytest.fixture
+    def model(self, parameter_set, parameters):
+        parameter_set.update(parameters.as_dict(parameters.true_value()))
+        return pybop.empirical.Thevenin(parameter_set=parameter_set)
+
+    @pytest.fixture
+    def dataset(self, model):
+        experiment = pybop.Experiment(
+            [
+                "Discharge at 0.5C for 2 minutes (4 seconds period)",
+                "Charge at 0.5C for 2 minutes (4 seconds period)",
+            ]
+        )
+        solution = model.predict(experiment=experiment)
+        return pybop.Dataset(
+            {
+                "Time [s]": solution["Time [s]"].data,
+                "Current function [A]": solution["Current [A]"].data,
+                "Voltage [V]": solution["Voltage [V]"].data,
+            }
+        )
+
+    @pytest.fixture
+    def problem(self, model, parameters, dataset):
+        return pybop.FittingProblem(model, parameters, dataset)
+
+    def test_classify_using_hessian(self, problem):
+        cost = pybop.RootMeanSquaredError(problem)
+        x = cost.parameters.true_value()
+        bounds = cost.parameters.get_bounds()
+        x0 = np.clip(x, bounds["lower"], bounds["upper"])
+        optim = pybop.Optimisation(cost=cost)
+        results = pybop.OptimisationResult(x=x0, optim=optim)
+
+        if np.all(x == np.asarray([0.05, 0.05])):
+            message = pybop.classify_using_hessian(results)
+            assert message == "The optimiser has located a minimum."
+        elif np.all(x == np.asarray([0.1, 0.05])):
+            message = pybop.classify_using_hessian(results)
+            assert message == (
+                "The optimiser has not converged to a stationary point."
+                " The result is near the upper bound of R0 [Ohm]."
+            )
+        elif np.all(x == np.asarray([0.05, 0.01])):
+            message = pybop.classify_using_hessian(results)
+            assert message == (
+                "The optimiser has not converged to a stationary point."
+                " The result is near the lower bound of R1 [Ohm]."
+            )
+        else:
+            raise Exception(f"Please add a check for these values: {x}")
+
+        if np.all(x == np.asarray([0.05, 0.05])):
+            cost = pybop.GaussianLogLikelihoodKnownSigma(problem, sigma0=0.002)
+            optim = pybop.Optimisation(cost=cost)
[…]
