+++ examples/scripts/comparison_examples/spm_descent.py
@@ -5,18 +5,18 @@ import pybop
 
 # Define model and use high-performant solver for sensitivities
 solver = pybamm.IDAKLUSolver()
-parameter_set = pybop.ParameterSet.pybamm("Chen2020")
+parameter_set = pybop.ParameterSet("Chen2020")
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
 
 # Fitting parameters
 parameters = pybop.Parameters(
     pybop.Parameter(
         "Negative electrode active material volume fraction",
-        prior=pybop.Gaussian(0.6, 0.01),
+        prior=pybop.Gaussian(0.6, 0.1),
     ),
     pybop.Parameter(
         "Positive electrode active material volume fraction",
-        prior=pybop.Gaussian(0.6, 0.01),
+        prior=pybop.Gaussian(0.6, 0.1),
     ),
 )
 
