### 2025-01-31 13c7d2e
observed_fisher method updated
