+++ pybop/costs/_likelihoods.py
@@ -5,7 +5,7 @@ import scipy.stats as stats
 
 import pybop
 from pybop.costs.base_cost import BaseCost
-from pybop.parameters.parameter import Parameter, Parameters, Inputs
+from pybop.parameters.parameter import Inputs, Parameter, Parameters
 from pybop.parameters.priors import BasePrior, JointLogPrior, Uniform
 from pybop.problems.base_problem import BaseProblem
 
