### 2025-01-31 251388e
fixed problem class initialization with parallelization workaround
