### 2025-01-31 d046b37
feat: adds verbose printing for pints' optimisers
