### 2025-01-31 7a9a216
refactor: adds evaluations, changelog entry
