### 2025-01-31 c249118
fix: avoids potential multistart infeasible condition
