### 2025-01-31 4cb16e5
Add test_gitt_pulse_fit
