+++ tests/unit/test_likelihoods.py
@@ -227,11 +227,15 @@ class TestLikelihoods:
         mock_dy = np.random.rand(1, n_data_points)  # 1 parameter
 
         # Compute observed Fisher Information Matrix with provided dy
-        fim_with_dy = likelihood.observed_fisher(y, mock_dy)
+        inputs = [0.5]
+        fim_with_dy = likelihood.observed_fisher(y, mock_dy, inputs)
 
         # Check that FIM is a 1x1 array
-        assert fim_with_dy.shape == (n_data_points,)
+        assert fim_with_dy.shape == (1,1)
 
         # Check that FIM is computed correctly
-        expected_fim = np.sum(np.square(mock_dy), axis=0) / n_data_points
+        _, grad = likelihood.__call__(inputs, calculate_grad=True)
+        shaped_grad = grad.reshape(-1, 1)
+        expected_fim = (shaped_grad @ shaped_grad.T) / n_data_points
+        
         np.testing.assert_allclose(fim_with_dy, expected_fim, rtol=1e-7)
