+++ tests/unit/test_likelihoods.py
@@ -231,11 +231,11 @@ class TestLikelihoods:
         fim_with_dy = likelihood.observed_fisher(y, mock_dy, inputs)
 
         # Check that FIM is a 1x1 array
-        assert fim_with_dy.shape == (1,1)
+        assert fim_with_dy.shape == (1, 1)
 
         # Check that FIM is computed correctly
         _, grad = likelihood.__call__(inputs, calculate_grad=True)
         shaped_grad = grad.reshape(-1, 1)
         expected_fim = (shaped_grad @ shaped_grad.T) / n_data_points
-        
+
         np.testing.assert_allclose(fim_with_dy, expected_fim, rtol=1e-7)
