### 2025-01-31 a00b399
Remove missed check after refactor, add to docstring
