### 2025-02-01 38f2d96
Include parameters set method pybamm for data
