+++ examples/scripts/comparison_examples/selecting_a_solver.py
@@ -6,7 +6,7 @@ import pybamm
 import pybop
 
 # Parameter set and model definition
-parameter_set = pybop.ParameterSet("Chen2020")
+parameter_set = pybop.ParameterSet().pybamm("Chen2020")
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
 
 solvers = [
