+++ examples/scripts/getting_started/multi_fitting.py
@@ -3,7 +3,7 @@ import numpy as np
 import pybop
 
 # Parameter set and model definition
-parameter_set = pybop.ParameterSet("Chen2020")
+parameter_set = pybop.ParameterSet().pybamm("Chen2020")
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
 
 # Create initial SOC, experiment objects
