+++ examples/scripts/comparison_examples/grouped_SPMe.py
@@ -12,7 +12,7 @@ layout_options = dict(
 plot_dict = pybop.plot.StandardPlot(layout_options=layout_options)
 
 # Unpack parameter values from Chen2020
-parameter_set = pybop.ParameterSet("Chen2020")
+parameter_set = pybop.ParameterSet().pybamm("Chen2020")
 
 # Fix the electrolyte diffusivity and conductivity
 ce0 = parameter_set["Initial concentration in electrolyte [mol.m-3]"]
