+++ examples/scripts/getting_started/multi_start_optimisation.py
@@ -5,7 +5,7 @@ import pybop
 
 # Define model and use the high-performant IDAKLU solver for sensitivities
 solver = pybamm.IDAKLUSolver(atol=1e-5, rtol=1e-5)
-parameter_set = pybop.ParameterSet("Chen2020")
+parameter_set = pybop.ParameterSet().pybamm("Chen2020")
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
 
 # Fitting parameters
