### 2025-02-02 f8145fb
Restore standalone tests
