### 2025-02-02 1a7ce81
Restore standalone tests (#652)
* Restore standalone tests

* Add pybamm_solution property to all base classes

* Increase coverage
