+++ tests/integration/test_applications.py
@@ -90,7 +90,7 @@ class TestApplications:
         )
 
         # Estimate the stoichiometry corresponding to the GITT-OCV
-        ocv_fit = pybop.StoichiometricFit(
+        ocv_fit = pybop.OCPCapacityToStoichiometry(
             ocv_dataset=ocv_dataset,
             ocv_function=ocv_function,
         )
