### 2025-02-03 01dc689
adds logic to skip fisher calculation for non-gradient models
