+++ pybop/optimisers/base_optimiser.py
@@ -333,6 +333,10 @@ class BaseOptimiser:
         return self._needs_sensitivities
 
     @property
+    def transformation(self):
+        return self._transformation
+
+    @property
     def minimising(self):
         return self._minimising
 
@@ -369,7 +373,7 @@ class OptimisationResult:
         self.optim = optim
         self.cost = self.optim.cost
         self.minimising = self.optim.minimising
-        self._transformation = self.optim._transformation  # noqa: SLF001
+        self._transformation = self.optim.transformation
         self.n_runs = 0
         self._best_run = None
         self._x = []
@@ -398,16 +402,16 @@ class OptimisationResult:
             )
 
             # Calculate Fisher Information if Likelihood
-            fisher = (
-                np.diagonal(self.cost.observed_fisher(x))
-                if isinstance(self.cost, BaseLikelihood)
-                else None
-            )
+            if isinstance(self.cost, BaseLikelihood):
+                fisher = self.cost.observed_fisher(x)
+                diag_fish = np.diag(fisher) if fisher is not None else None
+            else:
+                diag_fish = None
 
             self._extend(
                 x=[x],
                 final_cost=[final_cost],
-                fisher=[fisher],
+                fisher=[diag_fish],
                 n_iterations=[n_iterations],
                 n_evaluations=[n_evaluations],
                 time=[time],
