+++ tests/unit/test_likelihoods.py
@@ -29,10 +29,12 @@ class TestLikelihoods:
 
     @pytest.fixture
     def parameters(self):
-        return pybop.Parameter(
-            "Negative electrode active material volume fraction",
-            prior=pybop.Gaussian(0.5, 0.01),
-            bounds=[0.375, 0.625],
+        return pybop.Parameters(
+            pybop.Parameter(
+                "Negative electrode active material volume fraction",
+                prior=pybop.Gaussian(0.5, 0.01),
+                bounds=[0.375, 0.625],
+            )
         )
 
     @pytest.fixture
@@ -214,7 +216,23 @@ class TestLikelihoods:
             pybop.JaxGaussianLogLikelihoodKnownSigma,
         ],
     )
-    def test_fisher_matrix(self, likelihood_cls, one_signal_problem):
+    def test_fisher_matrix(
+        self, likelihood_cls, one_signal_problem, model, dataset, parameters
+    ):
         likelihood = likelihood_cls(one_signal_problem, sigma0=1e-3)
-        fisher = likelihood.observed_fisher([0.5, 0.03])
+        fisher = likelihood.observed_fisher([0.5])
         assert isinstance(fisher, np.ndarray)
+
+        # Test fisher does not compute for non-gradient available parameters
+
+        parameters.add(
+            pybop.Parameter(
+                "Negative particle radius [m]",
+                prior=pybop.Gaussian(6e-06, 0.1e-6),
+                bounds=[1e-6, 9e-6],
+            ),
+        )
+        problem = pybop.FittingProblem(model, parameters, dataset)
+        likelihood_non_grad = likelihood_cls(problem, sigma0=1e-3)
+        fisher = likelihood_non_grad.observed_fisher([0.5, 5e-06])
+        assert fisher is None
