### 2025-02-03 4f17323
examples: clean up missed items
