+++ examples/scripts/comparison_examples/maximum_likelihood.py
@@ -28,7 +28,7 @@ parameters = pybop.Parameters(
 )
 
 # Generate data
-sigma = 0.001
+sigma = 0.002
 experiment = pybop.Experiment(
     [
         (
@@ -60,7 +60,7 @@ signal = ["Voltage [V]", "Bulk open-circuit voltage [V]"]
 # Generate problem, cost function, and optimisation class
 problem = pybop.FittingProblem(model, parameters, dataset, signal=signal)
 likelihood = pybop.GaussianLogLikelihood(problem, sigma0=sigma * 4)
-optim = pybop.CMAES(
+optim = pybop.IRPropMin(
     likelihood,
     max_unchanged_iterations=20,
     min_iterations=20,
