### 2025-02-03 d5343e1
FIM added for PyBOP likelihood functions (#632)
* FIM added for PyBOP cost functions

* Pytest for FIM calculation added

* pytets updates

* style: pre-commit fixes

* Changelog updated, example file updated to show FIM calculation

* example file updated

* style: pre-commit fixes

* Update CHANGELOG.md

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

* Update pybop/costs/_likelihoods.py

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

* style: pre-commit fixes

* observed_fisher method updated

* style: pre-commit fixes

* unit
[…]
