### 2025-02-03 004f8d1
restores examples, updates to tell() comments
