### 2025-02-03 9850c88
Adds global parameter sensitivity analysis (#635)
* feat: adds global parameter sensitivity analysis, updates BaseCost to accept a list of proposals

* fix: missed example syntax, adds changelog entry

* tests: adds unit tests for sensitivity analysis

* tests: adds explicit dictionary type cost call

* docs: update docstrings

* fix: suggestion from review
