+++ examples/scripts/comparison_examples/scipy_minimize.py
@@ -37,8 +37,15 @@ dataset = pybop.Dataset(
 
 # Generate problem, cost function, and optimisation class
 problem = pybop.FittingProblem(model, parameters, dataset)
-cost = pybop.JaxSumSquaredError(problem)
-optim = pybop.SciPyMinimize(cost, max_iterations=100, method="L-BFGS-B", jac=True)
+cost = pybop.SumSquaredError(problem)
+optim = pybop.SciPyMinimize(
+    cost,
+    max_iterations=100,
+    multistart=1,
+    method="L-BFGS-B",
+    jac=True,
+    n_sensitivity_samples=256,
+)
 
 results = optim.run()
 
