+++ examples/scripts/comparison_examples/covariance_matrix_adaptation.py
@@ -44,7 +44,9 @@ signal = ["Voltage [V]", "Bulk open-circuit voltage [V]"]
 # Generate problem, cost function, and optimisation class
 problem = pybop.FittingProblem(model, parameters, dataset, signal=signal)
 cost = pybop.SumSquaredError(problem)
-optim = pybop.CMAES(cost, sigma0=0.25, max_unchanged_iterations=10, max_iterations=40)
+optim = pybop.CMAES(
+    cost, sigma0=0.25, max_unchanged_iterations=10, max_iterations=40, multistart=2
+)
 
 # Run the optimisation
 results = optim.run()
@@ -54,7 +56,7 @@ print("True parameters:", parameters.true_value())
 pybop.plot.dataset(dataset)
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.quick(problem, problem_inputs=results.x_best, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
