### 2025-02-03 08db668
tests: updates parameter range to align with default pybamm parameter object
