### 2025-02-03 08f41f7
tests: updates priors for x0 within bound clipping for RandomSearch test
