### 2025-02-03 d21c6c7
Align simulation options
