### 2025-02-04 5b8d899
Update geometry and copying
