### 2025-02-04 163ea4f
docs: adds additional information to the iRProp+ docstring
