### 2025-02-04 5e19d0f
iRProp+ Optimiser (#636)
* feat: adds iRProp+ optimiser

* removes redundant checks implemented in parent class, adds coverage

* fix: avoids potential multistart infeasible condition

* restores examples, updates to tell() comments

* tests: updates parameter range to align with default pybamm parameter object

* tests: updates priors for x0 within bound clipping for RandomSearch test

* docs: adds additional information to the iRProp+ docstring

* adds changelog entry
