+++ tests/unit/test_optimisation.py
@@ -31,8 +31,8 @@ class TestOptimisation:
     def one_parameter(self):
         return pybop.Parameter(
             "Positive electrode active material volume fraction",
-            prior=pybop.Gaussian(0.6, 0.02),
-            bounds=[0.58, 0.62],
+            prior=pybop.Gaussian(0.5, 0.02),
+            bounds=[0.48, 0.52],
         )
 
     @pytest.fixture
@@ -40,13 +40,13 @@ class TestOptimisation:
         return pybop.Parameters(
             pybop.Parameter(
                 "Negative electrode active material volume fraction",
-                prior=pybop.Gaussian(0.6, 0.2),
+                prior=pybop.Gaussian(0.6, 0.02),
                 bounds=[0.58, 0.62],
             ),
             pybop.Parameter(
                 "Positive electrode active material volume fraction",
-                prior=pybop.Gaussian(0.55, 0.05),
-                bounds=[0.53, 0.57],
+                prior=pybop.Gaussian(0.5, 0.05),
+                bounds=[0.48, 0.52],
             ),
         )
 
@@ -89,6 +89,7 @@ class TestOptimisation:
             (pybop.XNES, "Exponential Natural Evolution Strategy (xNES)", False),
             (pybop.PSO, "Particle Swarm Optimisation (PSO)", False),
             (pybop.IRPropMin, "iRprop-", True),
+            (pybop.IRPropPlus, "iRprop+", True),
             (pybop.NelderMead, "Nelder-Mead", False),
             (pybop.RandomSearch, "Random Search", False),
             (pybop.SimulatedAnnealing, "Simulated Annealing", False),
@@ -137,6 +138,7 @@ class TestOptimisation:
             pybop.XNES,
             pybop.PSO,
             pybop.IRPropMin,
+            pybop.IRPropPlus,
             pybop.NelderMead,
             pybop.CuckooSearch,
             pybop.RandomSearch,
@@ -189,7 +191,7 @@ class TestOptimisation:
         check_incorrect_update(optim)
 
         # Test multistart
-        multistart_optim = optimiser(cost, multistart=2, max_iterations=6)
+        multistart_optim = optimiser(cost, max_iterations=6, multistart=2)
         check_multistart(multistart_optim, 6, 2)
 
         if optimiser in [pybop.GradientDescent, pybop.AdamW, pybop.NelderMead]:
@@ -201,7 +203,7 @@ class TestOptimisation:
                 optim, {"upper": [np.inf], "lower": [0.57]}, should_raise=True
             )
         else:
-            bounds = {"upper": [0.63], "lower": [0.57]}
+            bounds = {"upper": [0.53], "lower": [0.47]}
             check_bounds_handling(optim, cost_bounds)
             optim = optimiser(cost=cost, bounds=bounds)
             assert optim.bounds == bounds
@@ -282,6 +284,7 @@ class TestOptimisation:
 
         if optimiser in [
             pybop.AdamW,
+            pybop.IRPropPlus,
             pybop.CuckooSearch,
             pybop.GradientDescent,
             pybop.RandomSearch,
@@ -370,7 +373,7 @@ class TestOptimisation:
         else:
             x0 = cost.parameters.initial_value()
             assert optim.x0 == x0
-            x0_new = np.array([0.6])
+            x0_new = np.array([0.5])
             optim = optimiser(cost=cost, x0=x0_new)
             assert optim.x0 == x0_new
             assert optim.x0 != x0
@@ -382,21 +385,17 @@ class TestOptimisation:
 
     def test_randomsearch_bounds(self, two_param_cost):
         # Test clip_candidates with bound
[…]
