### 2025-02-04 46e247f
Update example scripts
