### 2025-02-04 23d09df
Simplify ParameterSet.pybamm
