+++ tests/integration/test_monte_carlo.py
@@ -31,7 +31,7 @@ class Test_Sampling_SPM:
 
     @pytest.fixture
     def model(self):
-        parameter_set = pybop.ParameterSet.pybamm("Chen2020")
+        parameter_set = pybop.ParameterSet("Chen2020")
         x = self.ground_truth
         parameter_set.update(
             {
