### 2025-02-04 c811b60
Align simulation options (#656)
* Align simulation options

* Fix examples

* Try updating IDAKLU rtol

* Update geometry and copying

* Pass t_interp within predict

* Replace IDAKLU for experiment predictions

* Update example scripts

* style: pre-commit fixes

* Simplify ParameterSet.pybamm

* Add ParameterSet.copy

* Update CHANGELOG.md

* Relax tolerance by 0.001

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
