+++ pybop/costs/fitting_costs.py
@@ -74,211 +74,6 @@ class FittingCost(BaseCost):
         raise NotImplementedError
 
 
-class MeanSquaredError(FittingCost):
-    """
-    Mean square error (MSE) cost function.
-
-    Computes the mean square error between model predictions and the target
-    data, providing a measure of the differences between predicted values and
-    observed values.
-    """
-
-    def _error_measure(
-        self,
-        r: np.ndarray,
-        dy: Optional[np.ndarray] = None,
-    ) -> Union[float, tuple[float, np.ndarray]]:
-        e = np.mean(np.abs(r) ** 2)
-
-        if dy is not None:
-            de = 2 * np.mean((r * dy.T), axis=self.numpy_axis)
-            return e, de
-
-        return e
-
-
-class RootMeanSquaredError(FittingCost):
-    """
-    Root mean square error (RMSE) cost function.
-
-    Computes the root mean square error between model predictions and the target
-    data, providing a measure of the differences between predicted values and
-    observed values.
-    """
-
-    def _error_measure(
-        self,
-        r: np.ndarray,
-        dy: Optional[np.ndarray] = None,
-    ) -> Union[float, tuple[float, np.ndarray]]:
-        e = np.sqrt(np.mean(np.abs(r) ** 2))
-
-        if dy is not None:
-            de = np.mean((r * dy.T), axis=self.numpy_axis) / (e + np.finfo(float).eps)
-            return e, de
-
-        return e
-
-
-class MeanAbsoluteError(FittingCost):
-    """
-    Mean absolute error (MAE) cost function.
-
-    Computes the mean absolute error (MAE) between model predictions
-    and target data. The MAE is a measure of the average magnitude
-    of errors in a set of predictions, without considering their direction.
-    """
-
-    def _error_measure(
-        self,
-        r: np.ndarray,
-        dy: Optional[np.ndarray] = None,
-    ) -> Union[float, tuple[float, np.ndarray]]:
-        e = np.mean(np.abs(r))
-
-        if dy is not None:
-            sign_r = np.sign(r)
-            de = np.mean(sign_r * dy.T, axis=self.numpy_axis)
-            return e, de
-
-        return e
-
-
-class SumSquaredError(FittingCost):
-    """
-    Sum of squared error (SSE) cost function.
-
-    Computes the sum of the squares of the differences between model predictions
-    and target data, which serves as a measure of the total error between the
-    predicted and observed values.
-    """
-
-    def _error_measure(
-        self,
-        r: np.ndarray,
-        dy: Optional[np.ndarray] = None,
-    ) -> Union[float, tuple[float, np.ndarray]]:
-        e = np.sum(np.abs(r) ** 2)
-
-        if dy is not None:
-            de = 2 * np.sum((r * dy.T), axis=self.numpy_axis)
-            return e, de
-
-        return e
-
-
-class Minkowski(FittingCost):
-    """
-    The Minkowski distance is a generalisation of several distance metrics,
-    including the Euclidean and Manhattan distances. It is defined as:
-
-    .. math::
-        L_p(x, y) = ( \\sum_i |x_i - y_i|^p )^(1/p)
-
-    where p > 0 is the order of the Minkowski distance. For p ≥ 1, the
-    Minkowski distance is a metric. For 0 < p < 1, it is not a metric, as it
-    does not satisfy the triangle inequality, although a metric can be
-    obtained by removing the (1/p) exponent.
-
-    Special cases:
-
-    * p = 1: Manhattan distance
-    * p = 2: Euclidean distance
-    * p → ∞: Chebyshev distance (not implemented as yet)
-
-    This class implements the Minkowski distance as a cost function for
-    optimisation problems, allowing for flexible distance-based optimisation
[…]
