### 2025-02-04 8017a2e
Change mean to trapz
