### 2025-02-04 5fffa46
Update for multiple signals
