### 2025-02-04 8ee2011
Update for MultiFitting example
