### 2025-02-04 891c487
Update so trapz -> mean
