+++ tests/integration/test_spm_parameterisations.py
@@ -69,8 +69,8 @@ class Test_SPM_Parameterisation:
     def cost_cls(self, request):
         return request.param
 
-    def noise(self, sigma, values):
-        return np.random.normal(0, sigma, values)
+    def noisy(self, data, sigma):
+        return data + np.random.normal(0, sigma, len(data))
 
     @pytest.fixture(
         params=[
@@ -97,8 +97,7 @@ class Test_SPM_Parameterisation:
             {
                 "Time [s]": solution["Time [s]"].data,
                 "Current function [A]": solution["Current [A]"].data,
-                "Voltage [V]": solution["Voltage [V]"].data
-                + self.noise(self.sigma0, len(solution["Time [s]"].data)),
+                "Voltage [V]": self.noisy(solution["Voltage [V]"].data, self.sigma0),
             }
         )
 
@@ -195,12 +194,10 @@ class Test_SPM_Parameterisation:
             {
                 "Time [s]": solution["Time [s]"].data,
                 "Current function [A]": solution["Current [A]"].data,
-                "Voltage [V]": solution["Voltage [V]"].data
-                + self.noise(self.sigma0, len(solution["Time [s]"].data)),
-                "Bulk open-circuit voltage [V]": solution[
-                    "Bulk open-circuit voltage [V]"
-                ].data
-                + self.noise(self.sigma0, len(solution["Time [s]"].data)),
+                "Voltage [V]": self.noisy(solution["Voltage [V]"].data, self.sigma0),
+                "Bulk open-circuit voltage [V]": self.noisy(
+                    solution["Bulk open-circuit voltage [V]"].data, self.sigma0
+                ),
             }
         )
 
