### 2025-02-07 b74a5f6
Add output_variables property
