### 2025-02-07 8aa253d
corrected EP-BOLFI interface for merged changes
