### 2025-02-07 8decfd6
re-wrote parameterized costs to more easily implement new ones, added exponential fit costs
