+++ pybop/experimental/parameterized_costs.py
@@ -1,47 +1,54 @@
 from typing import Union
 
 import numpy as np
-from ep_bolfi.utility.fitting_functions import fit_sqrt
 from ep_bolfi.utility.preprocessing import find_occurrences
+from scipy.optimize import minimize
 
 from pybop import BaseCost
 
 
-class SquareRootFitOffset(BaseCost):
-    """
-    Square-root fit cost function.
+class ParameterizedCost(BaseCost):
+    """Base for defining cost functions based on fit parameters."""
 
-    Fits a square-root fit function and compares its offset between
-    model predictions and target data.
-    """
-
-    def __init__(
-        self,
-        problem,
-        time_start: float = None,
-        time_end: float = None,
-        threshold: float = 0.95,
-    ):
-        """
-        Parameters
-        ----------
-        problem : Inherited from pybop.BaseCost, see there.
-        time_start : float, optional
-            Set the time (in seconds) from which onwards the data shall be
-            fitted, counted from the start of the data. Default is the start.
-        time_end : float, optional
-            Set the time (in seconds) until which the data shall be fitted,
-            counted from the start of the data. Default is the end.
-        threshold : float, optional
-            A threshold for the R²-value of the square-root fit. If the initial
-            fit does not reach this threshold, data points at the end of the
-            time interval are culled until the fit to the remaining data
-            reaches it. Set to a value between 0 and 1, excluding 1 itself.
-        """
+    def __init__(self, problem, time_data: np.ndarray, feature: str):
         super().__init__(problem)
-        self.time_start = time_start
-        self.time_end = time_end
-        self.threshold = threshold
+        if feature not in self._supported_features:
+            raise ValueError(
+                "Feature '"
+                + feature
+                + "' not supported. Options: "
+                + str(self._supported_features)
+            )
+        self.feature = feature
+        self.time_data = time_data
+
+    @property
+    def _supported_features(self):
+        raise NotImplementedError
+
+    def _inverse_fit_function(self, y, *args):
+        raise NotImplementedError
+
+    def _fit_guess(self, t, y):
+        raise NotImplementedError
+
+    def _feature_selection(self, fit):
+        raise NotImplementedError
+
+    def _fit(self, t, y):
+        """
+        Uses SciPy to fit data. For numerical reasons, the fitting
+        involves applying the fit function to data and comparing to identity.
+        """
+        t = t - t[0]
+        fit_guess = self._fit_guess(t, y)
+        return self._feature_selection(
+            minimize(
+                lambda x: np.sum((t - self._inverse_fit_function(y, *x)) ** 2) ** 0.5,
+                x0=fit_guess,
+                method="trust-constr",
+            ).x
+        )
 
     def compute(
         self,
@@ -66,94 +73,138 @@ class SquareRootFitOffset(BaseCost):
             end_index = None
         error = np.asarray(
             [
-                fit_sqrt(
+                self._fit(
                     y["Time [s]"][start_index:end_index],
                     y[signal][start_index:end_index],
-                    self.threshold,
-                )[2][0]
-                - fit_sqrt(
-                    self._target["Time [s]"][start_index:end_index],
-                    self._target[signal][start_index:end_index],
-                    self.threshold,
-                )[2][0]
+                )
+                - self._fit(
+                    self.time_data[start_index:end_index],
+                    self.target[signal][start_index:end_index],
+                )
                 for signal in self.signal
             ]
         )
         return error.item() if self.n_outputs == 1 else np.sum(error)
 
 
-class SquareRootFitInverseSlope(BaseCost):
+class SquareRootFit(ParameterizedCost):
     """
     Square-root fit cost function.
 
-    Fits a square-root fit function and compares its inverse slope
[…]
