### 2025-02-07 113d090
Remove x0 from optim.log
