+++ pybop/plot/contour.py
@@ -197,11 +197,11 @@ def contour(
         )
 
         # Plot the initial guess
-        if optim.log["x0"] is not None:
+        if optim.x0 is not None:
             fig.add_trace(
                 go.Scatter(
-                    x=[optim.log["x0"][0]],
-                    y=[optim.log["x0"][1]],
+                    x=[optim.x0[0]],
+                    y=[optim.x0[1]],
                     mode="markers",
                     marker_symbol="x",
                     marker=dict(
