### 2025-02-07 5cfdc5f
Clarify x0 and bounds passing
