### 2025-02-07 3a7b0f7
Update bounds to dict
