### 2025-02-07 1273dd5
Sampler x0 corresponds to model parameters
