### 2025-02-08 115cff6
Separate out optimisation_result
