+++ pybop/optimisers/base_optimiser.py
@@ -1,12 +1,11 @@
 import warnings
 from copy import deepcopy
-from typing import Optional, Union
+from typing import Union
 
 import jax.numpy as jnp
 import numpy as np
-from pybamm import Solution
 
-from pybop import BaseCost, BaseLikelihood, Inputs, Parameter, Parameters
+from pybop import BaseCost, Inputs, OptimisationResult, Parameter, Parameters
 
 
 class BaseOptimiser:
@@ -352,318 +351,3 @@ class BaseOptimiser:
     @property
     def minimising(self):
         return self._minimising
-
-
-class OptimisationResult:
-    """
-    Stores the result of the optimisation.
-
-    Attributes
-    ----------
-    x : ndarray
-        The solution of the optimisation.
-    final_cost : float
-        The cost associated with the solution x.
-    n_iterations : int
-        Number of iterations performed by the optimiser.
-    scipy_result : scipy.optimize.OptimizeResult, optional
-        The result obtained from a SciPy optimiser.
-    pybamm_solution: pybamm.Solution or list[pybamm.Solution], optional
-        The best solution object(s) obtained from the optimisation.
-    """
-
-    def __init__(
-        self,
-        optim: BaseOptimiser,
-        x: Union[Inputs, np.ndarray] = None,
-        final_cost: Optional[float] = None,
-        sensitivities: Optional[dict] = None,
-        n_iterations: Optional[int] = None,
-        n_evaluations: Optional[int] = None,
-        time: Optional[float] = None,
-        scipy_result=None,
-    ):
-        self.optim = optim
-        self.cost = self.optim.cost
-        self.minimising = self.optim.minimising
-        self._transformation = self.optim.transformation
-        self.n_runs = 0
-        self._best_run = None
-        self._x = []
-        self._final_cost = []
-        self._sensitivities = None
-        self._fisher = []
-        self._n_iterations = []
-        self._n_evaluations = []
-        self._scipy_result = []
-        self._time = []
-        self._x0 = []
-        self._pybamm_solution = []
-
-        if x is not None:
-            # Transform the parameter values and update the sign of any final cost
-            # coming directly from an optimiser
-            x = self._transformation.to_model(x) if self._transformation else x
-            final_cost = (
-                final_cost * (1 if self.minimising else -1)
-                if final_cost is not None
-                else self.cost(x)
-            )
-            x0 = (
-                self.optim.parameters.initial_value()
-                if isinstance(self.optim, BaseOptimiser)
-                else None
-            )
-
-            # Evaluate the problem once more to update the solution
-            try:
-                self.cost(x)
-                pybamm_solution = self.cost.pybamm_solution
-            except Exception:
-                warnings.warn(
-                    "Failed to evaluate the model with best fit parameters.",
-                    UserWarning,
-                    stacklevel=2,
-                )
-                pybamm_solution = None
-
-            # Calculate Fisher Information if Likelihood
-            if isinstance(self.cost, BaseLikelihood):
-                fisher = self.cost.observed_fisher(x)
-                diag_fish = np.diag(fisher) if fisher is not None else None
-            else:
-                diag_fish = None
-
-            self._extend(
-                x=[x],
-                final_cost=[final_cost],
-                fisher=[diag_fish],
-                n_iterations=[n_iterations],
-                n_evaluations=[n_evaluations],
-                time=[time],
-                scipy_result=[scipy_result],
-                x0=[x0],
-                pybamm_solution=[pybamm_solution],
-            )
-
-    def add_result(self, result):
-        """Add a preprocessed OptimisationResult."""
-        self._extend(
-            x=result._x,  # noqa: SLF001
[…]
