+++ tests/integration/test_spm_parameterisations.py
@@ -167,7 +167,7 @@ class Test_SPM_Parameterisation:
         # Add sigma0 to ground truth for GaussianLogLikelihood
         if isinstance(optim.cost, pybop.GaussianLogLikelihood):
             self.ground_truth = np.concatenate(
-                (self.ground_truth, np.asarray([self.sigma0]))
+                (self.ground_truth[:2], np.asarray([self.sigma0]))
             )
 
         initial_cost = optim.cost(x0)
@@ -182,7 +182,7 @@ class Test_SPM_Parameterisation:
         else:
             assert initial_cost < results.final_cost
 
-        np.testing.assert_allclose(results.x[:2], self.ground_truth, atol=1.5e-2)
+        np.testing.assert_allclose(results.x, self.ground_truth, atol=1.5e-2)
         if isinstance(optim.cost, pybop.GaussianLogLikelihood):
             np.testing.assert_allclose(results.x[-1], self.sigma0, atol=5e-4)
 
@@ -250,7 +250,7 @@ class Test_SPM_Parameterisation:
 
         # Add sigma0 to ground truth for GaussianLogLikelihood
         if isinstance(two_signal_cost, pybop.GaussianLogLikelihood):
-            self.ground_truth = np.concatenate((self.ground_truth, combined_sigma0))
+            self.ground_truth = np.concatenate((self.ground_truth[:2], combined_sigma0))
 
         initial_cost = optim.cost(optim.parameters.initial_value())
         results = optim.run()
@@ -264,7 +264,7 @@ class Test_SPM_Parameterisation:
         else:
             assert initial_cost < results.final_cost
 
-        np.testing.assert_allclose(results.x[:2], self.ground_truth, atol=1.5e-2)
+        np.testing.assert_allclose(results.x, self.ground_truth, atol=1.5e-2)
         if isinstance(two_signal_cost, pybop.GaussianLogLikelihood):
             np.testing.assert_allclose(results.x[-2:], combined_sigma0, atol=5e-4)
 
@@ -310,11 +310,8 @@ class Test_SPM_Parameterisation:
         initial_state = {"Initial SoC": init_soc}
         experiment = pybop.Experiment(
             [
-                (
-                    "Discharge at 0.5C for 8 minutes (8 second period)",
-                    "Charge at 0.5C for 8 minutes (8 second period)",
-                )
+                "Discharge at 0.5C for 8 minutes (8 second period)",
+                "Charge at 0.5C for 8 minutes (8 second period)",
             ]
         )
-        sim = model.predict(initial_state=initial_state, experiment=experiment)
-        return sim
+        return model.predict(initial_state=initial_state, experiment=experiment)
