### 2025-02-08 860378e
Return to combined single_call
