+++ pybop/costs/base_cost.py
@@ -96,74 +96,64 @@ class BaseCost:
         # Convert dict to list for sequential computations
         if isinstance(inputs, dict):
             inputs = list(inputs.values())
-        input_list = np.atleast_2d(inputs)
-
-        minimising = self.minimising or not for_optimiser
-        sign = 1 if minimising else -1
+        inputs_list = np.atleast_2d(inputs)
 
         results = []
-        for input_val in input_list:
-            result = self._evaluate_single_input(
-                input_val,
+        for inputs in inputs_list:
+            result = self.single_call(
+                inputs,
                 calculate_grad=calculate_grad,
                 apply_transform=apply_transform,
-                sign=sign,
+                for_optimiser=for_optimiser,
             )
             results.append(result)
 
-        return results[0] if len(input_list) == 1 else results
+        return results[0] if len(inputs_list) == 1 else results
 
-    def _evaluate_single_input(
+    def single_call(
         self,
-        input_value: Union[Inputs, np.ndarray],
+        inputs: Union[Inputs, np.ndarray],
         calculate_grad: bool,
         apply_transform: bool,
-        sign: int,
+        for_optimiser: bool = False,
     ) -> Union[float, tuple[float, np.ndarray]]:
-        """Evaluate cost (and optional gradient) for a single input."""
+        """Evaluate the cost and (optionally) the gradient for a single set of inputs."""
         # Setup input transformation
         self.has_transform = self.transformation is not None and apply_transform
-        model_inputs = self.parameters.verify(self._apply_transformations(input_value))
+        model_inputs = self.parameters.verify(self._apply_transformations(inputs))
         self.parameters.update(values=list(model_inputs.values()))
 
-        if self._has_separable_problem:
-            return self._evaluate_separable_problem(
-                input_value, calculate_grad=calculate_grad, sign=sign
-            )
-
-        return self._evaluate_non_separable_problem(
-            calculate_grad=calculate_grad, sign=sign
-        )
-
-    def _evaluate_separable_problem(
-        self, input_value: Union[Inputs, np.ndarray], calculate_grad: bool, sign: int
-    ) -> Union[float, tuple[float, np.ndarray]]:
-        """Evaluation for separable problems."""
-        if calculate_grad:
-            y, dy = self.problem.evaluateS1(self.problem.parameters.as_dict())
-            cost, grad = self.compute(y, dy=dy)
-
-            if self.has_transform and np.isfinite(cost):
-                jac = self.transformation.jacobian(input_value)
-                grad = np.matmul(grad, jac)
-
-            return cost * sign, grad * sign
-
-        y = self.problem.evaluate(self.problem.parameters.as_dict())
-        return self.compute(y, dy=None) * sign
-
-    def _evaluate_non_separable_problem(
-        self, calculate_grad: bool, sign: int
-    ) -> Union[float, tuple[float, np.ndarray]]:
-        """Evaluation for non-separable problems."""
         y = self.DeferredPrediction
         dy = self.DeferredPrediction if calculate_grad else None
 
+        minimising = self.minimising or not for_optimiser
+
+        if self._has_separable_problem:
+            if calculate_grad:
+                y, dy = self.problem.evaluateS1(self.problem.parameters.as_dict())
+                cost, grad = self.compute(y, dy=dy)
+
+                if self.has_transform and np.isfinite(cost):
+                    jac = self.transformation.jacobian(inputs)
+                    grad = np.matmul(grad, jac)
+
+                return cost * (1 if minimising else -1), grad * (
+                    1 if minimising else -1
+                )
+
+            y = self.problem.evaluate(self.problem.parameters.as_dict())
[…]
