+++ pybop/optimisers/scipy_optimisers.py
@@ -77,7 +77,9 @@ class BaseSciPyOptimiser(BaseOptimiser):
 
         # Choose method to evaluate
         def fun(x):
-            return self.cost_call(x, calculate_grad=self._needs_sensitivities)
+            return self.call_cost(
+                x, cost=self.cost, calculate_grad=self._needs_sensitivities
+            )
 
         # Create evaluator object
         self.evaluator = SciPyEvaluator(fun)
