### 2025-02-08 7fb7fc9
Align sampler with cost_call
