### 2025-02-08 8b8bb2c
feat: Adds DFN example, optional rebuild_on_evaluation arg to FittingProblem, synthetic data generation
