+++ examples/scripts/battery_parameterisation/simple_dfn.py
@@ -1,0 +1,72 @@
+import numpy as np
+import pybamm
+
+import pybop
+
+# Define model and use high-performant solver for sensitivities
+solver = pybamm.IDAKLUSolver()
+parameter_set = pybop.ParameterSet("Chen2020")
+model = pybop.lithium_ion.DFN(parameter_set=parameter_set, solver=solver)
+
+# Fitting parameters
+parameters = pybop.Parameters(
+    pybop.Parameter(
+        "Negative electrode active material volume fraction",
+        prior=pybop.Gaussian(0.68, 0.05),
+        initial_value=0.55,
+        bounds=[0.4, 0.9],
+    ),
+    pybop.Parameter(
+        "Positive electrode active material volume fraction",
+        prior=pybop.Gaussian(0.58, 0.05),
+        initial_value=0.55,
+        bounds=[0.4, 0.9],
+    ),
+)
+csv_data = np.loadtxt(
+    "../../data/synthetic/charge_discharge_75.csv", delimiter=",", skiprows=1
+)
+
+# Form dataset
+dataset = pybop.Dataset(
+    {
+        "Time [s]": csv_data[:, 0],
+        "Current function [A]": csv_data[:, 1],
+        "Voltage [V]": csv_data[:, 2],
+        "Bulk open-circuit voltage [V]": csv_data[:, 3],
+    }
+)
+
+signal = ["Voltage [V]", "Bulk open-circuit voltage [V]"]
+# Generate problem, cost function, and optimisation class
+problem = pybop.FittingProblem(
+    model,
+    parameters,
+    dataset,
+    signal=signal,
+    initial_state={"Initial open-circuit voltage [V]": csv_data[0, 2]},
+    build_on_evaluation=False,
+)
+cost = pybop.SumofPower(problem, p=2.5)
+
+optim = pybop.IRPropMin(
+    cost,
+    verbose=True,
+    max_iterations=100,
+    max_unchanged_iterations=45,
+)
+
+# Run optimisation
+results = optim.run()
+
+# Plot the timeseries output
+pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+
+# Plot convergence
+pybop.plot.convergence(optim)
+
+# Plot the parameter traces
+pybop.plot.parameters(optim)
+
+# Plot the cost landscape with optimisation path
+pybop.plot.surface(optim)
