+++ examples/data/synthetic/README.md
@@ -1,0 +1,5 @@
+# Data description
+The data files in this directory have the following metadata
+
+`charge_discharge_75.csv`
+- Voltage, Current, and Time data for a Chen2020/DFN combination that started at 75% SOC. All other parameters were left default to the Chen2020 dataset. 1mV of noise is applied.
