### 2025-02-08 bf2dd57
Fix test_spm_parameterisations
