### 2025-02-09 cf6d447
Move transformation to cost interface
