### 2025-02-09 f4d1e1b
Move minimising to cost interface
