### 2025-02-09 a8f6c16
Split apply_transformation
