### 2025-02-09 aadf126
Align contour with current functionality
