### 2025-02-09 6ca1302
Fix Gaussian logpdf gradient
