### 2025-02-10 66c53e6
Change SumofPower to SumOfPower
