### 2025-02-10 d15d389
examples: uses previously generated data as .csv instead of inline generated data, adds dfn pulse fit example
