+++ examples/scripts/battery_parameterisation/simple_pulse_fit.py
@@ -1,8 +1,14 @@
 import os
 
 import numpy as np
+import pybamm
 
 import pybop
+
+# This example introduces pulse fitting
+# within PyBOP. Data is loaded from a local `csv` file
+# and particle diffusivity for a DFN model is performed.
+# Additionally, uncertainty metrics are computed.
 
 # Get the current directory location and convert to absolute path
 current_dir = os.path.dirname(os.path.abspath(__file__))
@@ -12,29 +18,27 @@ dataset_path = os.path.join(
 
 # Define model and use high-performant solver for sensitivities
 parameter_set = pybop.ParameterSet("Chen2020")
-model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
+model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=pybamm.IDAKLUSolver())
 
-# Construct the fitting parameters
-# with initial values sampled from a different distribution
+# Fitting parameters
 parameters = pybop.Parameters(
     pybop.Parameter(
-        "Negative electrode active material volume fraction",
-        prior=pybop.Uniform(0.3, 0.9),
-        bounds=[0.3, 0.8],
-        initial_value=pybop.Uniform(0.4, 0.75).rvs()[0],
-        true_value=parameter_set["Negative electrode active material volume fraction"],
+        "Negative particle diffusivity [m2.s-1]",
+        prior=pybop.Gaussian(4e-14, 1e-14),
+        transformation=pybop.LogTransformation(),
+        bounds=[1e-14, 1e-13],
     ),
     pybop.Parameter(
-        "Positive electrode active material volume fraction",
-        prior=pybop.Uniform(0.3, 0.9),
-        initial_value=pybop.Uniform(0.4, 0.75).rvs()[0],
-        true_value=parameter_set["Positive electrode active material volume fraction"],
-        # no bounds
+        "Positive particle diffusivity [m2.s-1]",
+        prior=pybop.Gaussian(7e-15, 1e-15),
+        transformation=pybop.LogTransformation(),
+        bounds=[1e-15, 1e-14],
     ),
 )
 
 # Import the synthetic dataset
 csv_data = np.loadtxt(dataset_path, delimiter=",", skiprows=1)
+
 
 # Form dataset
 dataset = pybop.Dataset(
@@ -46,20 +50,28 @@ dataset = pybop.Dataset(
 )
 
 # Generate problem, cost function, and optimisation class
+# In this example, we initialise the DFN at the first voltage
+# point in `csv_data`, an optimise without rebuilding the
+# model on every evaluation.
 problem = pybop.FittingProblem(
     model,
     parameters,
     dataset,
     initial_state={"Initial open-circuit voltage [V]": csv_data[0, 2]},
-)
-likelihood = pybop.GaussianLogLikelihoodKnownSigma(problem, sigma0=2e-3 * 1.05)
-posterior = pybop.LogPosterior(likelihood)
-optim = pybop.IRPropMin(
-    posterior, max_iterations=125, max_unchanged_iterations=60, sigma0=0.01
+    build_on_evaluation=False,
 )
 
+likelihood = pybop.GaussianLogLikelihoodKnownSigma(problem, sigma0=2e-4)
+
+optim = pybop.XNES(
+    likelihood,
+    verbose=True,
+    max_iterations=100,
+    max_unchanged_iterations=45,
+)
+
+# Run optimisation
 results = optim.run()
-print(parameters.true_value())
 
 # Plot the timeseries output
 pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
@@ -70,6 +82,5 @@ pybop.plot.convergence(optim)
[…]
