### 2025-02-10 a885127
fixed time domain issue with parameterized costs
