+++ pybop/experimental/test_ep_bolfi.py
@@ -2,8 +2,9 @@ import numpy as np
 from _ep_bolfi import EP_BOLFI
 from multivariate_parameters import MultivariateParameters
 from multivariate_priors import MultivariateGaussian
+from parameterized_costs import SquareRootFit
 from plot_bayes import bayes
-from pybamm import print_citations
+from pybamm import Experiment, print_citations
 
 import pybop
 
@@ -11,12 +12,28 @@ parameter_set = pybop.ParameterSet.pybamm("Chen2020")
 original_diffusivity = parameter_set["Positive particle diffusivity [m2.s-1]"]
 model = pybop.lithium_ion.DFN(parameter_set=parameter_set)
 
+"""
 t_eval = np.arange(0, 901, 3)
 values = model.predict(t_eval=t_eval)
-
 dataset = pybop.Dataset(
     {
         "Time [s]": t_eval,
+        "Current function [A]": values["Current [A]"].data,
+        "Voltage [V]": values["Voltage [V]"].data,
+    }
+)
+"""
+
+experiment = Experiment(
+    [
+        "Discharge at 1 C for 15 minutes (1 second period)",
+        "Rest for 15 minutes (1 second period)",
+    ]
+)
+values = model.predict(experiment=experiment)
+dataset = pybop.Dataset(
+    {
+        "Time [s]": values["Time [s]"].data,
         "Current function [A]": values["Current [A]"].data,
         "Voltage [V]": values["Voltage [V]"].data,
     }
@@ -34,19 +51,28 @@ unknowns = MultivariateParameters(
 
 if __name__ == "__main__":
     problem = pybop.FittingProblem(
-        model, unknowns, dataset, signal=["Voltage [V]"], parallelizable=True
+        model,
+        unknowns,
+        dataset,
+        additional_variables=["Time [s]"],
+        signal=["Voltage [V]"],
+        parallelizable=True,
     )
-    cost = pybop.WeightedCost(pybop.SumSquaredError(problem))
+    # cost = pybop.WeightedCost(pybop.SumSquaredError(problem))
+    cost = pybop.WeightedCost(
+        SquareRootFit(problem, values["Time [s]"].data, "inverse_slope", time_end=90)
+    )
     # Override the forced Parameters class in BaseCost instantiation.
     cost.parameters = unknowns
     optim = EP_BOLFI(
         cost,
-        t_eval=t_eval,
+        # t_eval=t_eval,
+        experiment=experiment,
         ep_iterations=1,
         ep_dampener=0,
-        bolfi_initial_evidence=15,
-        bolfi_total_evidence=30,
-        bolfi_posterior_samples=20,
+        bolfi_initial_evidence=15 // 5,
+        bolfi_total_evidence=30 // 5,
+        bolfi_posterior_samples=20 // 5,
         verbose=True,
     )
 
