### 2025-02-10 839ab04
examples: updates pulse fitting example, adds corresponding data, README update
