+++ examples/data/synthetic/README.md
@@ -1,5 +1,10 @@
 # Data description
 The data files in this directory have the following metadata
 
-`charge_discharge_75.csv`
-- Voltage, Current, and Time data for a Chen2020/DFN combination that started at 75% SOC. All other parameters were left default to the Chen2020 dataset. 1mV of noise is applied.
+`{model}_charge_discharge_{soc}.csv`
+- Time, current, terminal voltage, and open-circuit voltage data for a Chen2020 parameters at the corresponding SOC. 1mV of noise is applied to both voltage signals.
+- See `discharge_charge_data_gen.py` for reference
+
+`{model}_pulse_{soc}.csv`
+- Time, current, terminal voltage, and open-circuit voltage data for a Chen2020 parameters at the corresponding SOC. 1mV of noise is applied to both voltage signals.
+- See `pulse_data_gen.py` for reference
