### 2025-02-10 f74489f
examples: updates data_gen scripts, corresponding data, and adds docstring and changelog entries
