### 2025-02-10 57e283b
fix: restore cost gradient signal combination axes
