### 2025-02-11 d9d855b
fixed parameterized cost example
