+++ pybop/experimental/test_ep_bolfi.py
@@ -1,16 +1,29 @@
+from copy import deepcopy
+
 import numpy as np
 from _ep_bolfi import EP_BOLFI
 from multivariate_parameters import MultivariateParameters
 from multivariate_priors import MultivariateGaussian
 from parameterized_costs import SquareRootFit
 from plot_bayes import bayes
-from pybamm import Experiment, print_citations
+from pybamm import CasadiSolver, Experiment, print_citations
 
 import pybop
 
 parameter_set = pybop.ParameterSet.pybamm("Chen2020")
 original_diffusivity = parameter_set["Positive particle diffusivity [m2.s-1]"]
 model = pybop.lithium_ion.DFN(parameter_set=parameter_set)
+model.solver = CasadiSolver(
+    rtol=1e-5,
+    atol=1e-5,
+    root_tol=1e-3,
+    max_step_decrease_count=10,
+    extra_options_setup={
+        "disable_internal_warnings": True,
+        "newton_scheme": "tfqmr",
+    },
+    return_solution_if_failed_early=True,
+)
 
 """
 t_eval = np.arange(0, 901, 3)
@@ -30,7 +43,7 @@ experiment = Experiment(
         "Rest for 15 minutes (1 second period)",
     ]
 )
-values = model.predict(experiment=experiment)
+values = deepcopy(model).predict(experiment=experiment)
 dataset = pybop.Dataset(
     {
         "Time [s]": values["Time [s]"].data,
@@ -43,10 +56,10 @@ unknowns = MultivariateParameters(
     pybop.Parameter(
         "Positive particle diffusivity [m2.s-1]",
         initial_value=10 * original_diffusivity,
-        bounds=[original_diffusivity / 100, original_diffusivity * 100],
+        bounds=[original_diffusivity / 1000, original_diffusivity * 1000],
         transformation=pybop.LogTransformation(),
     ),
-    prior=MultivariateGaussian([np.log(original_diffusivity)], [[np.log(10)]]),
+    prior=MultivariateGaussian([np.log(original_diffusivity)], [[np.log(100)]]),
 )
 
 if __name__ == "__main__":
@@ -70,9 +83,9 @@ if __name__ == "__main__":
         experiment=experiment,
         ep_iterations=1,
         ep_dampener=0,
-        bolfi_initial_evidence=15 // 5,
-        bolfi_total_evidence=30 // 5,
-        bolfi_posterior_samples=20 // 5,
+        bolfi_initial_evidence=10,
+        bolfi_total_evidence=20,
+        bolfi_posterior_samples=10,
         verbose=True,
     )
 
