### 2025-02-12 bc07292
feat: small simplification to domain selection for parameter checks
