### 2025-02-13 8afef11
Apply suggestions from code review
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
