### 2025-02-13 4d0b9cb
Pass transformations and invert_cost to CostInterface
