+++ tests/integration/test_thevenin_parameterisation.py
@@ -106,7 +106,7 @@ class TestTheveninParameterisation:
 
         # Assertions
         if not np.allclose(x0, self.ground_truth, atol=1e-5):
-            if optim.minimising:
+            if results.minimising:
                 assert initial_cost > results.final_cost
             else:
                 assert initial_cost < results.final_cost
