### 2025-02-13 6759b82
Transform failed gradients to preserve direction
