### 2025-02-13 b2e472f
Pass x0 and bounds in model space
