+++ pybop/problems/multi_fitting_problem.py
@@ -138,7 +138,9 @@ class MultiFittingProblem(BaseProblem):
 
         combined_domain = []
         combined_signal = []
-        all_derivatives = []
+        dy = dict.fromkeys(self.parameters.keys())
+        for key in self.parameters.keys():
+            dy[key] = {"Combined signal": []}
 
         for problem in self.problems:
             problem_inputs = problem.parameters.as_dict()
@@ -153,13 +155,13 @@ class MultiFittingProblem(BaseProblem):
             for signal in problem.signal:
                 combined_domain.extend(domain_data)
                 combined_signal.extend(problem_output[signal])
-            all_derivatives.append(dyi)
+                for key in self.parameters.keys():
+                    dy[key]["Combined signal"].extend(dyi[key][signal])
 
         y = {
             self.domain: np.asarray(combined_domain),
             "Combined signal": np.asarray(combined_signal),
         }
-        dy = np.concatenate(all_derivatives) if all_derivatives else None
 
         return (y, dy)
 
