+++ pybop/costs/fitting_costs.py
@@ -18,35 +18,39 @@ class FittingCost(BaseCost):
 
     def __init__(self, problem):
         super().__init__(problem)
-        self.output_axis = 0 if self.n_outputs > 1 else 1
 
     def compute(
         self,
         y: dict,
-        dy: Optional[np.ndarray] = None,
+        dy: Optional[dict] = None,
     ) -> Union[float, tuple[float, np.ndarray]]:
         """
         Computes the cost function for the given predictions.
 
         Parameters
         ----------
-        y : dict
+        y : dict[str, np.ndarray[np.float64]]
             The dictionary of predictions with keys designating the signals for fitting.
-        dy : np.ndarray, optional
-            The corresponding gradient with respect to the parameters for each signal.
+        dy : dict[str, dict[str, np.ndarray]], optional
+            The corresponding sensitivities to each parameter for each signal.
 
         Returns
         -------
-        tuple or float
+        np.float64 or tuple[np.float64, np.ndarray[np.float64]]
             If dy is not None, returns a tuple containing the cost (float) and the
-            gradient (np.ndarray), otherwise returns only the computed cost (float).
+            sensitivities with dimensions (len(parameters), len(signal), len(domain_data)),
+            otherwise returns only the cost.
         """
         # Early return if the prediction is not verified
         if not self.verify_prediction(y):
             return (np.inf, self.grad_fail) if dy is not None else np.inf
 
-        # Compute the residual
+        # Compute the residual for all signals
         r = np.asarray([y[signal] - self._target[signal] for signal in self.signal])
+
+        # Extract the sensitivities for all signals and parameters
+        if dy is not None:
+            dy = self.compute_model_parameter_sensitivities(dy)
 
         return self._error_measure(r=r, dy=dy)
 
@@ -61,15 +65,18 @@ class FittingCost(BaseCost):
         Parameters
         ----------
         r : np.ndarray
-            The residual difference between the model prediction and the target.
+            The residual difference between the model prediction and the target. The
+            dimensions of r are (len(signal), len(domain_data)).
         dy : np.ndarray, optional
             The corresponding gradient with respect to the parameters for each signal.
+            The dimensions of dy are (len(parameters), len(signal), len(domain_data)).
 
         Returns
         -------
-        tuple or float
+        np.float64 or tuple[np.float64, np.ndarray[np.float64]]
             If dy is not None, returns a tuple containing the cost (float) and the
-            gradient (np.ndarray), otherwise returns only the computed cost (float).
+            gradient with dimensions (len(parameters), len(signal), len(domain_data)),
+            otherwise returns only the cost.
         """
         raise NotImplementedError
 
