### 2025-02-16 9834597
Align samplers and fix cost gradient (#661)
* Remove x0 from optim.log

* Clarify x0 and bounds passing

* Update bounds to dict

* Sampler x0 corresponds to model parameters

* Separate out optimisation_result

* Update integration tests

* Return to combined single_call

* Align sampler with cost_call

* Fix test_spm_parameterisations

* Update parameters.get_bounds

* Move transformation to cost interface

* Rename test_cost_transformations

* Move minimising to cost interface

* Split apply_transformation

* Update transformation example and plotting

* Update CHANGELOG.md

* Improve trans
[…]
