+++ tests/integration/test_thevenin_parameterisation.py
@@ -40,13 +40,13 @@ class TestTheveninParameterisation:
             pybop.Parameter(
                 "R0 [Ohm]",
                 prior=pybop.Gaussian(0.05, 0.01),
-                bounds=[0, 0.1],
+                bounds=[1e-6, 0.1],
                 transformation=pybop.LogTransformation(),
             ),
             pybop.Parameter(
                 "R1 [Ohm]",
                 prior=pybop.Gaussian(0.05, 0.01),
-                bounds=[0, 0.1],
+                bounds=[1e-6, 0.1],
                 transformation=pybop.LogTransformation(),
             ),
         )
@@ -70,8 +70,8 @@ class TestTheveninParameterisation:
     @pytest.mark.parametrize(
         "optimiser, method",
         [
-            (pybop.SciPyMinimize, "trust-constr"),
             (pybop.SciPyMinimize, "SLSQP"),
+            (pybop.SciPyMinimize, "trust-constr"),
             (pybop.SciPyMinimize, "L-BFGS-B"),
             (pybop.SciPyMinimize, "COBYLA"),
             (pybop.GradientDescent, ""),
@@ -106,7 +106,7 @@ class TestTheveninParameterisation:
 
         # Assertions
         if not np.allclose(x0, self.ground_truth, atol=1e-5):
-            if optim.minimising:
+            if results.minimising:
                 assert initial_cost > results.final_cost
             else:
                 assert initial_cost < results.final_cost

[…]
