### 2025-02-17 8d600c3
feat: removes events from base pybamm model by default, w/ option to reapply
