+++ tests/unit/test_cost.py
@@ -164,6 +164,7 @@ class TestCosts:
 
             # Test infeasible locations
             cost.problem.model.allow_infeasible_solutions = False
+            cost.problem.model.apply_events()
             assert cost([1.1]) == np.inf
             assert cost([1.1], calculate_grad=True) == (np.inf, cost._de)
             assert cost([0.01]) == np.inf
