+++ examples/scripts/comparison_examples/exponential_natural_evolution.py
@@ -22,7 +22,7 @@ parameters = pybop.Parameters(
 
 sigma = 0.001
 t_eval = np.arange(0, 900, 3)
-values = model.predict(t_eval=t_eval)
+values = model.predict(t_eval=t_eval, initial_state={"Initial SoC": 0.2})
 corrupt_values = values["Voltage [V]"].data + np.random.normal(0, sigma, len(t_eval))
 
 # Form dataset
@@ -35,7 +35,12 @@ dataset = pybop.Dataset(
 )
 
 # Generate problem, cost function, and optimisation class
-problem = pybop.FittingProblem(model, parameters, dataset)
+problem = pybop.FittingProblem(
+    model,
+    parameters,
+    dataset,
+    # initial_state={"Initial open-circuit voltage [V]": dataset["Voltage [V]"][0]},
+)
 cost = pybop.SumSquaredError(problem)
 optim = pybop.XNES(cost, max_iterations=100)
 
