### 2025-02-17 187bed8
feat: moves infeasible solution interface to problem class, updates tests
