+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Breaking Changes
 
+- [#672](https://github.com/pybop-team/PyBOP/pull/672) - Removes pybamm model events by default, removes `allow_infeasible_solutions` from optimisers and adds `infeasible_solutions` arg to problem classes.
 - [#661](https://github.com/pybop-team/PyBOP/pull/661) - Adds `pybop.CostInterface` which aligns the optimisers and samplers with a unified `call_cost` in which transformations and sign inversions are applied. Also includes bug fixes for transformations and gradient calculations.
 
 
