### 2025-02-17 99a5731
restores classification, adds direct event unit tests
