### 2025-02-17 e72a69a
fix: resets model solver for consistent event application, fixes tests
