+++ tests/unit/test_models.py
@@ -145,28 +145,27 @@ class TestModels:
                 model.predict(t_eval=t_eval, inputs=inputs)
 
     def test_removal_and_apply_events(self, model):
-        if isinstance(model, (pybop.lithium_ion.SPM, pybop.lithium_ion.SPMe)):
-            model = (
-                model.new_copy()
-            )  # Copy the model, so building does not affect later tests
-            model.build()
-            assert not model._built_model.events
-            t_eval = np.linspace(0, 1e3, 5)
-            inputs = {
-                "Negative electrode active material volume fraction": 0.9,
-                "Positive electrode active material volume fraction": 0.9,
-            }
-            assert (
-                len(model.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
-                == 5
-            )
+        if isinstance(model, pybop.lithium_ion.SPM):
+            # Test with already built/solved model and new model instances
+            for mdl in [model, model.new_copy()]:
+                mdl.build()
+                assert not mdl._built_model.events
+                t_eval = np.linspace(0, 1e3, 3)
+                inputs = {
+                    "Negative electrode active material volume fraction": 0.5,
+                    "Positive electrode active material volume fraction": 0.5,
+                }
+                assert (
+                    len(mdl.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
+                    == 3
+                )
 
-            # Reapply events
-            model.apply_events()
-            assert (
-                len(model.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
-                <= 5
-            )
+                # Reapply events
+                mdl.apply_events()
+                assert (
+                    len(mdl.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
+                    <= 3
+                )
 
     def test_build(self, model):
         if isinstance(model, pybop.lithium_ion.SPMe):
