### 2025-02-17 d42edbb
Change SumofPower to SumOfPower
