+++ examples/scripts/comparison_examples/adamw.py
@@ -57,7 +57,7 @@ dataset = pybop.Dataset(
 signal = ["Voltage [V]", "Bulk open-circuit voltage [V]"]
 # Generate problem, cost function, and optimisation class
 problem = pybop.FittingProblem(model, parameters, dataset, signal=signal)
-cost = pybop.SumofPower(problem, p=2.5)
+cost = pybop.SumOfPower(problem, p=2.5)
 
 optim = pybop.AdamW(
     cost,
