### 2025-02-17 0e4f30d
Pass dy as dict and update documentation
