+++ pybop/costs/_likelihoods.py
@@ -56,10 +56,9 @@ class BaseLikelihood(BaseCost):
 
 class BaseMetaLikelihood(BaseLikelihood):
     """
-    Base class for likelihood classes which have a meta-likelihood
-    such as `LogPosterior` or `ScaledLoglikelihood`. This class
-    points the required attributes towards the composed likelihood
-    class.
+    Base class for likelihood classes which have a meta-likelihood such as `LogPosterior` or
+    `ScaledLoglikelihood`. This class points the required attributes towards the composed
+    likelihood class.
     """
 
     def __init__(self, log_likelihood: BaseLikelihood):
@@ -85,16 +84,15 @@ class BaseMetaLikelihood(BaseLikelihood):
 
 class GaussianLogLikelihoodKnownSigma(BaseLikelihood):
     """
-    This class represents a Gaussian Log Likelihood with a known sigma,
-    which assumes that the data follows a Gaussian distribution and computes
-    the log-likelihood of observed data under this assumption.
+    This class represents a Gaussian log-likelihood with a known sigma, which computes the
+    log-likelihood under the assumption that measurement noise on the target data follows a
+    Gaussian distribution.
 
     Parameters
     ----------
     sigma0 : scalar or array
-        Initial standard deviation around ``x0``. Either a scalar value (one
-        standard deviation for all coordinates) or an array with one entry
-        per dimension.
+        Initial standard deviation around ``x0``. Either a scalar value (one standard deviation
+        for all coordinates) or an array with one entry per dimension.
     """
 
     def __init__(self, problem: BaseProblem, sigma0: Union[list[float], float]):
@@ -120,13 +118,14 @@ class GaussianLogLikelihoodKnownSigma(BaseLikelihood):
 
         # Calculate residuals and error
         r = np.asarray([self._target[signal] - y[signal] for signal in self.signal])
-        e = np.sum(self._offset + self._multip * np.sum(np.real(r * np.conj(r))))
+        l = np.sum(self._offset + self._multip * np.sum(np.real(r * np.conj(r))))
 
         if dy is not None:
-            dl = np.sum((np.sum((r * dy.T), axis=2) / self.sigma2), axis=1)
-            return e, dl
+            dy = self.stack_sensitivities(dy)
+            dl = np.sum((np.sum((r * dy), axis=2) / self.sigma2), axis=1)
+            return l, dl
 
-        return e
+        return l
 
     def check_sigma0(self, sigma0: Union[np.ndarray, float]):
         """
@@ -145,12 +144,11 @@ class GaussianLogLikelihoodKnownSigma(BaseLikelihood):
 
 class GaussianLogLikelihood(BaseLikelihood):
     """
-    This class represents a Gaussian Log Likelihood, which assumes that the
-    data follows a Gaussian distribution and computes the log-likelihood of
-    observed data under this assumption.
+    This class represents a Gaussian log-likelihood, which computes the log-likelihood under
+    the assumption that measurement noise on the target data follows a Gaussian distribution.
 
-    This class estimates the standard deviation of the Gaussian distribution
-    alongside the parameters of the model.
+    This class estimates the standard deviation of the Gaussian distribution alongside the
+    parameters of the model.
 
     Attributes
     ----------
@@ -232,10 +230,19 @@ class GaussianLogLikelihood(BaseLikelihood):
 
         This method only computes the likelihood, without calling problem.evaluate().
 
+        Parameters
+        ----------
+        y : dict[str, np.ndarray[np.float64]]
+            The dictionary of predictions with keys designating the signals for fitting.
+        dy : dict[str, dict[str, np.ndarray]], optional
[…]
