+++ pybop/problems/base_problem.py
@@ -84,7 +84,6 @@ class BaseProblem:
             self.domain = "Frequency [Hz]" if self.eis else "Time [s]"
 
         # If model.solver is IDAKLU, set output vars for improved performance
-        self.output_vars = tuple(self.signal + self.additional_variables)
         if self._model is not None and isinstance(self._model.solver, IDAKLUSolver):
             self._solver_copy = self._model.solver.copy()
             self._model.solver = IDAKLUSolver(
@@ -94,7 +93,7 @@ class BaseProblem:
                 root_tol=self._solver_copy.root_tol,
                 extrap_tol=self._solver_copy.extrap_tol,
                 options=self._solver_copy._options,  # noqa: SLF001
-                output_variables=self.output_vars,
+                output_variables=tuple(self.output_variables),
             )
 
     def set_initial_state(self, initial_state: Optional[dict] = None):
@@ -115,6 +114,10 @@ class BaseProblem:
     @property
     def n_outputs(self):
         return len(self.signal)
+
+    @property
+    def output_variables(self):
+        return list(set(self.signal + self.additional_variables))
 
     def evaluate(self, inputs: Inputs, eis=False):
         """

[…]
