### 2025-02-17 3dac9e4
tests: adds coverage
