### 2025-02-17 bd7c9c7
Add message to OptimisationResult
