+++ tests/unit/test_optimisation.py
@@ -600,6 +600,7 @@ class TestOptimisation:
             f"  Optimisation time: {results.time} seconds\n"
             f"  Number of iterations: {results.n_iterations}\n"
             f"  Number of evaluations: {results.n_evaluations}\n"
+            f"  Reason for stopping: {results.message}\n"
             f"  SciPy result available: No\n"
             f"  PyBaMM Solution available: Yes"
         )
@@ -648,6 +649,7 @@ class TestOptimisation:
             f"  Optimisation time: {results.time} seconds\n"
             f"  Number of iterations: {results.n_iterations}\n"
             f"  Number of evaluations: {results.n_evaluations}\n"
+            f"  Reason for stopping: {results.message}\n"
             f"  SciPy result available: No\n"
             f"  PyBaMM Solution available: Yes"
         )
