+++ pybop/optimisers/base_pints_optimiser.py
@@ -339,8 +339,6 @@ class BasePintsOptimiser(BaseOptimiser):
             raise
 
         total_time = time() - start_time
-        if self.verbose:
-            print("Halt: " + halt_message)
 
         # Save post-run statistics
         self._evaluations = evaluations
@@ -361,6 +359,7 @@ class BasePintsOptimiser(BaseOptimiser):
             n_iterations=self._iterations,
             n_evaluations=self._evaluations,
             time=total_time,
+            message=halt_message,
         )
 
     def f_guessed_tracking(self):
