### 2025-02-17 830c48d
Update test_halting
