### 2025-02-17 61132e7
tests: add physical_viability coverage through optim arg, test on additional models
