+++ tests/unit/test_optimisation.py
@@ -711,7 +711,7 @@ class TestOptimisation:
     def test_optimisation_results(self, cost):
         # Construct OptimisationResult
         cost.problem.model.apply_events()
-        optim = pybop.Optimisation(cost=cost)
+        optim = pybop.Optimisation(cost=cost, physical_viability=True)
         results = pybop.OptimisationResult(optim=optim, x=[1e-3], n_iterations=1)
 
         # Asserts
