### 2025-02-18 6aec39d
tests: copies on entry to maintain model state
