+++ tests/unit/test_models.py
@@ -145,31 +145,32 @@ class TestModels:
                 model.predict(t_eval=t_eval, inputs=inputs)
 
     def test_removal_and_apply_events(self, model):
+        mdl = model.copy()
         skipped_models = (
             pybop.lithium_ion.WeppnerHuggins,
             pybop.empirical.Thevenin,
             pybop.lithium_ion.GroupedSPMe,
         )
-        if not isinstance(model, skipped_models):
+        if not isinstance(mdl, skipped_models):
             # Test with already built/solved model and new model instances
-            for mdl in [model, model.new_copy()]:
-                mdl.build()
-                assert not mdl._built_model.events
-                assert mdl.events is not None
+            for _mdl in [mdl, mdl.new_copy()]:
+                _mdl.build()
+                assert not _mdl._built_model.events
+                assert _mdl.events is not None
                 t_eval = np.linspace(0, 1e3, 3)
                 inputs = {
                     "Negative electrode active material volume fraction": 0.5,
                     "Positive electrode active material volume fraction": 0.5,
                 }
                 assert (
-                    len(mdl.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
+                    len(_mdl.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
                     == 3
                 )
 
                 # Reapply events
-                mdl.apply_events()
+                _mdl.apply_events()
                 assert (
-                    len(mdl.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
+                    len(_mdl.simulate(t_eval=t_eval, inputs=inputs)["Voltage [V]"].data)
                     <= 3
                 )
 
