### 2025-02-18 0c83f42
adds _events to model.clear(), updates copy to new_copy
