### 2025-02-18 0a82ade
Add weighting to error measures
