### 2025-02-18 c494042
Update test_cost.py
