+++ tests/unit/test_cost.py
@@ -39,11 +39,7 @@ class TestCosts:
 
     @pytest.fixture
     def experiment(self):
-        return pybop.Experiment(
-            [
-                ("Discharge at 1C for 10 minutes (20 second period)"),
-            ]
-        )
+        return pybop.Experiment(["Discharge at 1C for 10 minutes (20 second period)"])
 
     @pytest.fixture
     def dataset(self, model, experiment):
@@ -56,33 +52,31 @@ class TestCosts:
             }
         )
 
-    @pytest.fixture
-    def signal(self):
-        return "Voltage [V]"
+    def test_base(self, model, parameters, dataset):
+        problem = pybop.FittingProblem(model, parameters, dataset)
+        for cost_class in [pybop.BaseCost, pybop.FittingCost, pybop.DesignCost]:
+            base_cost = cost_class(problem)
+            assert base_cost.problem == problem
+            with pytest.raises(NotImplementedError):
+                base_cost([0.5])
 
     @pytest.fixture(params=[2.5, 3.777])
-    def problem(self, model, parameters, dataset, signal, request):
+    def problem(self, model, parameters, dataset, request):
         cut_off = request.param
         model.parameter_set.update({"Lower voltage cut-off [V]": cut_off})
-        problem = pybop.FittingProblem(model, parameters, dataset, signal=signal)
-        return problem
+        return pybop.FittingProblem(model, parameters, dataset)
 
     @pytest.fixture(
         params=[
-            pybop.RootMeanSquaredError,
-            pybop.SumSquaredError,
-            pybop.Minkowski,
-            pybop.SumOfPower,
+            pybop.MeanAbsoluteError,
             pybop.ObserverCost,
             pybop.LogPosterior,
         ]
     )
-    def cost(self, problem, request):
+    def fitting_cost(self, problem, request):
         cls = request.param
-        if cls in [pybop.SumSquaredError, pybop.RootMeanSquaredError]:
+        if cls is pybop.MeanAbsoluteError:
             return cls(problem)
-        elif cls in [pybop.Minkowski, pybop.SumOfPower]:
-            return cls(problem, p=2)
         elif cls is pybop.LogPosterior:
             return cls(pybop.GaussianLogLikelihoodKnownSigma(problem, sigma0=0.002))
         elif cls is pybop.ObserverCost:
@@ -106,19 +100,12 @@ class TestCosts:
                     process=process,
                     measure=1e-4,
                     dataset=dataset,
-                    signal=problem.signal,
                 ),
             )
 
-    def test_base(self, model, parameters, dataset):
-        problem = pybop.FittingProblem(model, parameters, dataset)
-        for cost_class in [pybop.BaseCost, pybop.FittingCost, pybop.DesignCost]:
-            base_cost = cost_class(problem)
-            assert base_cost.problem == problem
-            with pytest.raises(NotImplementedError):
-                base_cost([0.5])
+    def test_fitting_costs(self, fitting_cost, parameters):
+        cost = fitting_cost
 
-    def test_costs(self, cost, parameters):
         # Test cost direction
         if isinstance(cost, pybop.BaseLikelihood):
             higher_cost = cost([0.52])
@@ -135,43 +122,62 @@ class TestCosts:
         assert np.isscalar(cost(parameters.as_dict()))
 
         # Test UserWarnings
-        if isinstance(cost, (pybop.SumSquaredError, pybop.RootMeanSquaredError)):
+        if isinstance(cost, (pybop.MeanAbsoluteError)):
             assert cost([0.5]) >= 0
-            with pytest.warns(UserWarning) as record:
+            with pytest.warns(UserWarning):
                 cost([1.1])
 
[…]
