### 2025-02-18 0931c5d
Revert "Merge branch 'non-uniform-mean' of https://github.com/pybop-team/PyBOP into non-uniform-mean"
This reverts commit e0b6a5668c2d5001079ff94d99818d7f09a061ee, reversing
changes made to c494042cd9ad11e43939742fb1bb6ec2b26e9163.
