+++ examples/scripts/battery_parameterisation/ocp_averaging.py
@@ -4,52 +4,123 @@ import pybop
 
 # Generate some synthetic data for testing
 parameter_set = pybop.ParameterSet("Chen2020")
-ocp_function = parameter_set["Positive electrode OCP [V]"]
+model = pybop.lithium_ion.SPMe(parameter_set=parameter_set)
 
-
-def noise(sigma):
-    return np.random.normal(0, sigma, 91)
-
-
-discharge_sto = np.linspace(0, 0.9, 91)
-discharge_voltage = ocp_function(discharge_sto + 0.02) + noise(1e-3)
-charge_sto = np.linspace(1, 0.1, 91)
-charge_voltage = ocp_function(charge_sto - 0.02) + noise(1e-3)
-
-# Create the charge and discharge datasets
-discharge_dataset = pybop.Dataset(
-    {"Stoichiometry": discharge_sto, "Voltage [V]": discharge_voltage}
+# Create representative charge and discharge datasets
+discharge_solution = model.predict(
+    initial_state={"Initial SoC": 1},
+    experiment=pybop.Experiment(["Discharge at C/10 until 2.5 V"]),
 )
-charge_dataset = pybop.Dataset(
-    {"Stoichiometry": charge_sto, "Voltage [V]": charge_voltage}
+discharge_dataset_fullcell = pybop.Dataset(
+    {
+        "Stoichiometry": discharge_solution["Negative electrode stoichiometry"].data,
+        "Voltage [V]": discharge_solution["Voltage [V]"].data,
+    }
+)
+discharge_dataset_positive = pybop.Dataset(
+    {
+        "Stoichiometry": discharge_solution["Positive electrode stoichiometry"].data,
+        "Voltage [V]": np.mean(
+            discharge_solution[
+                "Positive electrode surface potential difference [V]"
+            ].data,
+            axis=0,
+        ),
+    }
+)
+charge_dataset_negative = pybop.Dataset(
+    {
+        "Stoichiometry": discharge_solution["Negative electrode stoichiometry"].data,
+        "Voltage [V]": np.mean(
+            discharge_solution[
+                "Negative electrode surface potential difference [V]"
+            ].data,
+            axis=0,
+        ),
+    }
+)
+charge_solution = model.predict(
+    initial_state={"Initial SoC": 0},
+    experiment=pybop.Experiment(["Charge at C/10 until 4.2 V"]),
+)
+charge_dataset_fullcell = pybop.Dataset(
+    {
+        "Stoichiometry": charge_solution["Negative electrode stoichiometry"].data,
+        "Voltage [V]": charge_solution["Voltage [V]"].data,
+    }
+)
+charge_dataset_positive = pybop.Dataset(
+    {
+        "Stoichiometry": charge_solution["Positive electrode stoichiometry"].data,
+        "Voltage [V]": np.mean(
+            charge_solution["Positive electrode surface potential difference [V]"].data,
+            axis=0,
+        ),
+    }
+)
+discharge_dataset_negative = pybop.Dataset(
+    {
+        "Stoichiometry": charge_solution["Negative electrode stoichiometry"].data,
+        "Voltage [V]": np.mean(
+            charge_solution["Negative electrode surface potential difference [V]"].data,
+            axis=0,
+        ),
+    }
 )
 
-# Estimate the shift and generate the average open-circuit potential
-ocp_average = pybop.OCPAverage(
-    discharge_dataset,
-    charge_dataset,
-    allow_stretching=False,
-)
 
-# Create a composite open-circuit potential from charge and discharge
-ocp_blend = pybop.OCPBlend(discharge_dataset, charge_dataset)
+for discharge_dataset, charge_dataset in zip(
+    [
[…]
