### 2025-02-18 f345bc4
Update GITT methods
