### 2025-02-18 ca017bf
Fix gitt_fitting and add current check
