### 2025-02-18 56bc36d
Add tests on custom models
