### 2025-02-18 d5b0750
Add test_gitt_fit
