### 2025-02-18 4989a74
Remove working electrode option
