### 2025-02-20 5617526
Enable variable tau_d and R0
