### 2025-02-20 d97134e
Update gitt parameters and store costs
