### 2025-02-20 980ea78
Fixes #677 - observer quick plotting (#678)
* feat: updates observer.evaluate to align w/ BaseProblem.evaluate

* adds changelog entry
