+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Optimisations
 
+- [#676](https://github.com/pybop-team/PyBOP/pull/676) - Update the format of the problem sensitivities to a dict.
 - [#681](https://github.com/pybop-team/PyBOP/pull/681) - Update the spatial variable defaults of the `GroupedSPMe` model.
 
 ## Bug Fixes
