### 2025-02-25 #7653 @jgostick
Hi Authors, thanks for this submission, which appears to be in pretty good shape.  One problem we need to address though: The paper is WAY too long.  According to the [guidelines](https://joss.readthedocs.io/en/latest/paper.html) it should 250-1000 words.  By my quick count your paper.md file has 2700 words, not including the reference list!  The point of the JOSS paper is the code, not the paper.  The information you have in your paper.md is actually more appropriate for your documentation.  There is another way to think about it: The less your write the paper, the longer it will remain valid as your code evolves.

Anyway, you work on re-writing the paper and I'll start finding reviewers.  If you have any suggestions, please let me know.  The first half-dozen that spring to my mind all have a COI with this paper, so suggestions would be helpful.
