### 2025-02-25 06975d8
Pass sensitivities as dict and remove transpose (#676)
* Change SumofPower to SumOfPower

* Pass dy as dict and update documentation

* Update documentation

* Update documentation

* Update CHANGELOG.md
