### 2025-02-25 e01de2f
Rename plot.quick to plot.problem
