+++ examples/scripts/battery_parameterisation/simple_ecm.py
@@ -87,7 +87,7 @@ parameter_set.export_parameters(
 pybop.plot.dataset(dataset)
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
