+++ examples/scripts/comparison_examples/nelder_mead.py
@@ -65,7 +65,7 @@ optim = pybop.NelderMead(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
