+++ examples/scripts/comparison_examples/gradient_descent.py
@@ -49,7 +49,7 @@ optim = pybop.GradientDescent(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
