+++ examples/scripts/comparison_examples/maximum_a_posteriori.py
@@ -76,7 +76,7 @@ results = optim.run()
 print("True parameters:", parameters.true_value())
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
