+++ examples/scripts/comparison_examples/stochastic_natural_evolution.py
@@ -41,7 +41,7 @@ optim = pybop.SNES(cost, max_iterations=100)
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
