+++ examples/scripts/comparison_examples/adamw.py
@@ -78,7 +78,7 @@ optim.optimiser.b2 = 0.9
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
