+++ examples/scripts/comparison_examples/cuckoo.py
@@ -63,7 +63,7 @@ optim = pybop.Optimisation(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
