+++ examples/scripts/comparison_examples/exponential_decay.py
@@ -59,7 +59,7 @@ optim = pybop.AdamW(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
