+++ examples/scripts/comparison_examples/maximum_likelihood.py
@@ -72,7 +72,7 @@ optim = pybop.IRPropMin(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
