+++ examples/scripts/getting_started/functional_parameters.py
@@ -84,7 +84,7 @@ optim = pybop.SciPyMinimize(cost, sigma0=0.1, max_iterations=125)
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
