+++ examples/scripts/design_optimisation/maximising_energy.py
@@ -84,7 +84,7 @@ print(f"Initial volumetric energy density: {cost2(optim.x0):.2f} Wh.m-3")
 print(f"Optimised volumetric energy density: {cost2(results.x):.2f} Wh.m-3")
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot the cost landscape with optimisation path
 pybop.plot.surface(optim)
