### 2025-03-04 1a9e89f
refactor: EIS methods to mixin class
