+++ pybop/models/eis_methods.py
@@ -10,110 +10,119 @@ from pybop import SymbolReplacer
 from pybop.parameters.parameter import Inputs
 
 
-def set_up_for_eis(model):
+class EISMixin:
     """
-    Set up the model for electrochemical impedance spectroscopy (EIS) simulations.
-
-    This method sets up the model for EIS simulations by adding the necessary
-    algebraic equations and variables to the model.
-    Originally developed by pybamm-eis: https://github.com/pybamm-team/pybamm-eis
-
-    Parameters
-    ----------
-    model : pybamm.Model
-        The PyBaMM model to be used for EIS simulations.
+    A mixin class that adds Electrochemical Impedance Spectroscopy (EIS) functionality
+    to pybop.BaseModel
     """
-    V_cell = pybamm.Variable("Voltage variable [V]")
-    model.variables["Voltage variable [V]"] = V_cell
-    V = model.variables["Voltage [V]"]
 
-    # Add algebraic equation for the voltage
-    model.algebraic[V_cell] = V_cell - V
-    model.initial_conditions[V_cell] = model.param.ocv_init
+    def set_up_for_eis(self):
+        """
+        Set up the model for electrochemical impedance spectroscopy (EIS) simulations.
 
-    # Create the FunctionControl submodel and extract variables
-    external_circuit_variables = pybamm.external_circuit.FunctionControl(
-        model.param, None, model.options, control="algebraic"
-    ).get_fundamental_variables()
+        This method sets up the model for EIS simulations by adding the necessary
+        algebraic equations and variables to the model.
+        Originally developed by pybamm-eis: https://github.com/pybamm-team/pybamm-eis
 
-    # Define the variables to replace
-    symbol_replacement_map = {}
-    for name, variable in external_circuit_variables.items():
-        if name in model.variables.keys():
-            symbol_replacement_map[model.variables[name]] = variable
+        Parameters
+        ----------
+        model : pybamm.Model
+            The PyBaMM model to be used for EIS simulations.
+        """
+        V_cell = pybamm.Variable("Voltage variable [V]")
+        self.pybamm_model.variables["Voltage variable [V]"] = V_cell
+        V = self.pybamm_model.variables["Voltage [V]"]
 
-    # Don't replace initial conditions, as these should not contain
-    # Variable objects
-    replacer = SymbolReplacer(symbol_replacement_map, process_initial_conditions=False)
-    replacer.process_model(model, inplace=True)
+        # Add algebraic equation for the voltage
+        self.pybamm_model.algebraic[V_cell] = V_cell - V
+        self.pybamm_model.initial_conditions[V_cell] = self.pybamm_model.param.ocv_init
 
-    # Add an algebraic equation for the current density variable
-    # External circuit submodels are always equations on the current
-    I_cell = model.variables["Current variable [A]"]
-    I = model.variables["Current [A]"]
-    I_applied = pybamm.FunctionParameter("Current function [A]", {"Time [s]": pybamm.t})
-    model.algebraic[I_cell] = I - I_applied
-    model.initial_conditions[I_cell] = 0
+        # Create the FunctionControl submodel and extract variables
+        external_circuit_variables = pybamm.external_circuit.FunctionControl(
+            self.pybamm_model.param,
+            None,
+            self.pybamm_model.options,
+            control="algebraic",
+        ).get_fundamental_variables()
 
+        # Define the variables to replace
+        symbol_replacement_map = {}
+        for name, variable in external_circuit_variables.items():
+            if name in self.pybamm_model.variables.keys():
+                symbol_replacement_map[self.pybamm_model.variables[name]] = variable
 
-def initialise_eis_simulation(model, inputs: Optional[Inputs] = None):
-    """
[…]
