### 2025-03-05 #7874 @editorialbot
[REVIEW]: PyBOP: A Python package for battery model optimisation and parameterisation
**Submitting author:** @bradyplanden (<a href="http://orcid.org/0000-0002-1082-9125">Brady Planden</a>)
**Repository:** https://github.com/pybop-team/PyBOP
**Branch with paper.md** (empty if default branch): joss
**Version:** v25.11
