### 2025-03-05 #7874 @jgostick
Hi @nickbrady and @leeberty84, thanks very much for agreeing to review this package. There is a link to reviewer guidelines at the top of this thread but here is the short version:

- Create a 'reviewer checklist' for yourself
- Start playing with the package and checking off the boxes
- As you find issues or problems, open issues on the PyBOP repo
- Once you've ticked all the boxes, then your job in done!
