### 2025-03-05 9b62f5c
refactor: restores `t_eval` and `f_eval` args to model.simulate/S1
