### 2025-03-05 cab8f85
tests: adds coverage
