### 2025-03-05 3c3a0fb
refactor: restores SimulateS1 args, adds changelog entry
