+++ pybop/models/base_model.py
@@ -449,7 +449,6 @@ class BaseModel(EISMixin):
         inputs: Inputs,
         t_eval: np.array,
         initial_state: Optional[dict] = None,
-        eis: bool = False,
     ):
         """
         Perform the forward model simulation with sensitivities.
@@ -477,11 +476,6 @@ class BaseModel(EISMixin):
         ValueError
             If the model has not been built before simulation.
         """
-        if eis is True:
-            raise ValueError(
-                "EIS predictions don't currently support gradient information"
-            )
-
         inputs = self.parameters.verify(inputs)
 
         if initial_state is not None or any(
