### 2025-03-05 fdd51b8
tests: reduces corruption in map estimate
