+++ tests/integration/test_monte_carlo_thevenin.py
@@ -127,7 +127,7 @@ class TestSamplingThevenin:
         ],
     )
     def test_sampling_thevenin(self, sampler, posterior, map_estimate):
-        x0 = np.clip(map_estimate + np.random.normal(0, 1e-3, size=2), 1e-4, 1e-1)
+        x0 = np.clip(map_estimate + np.random.normal(0, 5e-4, size=2), 1e-4, 1e-1)
         common_args = {
             "log_pdf": posterior,
             "chains": 2,
