### 2025-03-13 bfd5bd6
tests: extends SliceRankShrinking iterations
