### 2025-03-18 #7874 @jgostick
Hi @BradyPlanden, I have secured 2 reviewers for your submission, and have pinged them both via email this evening to ask how things are progressing.  While we away word from them, can you please address my earlier request to shorten your paper substantially?  It should talk about the generate point of the software and how it relates to other options, and maybe a bit about the design (in the context how it is better/different that other works), but it should not have anything like examples or documentation.  So get started re-working that while the reviewers are playing with the code itself.
