### 2025-03-18 #7874 @BradyPlanden
Hi @jgostick, thanks for organising and progressing the reviews. Yes, we will start iterating on the article to get it closer to the recommended submission size. Do let me know if you need anything further while we await the reviews.
