### 2025-03-18 63a2d92
feat: improvements to the BaseSampler and BasePintsSampler classes. Fixes incorrect attribute access between child/parent
