+++ pybop/samplers/base_sampler.py
@@ -1,4 +1,3 @@
-import logging
 from typing import Union
 
 import numpy as np
@@ -9,27 +8,29 @@ from pybop import CostInterface, LogPosterior
 
 class BaseSampler(CostInterface):
     """
-    Base class for Monte Carlo samplers. Samplers perform maximisation of the cost
-    function by default.
+    Base class for Monte Carlo samplers.
+
+    Parameters
+    ----------
+    log_pdf : pybop.LogPosterior or List[pybop.LogPosterior]
+        The posterior or PDF to be sampled.
+    chains : int
+        Number of chains to be used.
+    x0
+        List-like initial values of the parameters for Monte Carlo sampling.
+    cov0
+        The covariance matrix to be sampled.
+
+    Note: Samplers perform maximisation of the Posterior by default.
     """
 
     def __init__(
-        self, log_pdf: LogPosterior, x0, chains: int, cov0: Union[np.ndarray, float]
+        self,
+        log_pdf: Union[LogPosterior, list[LogPosterior]],
+        x0,
+        chains: int,
+        cov0: Union[np.ndarray, float],
     ):
-        """
-        Initialise the base sampler.
-
-        Parameters
-        ----------
-        log_pdf : pybop.LogPosterior or List[pybop.LogPosterior]
-            The posterior or PDF to be sampled.
-        chains : int
-            Number of chains to be used.
-        x0
-            List-like initial values of the parameters for Monte Carlo sampling.
-        cov0
-            The covariance matrix to be sampled.
-        """
         self._log_pdf = log_pdf
         self._cov0 = cov0
 
@@ -56,12 +57,14 @@ class BaseSampler(CostInterface):
         super().__init__(transformation=transformation)
 
         # Check initial conditions
-        if x0 is not None and len(x0) != self.n_parameters:
-            raise ValueError("x0 must have the same number of parameters as log_pdf")
-
-        # Set initial values, if x0 is None, initial values are unmodified
         if x0 is not None:
+            if len(x0) != self.n_parameters:
+                raise ValueError(
+                    "x0 must have the same number of parameters as log_pdf"
+                )
             self.parameters.update(initial_values=x0)
+
+        # Update x0 w/ transformation if applicable - reshape to align with chains
         self._x0 = self.parameters.reset_initial_value(apply_transform=True).reshape(
             1, -1
         )
@@ -111,87 +114,11 @@ class BaseSampler(CostInterface):
             If True, use as many worker processes as there are CPU cores. If an integer, use that many workers.
             If False or 0, disable parallelism (default: False).
         """
+        self._parallel = bool(parallel is True or parallel >= 1)
+
         if parallel is True:
-            self._parallel = True
             self._n_workers = ParallelEvaluator.cpu_count()
         elif parallel >= 1:
-            self._parallel = True
             self._n_workers = int(parallel)
         else:
-            self._parallel = False
             self._n_workers = 1
-
-    def _ask_for_samples(self):
-        if self._single_chain:
-            return [self._samplers[i].ask() for i in self._active]
-        else:
-            return self._samplers[0].ask()
-
-    def _check_initial_phase(self):
-        # Set initial phase if needed
-        if self._initial_phase:
-            for sampler in self._samplers:
-                sampler.set_initial_phase(True)
-
-    def _end_initial_phase(self):
-        for sampler in self._samplers:
-            sampler.set_initial_phase(False)
-        if self._log_to_screen:
-            logging.info("Initial phase completed.")
-
-    def _initialise_storage(self):
-        self._prior = None
-        if isinstance(self._log_pdf, LogPosterior):
-            self._prior = self._log_pdf.prior
-
-        # Storage of the received samples
-        self._sampled_logpdf = np.zeros(self._n_chains)
[…]
