+++ pybop/samplers/base_pints_sampler.py
@@ -1,7 +1,7 @@
 import logging
 import time
 from functools import partial
-from typing import Optional
+from typing import Optional, Union
 
 import numpy as np
 from pints import (
@@ -11,7 +11,13 @@ from pints import (
     SingleChainMCMC,
 )
 
-from pybop import BaseCost, BaseSampler, LogPosterior
+from pybop import (
+    BaseCost,
+    BaseSampler,
+    LogPosterior,
+    MultiChainProcessor,
+    SingleChainProcessor,
+)
 
 
 class BasePintsSampler(BaseSampler):
@@ -21,11 +27,28 @@ class BasePintsSampler(BaseSampler):
     This class extends the BaseSampler class to provide a common interface for
     PINTS samplers. The class provides a sample() method that can be used to
     sample from the posterior distribution using a PINTS sampler.
+
+    Parameters
+    ----------
+    log_pdf : pybop.LogPosterior or List[pybop.LogPosterior] or callable
+        An object to be sampled, which can be either a pybop.LogPosterior or callable function.
+    sampler : pybop.Sampler
+        The sampling algorithm to use.
+    chains : int
+        Number of chains to run concurrently. Each chain contains separate markov samples from
+        the log_pdf.
+    x0 : numpy.ndarray
+        Initial values of the parameters for the optimisation.
+    cov0 : list-like
+        Initial covariance for the chains in the parameters. Either a scalar value
+        (same for all coordinates) or an array with one entry per dimension.
+    kwargs
+        Additional keyword arguments.
     """
 
     def __init__(
         self,
-        log_pdf: LogPosterior,
+        log_pdf: Union[LogPosterior, list[LogPosterior]],
         sampler,
         chains: int = 1,
         warm_up=None,
@@ -33,24 +56,6 @@ class BasePintsSampler(BaseSampler):
         cov0=0.1,
         **kwargs,
     ):
-        """
-        Initialise the base PINTS sampler.
-
-        Parameters
-        ----------
-        log_pdf : pybop.LogPosterior or List[pybop.LogPosterior]
-            The distribution(s) to be sampled.
-        chains : int
-            Number of chains to be used.
-        sampler
-            The sampler class to be used.
-        x0 : list
-            Initial states for the chains.
-        cov0
-            Initial standard deviation for the chains.
-        kwargs
-            Additional keyword arguments.
-        """
         super().__init__(log_pdf, x0, chains, cov0)
 
         # Set kwargs
@@ -64,6 +69,7 @@ class BasePintsSampler(BaseSampler):
         self._parallel = kwargs.get("parallel", False)
         self._verbose = kwargs.get("verbose", False)
         self.sampler = sampler
+        self._prior = None
         self.iter_time = float(0)
         self._iteration = 0
         self._loop_iters = 0
@@ -106,9 +112,17 @@ class BasePintsSampler(BaseSampler):
         if self._initial_phase:
             self.set_initial_phase_iterations()
 
-        # Parallelisation (Might be able to move into parent class)
-        self._n_workers = 1
+        # Set parallelisation
         self.set_parallel(self._parallel)
+
+    def _initialize_chain_processor(self):
+        """
+        Initialise the appropriate chain processor based on configuration.
+        """
+        if self._single_chain:
+            self._chain_processor = SingleChainProcessor(self)
+        else:
+            self._chain_processor = MultiChainProcessor(self)
 
     def run(self) -> Optional[np.ndarray]:
         """
@@ -148,6 +162,7 @@ class BasePintsSampler(BaseSampler):
 
         self._initialise_logging()
         self._check_stopping_criteria()
+        self._initialize_chain_processor()
 
[…]
