### 2025-03-24 9bd42e8
Enable non-stationary EIS
