### 2025-03-24 050fc2a
Update current when given time
