### 2025-03-24 3b9833e
refactor: align store_samples method, update property attributes
