### 2025-03-24 b8c6fa6
refactor: addition property alignment, code simplication
