### 2025-03-24 c02e9c6
tests: adds set_parallel coverage for pints optimisers
