+++ pybop/optimisers/base_pints_optimiser.py
@@ -397,14 +397,13 @@ class BasePintsOptimiser(BaseOptimiser):
             If True, use as many worker processes as there are CPU cores. If an integer, use that many workers.
             If False or 0, disable parallelism (default: False).
         """
+        self._parallel = bool(parallel is True or parallel >= 1)
+
         if parallel is True:
-            self._parallel = True
             self._n_workers = PintsParallelEvaluator.cpu_count()
         elif parallel >= 1:
-            self._parallel = True
             self._n_workers = int(parallel)
         else:
-            self._parallel = False
             self._n_workers = 1
 
     def set_max_iterations(self, iterations="default"):
