### 2025-03-24 ec76b73
refactor: removes unused property, formatting
