### 2025-03-24 1dceb82
fix: clearer logic for intermediate sampler steps
