### 2025-03-25 26b50c0
Add finite difference check
