### 2025-03-25 1bad045
Enable domain-based means (#659)
* Add weighting to error measures

* Update test_cost.py

* Tune test_error_weighting

* Add brackets for clarity

* Update CHANGELOG.md

* Add finite difference check
