+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- [#659](https://github.com/pybop-team/PyBOP/pull/659) - Enables user-defined weightings of the error measures.
 - [#674](https://github.com/pybop-team/PyBOP/issues/674) - Adds the reason for stopping to the `OptimisationResult`.
 
 ## Optimisations
