### 2025-03-25 e249361
refactor: removes build_on_evaluation arg for `model.set_initial_condition`
