+++ examples/scripts/comparison_examples/irpropmin.py
@@ -33,8 +33,10 @@ parameters = pybop.Parameters(
     ),
 )
 
-# Import the synthetic dataset
+# Import the synthetic dataset, set model initial state
 csv_data = np.loadtxt(dataset_path, delimiter=",", skiprows=1)
+initial_state = {"Initial open-circuit voltage [V]": csv_data[0, 2]}
+model.set_initial_state(initial_state=initial_state)
 
 # Form dataset
 dataset = pybop.Dataset(
@@ -50,7 +52,6 @@ problem = pybop.FittingProblem(
     model,
     parameters,
     dataset,
-    initial_state={"Initial open-circuit voltage [V]": csv_data[0, 2]},
 )
 likelihood = pybop.GaussianLogLikelihoodKnownSigma(problem, sigma0=2e-3 * 1.05)
 posterior = pybop.LogPosterior(likelihood)
