+++ examples/scripts/comparison_examples/adamw.py
@@ -32,8 +32,10 @@ parameters = pybop.Parameters(
     ),
 )
 
-# Import the synthetic dataset
+# Import the synthetic dataset, set model initial state
 csv_data = np.loadtxt(dataset_path, delimiter=",", skiprows=1)
+initial_state = {"Initial open-circuit voltage [V]": csv_data[0, 2]}
+model.set_initial_state(initial_state=initial_state)
 
 # Form dataset
 dataset = pybop.Dataset(
@@ -52,7 +54,6 @@ problem = pybop.FittingProblem(
     parameters,
     dataset,
     signal=signal,
-    initial_state={"Initial open-circuit voltage [V]": csv_data[0, 2]},
 )
 cost = pybop.SumOfPower(problem, p=2.5)
 
