### 2025-03-25 df79ec2
docs: Update FittingProblem docstring
