### 2025-03-26 529527d
fix: converts `physical_viability` arg to `check_physical_viability` w/ kwarg support
