### 2025-03-26 b0dd16b
docs: updates BasePintsSampler docstring
