### 2025-03-26 d67654b
MCMC Sampler Improvements (#692)
* feat: improvements to the BaseSampler and BasePintsSampler classes. Fixes incorrect attribute access between child/parent

* fix: removes dangling code, aligns method name

* test: adds coverage

* refactor: align store_samples method, update property attributes

* refactor: addition property alignment, code simplication

* tests: adds set_parallel coverage for pints optimisers

* refactor: removes unused property, formatting

* fix: clearer logic for intermediate sampler steps

* docs: updates BasePintsSampler docstring
