### 2025-03-27 bc0ec62
refactor: adds print frequency optional arg, optim.log as dataclass
