### 2025-03-27 bc1bb22
test: adds coverage, update optimisation_options integration test
