### 2025-03-27 03e37c3
refactor: renames CostInterface._inverted_cost to _inverts_cost
