+++ pybop/optimisers/_cost_interface.py
@@ -29,7 +29,7 @@ class CostInterface:
             return [self._transformation.to_model(values) for values in list_of_values]
         return list_of_values
 
-    def _inverted_cost(self, cost):
+    def _inverts_cost(self, cost):
         """
         Returns the true cost if the optimiser is operating in the inverted space
         else returns the cost as is.
