+++ tests/unit/test_cost.py
@@ -251,6 +251,9 @@ class TestCosts:
         np.testing.assert_allclose(e, eR, rtol=1e-2, atol=1e-9)
         np.testing.assert_allclose(de, deR, rtol=1e-2, atol=1e-9)
 
+        # Check that the sum (and therefore mean) are the same as an even weighting
+        np.testing.assert_allclose(np.sum(costR.weighting), len(costR.weighting))
+
         # Check gradient calculation using finite difference
         delta = 1e-6 * x[0]
         cost_right = costR(x[0] + delta / 2)
