### 2025-03-28 b10e042
examples: adds `verbose=True` to key getting_started examples
