### 2025-03-28 ef29f92
Verbose optimiser output (#650)
* feat: adds verbose printing for pints' optimisers

* refactor: adds evaluations, changelog entry

* Remove missed check after refactor, add to docstring

* refactor: adds print frequency optional arg, optim.log as dataclass

* test: adds coverage, update optimisation_options integration test

* refactor: renames CostInterface._inverted_cost to _inverts_cost

* docs: updates for release_workflow.md

* examples: adds `verbose=True` to key getting_started examples
