+++ pybop/optimisers/base_optimiser.py
@@ -1,5 +1,6 @@
 import warnings
 from copy import deepcopy
+from dataclasses import dataclass, field
 
 import jax.numpy as jnp
 import numpy as np
@@ -11,6 +12,20 @@ from pybop import (
     Parameter,
     Parameters,
 )
+
+
+@dataclass
+class OptimisationLog:
+    """Stores optimisation progress data."""
+
+    iterations: list[int] = field(default_factory=list)
+    evaluations: list[int] = field(default_factory=list)
+    x: list[list[float]] = field(default_factory=list)
+    x_best: list[list[float]] = field(default_factory=list)
+    x_search: list[list[float]] = field(default_factory=list)
+    x0: list[list[float]] = field(default_factory=list)
+    cost: list[float] = field(default_factory=list)
+    cost_best: list[float] = field(default_factory=list)
 
 
 class BaseOptimiser(CostInterface):
@@ -38,7 +53,9 @@ class BaseOptimiser(CostInterface):
             (same for all coordinates) or an array with one entry per dimension.
             Not all methods will use this information.
         verbose : bool, optional
-            If True, the optimisation progress is printed (default: False).
+            If True, the optimisation progress and final result is printed (default: False).
+        verbose_print_rate : int, optional
+            The frequency in iterations to print the optimisation progress (default: 50).
         physical_viability : bool, optional
             If True, the feasibility of the optimised parameters is checked (default: False).
         allow_infeasible_solutions : bool, optional
@@ -58,15 +75,17 @@ class BaseOptimiser(CostInterface):
         # First set attributes to default values
         self.parameters = Parameters()
         self.x0 = optimiser_kwargs.get("x0", None)
-        self.log = dict(x=[], x_best=[], x_search=[], cost=[], cost_best=[])
         self.bounds = None
         self.sigma0 = 0.02
-        self.verbose = True
+        self.verbose = False
+        self.verbose_print_rate = 50
         self._needs_sensitivities = False
         self.physical_viability = False
         self.allow_infeasible_solutions = False
         self.default_max_iterations = 1000
         self.result = None
+        self._iter_count = 0
+        self.log = OptimisationLog()
         transformation = None
         invert_cost = False
 
@@ -145,6 +164,9 @@ class BaseOptimiser(CostInterface):
 
         # Set other options
         self.verbose = self.unset_options.pop("verbose", self.verbose)
+        self.verbose_print_rate = self.unset_options.pop(
+            "verbose_print_rate", self.verbose_print_rate
+        )
         if "allow_infeasible_solutions" in self.unset_options.keys():
             self.set_allow_infeasible_solutions(
                 self.unset_options.pop("allow_infeasible_solutions")
@@ -224,56 +246,92 @@ class BaseOptimiser(CostInterface):
 
         return self.cost.sensitivity_analysis(self.n_samples_sensitivity)
 
-    def log_update(self, x=None, x_best=None, cost=None, cost_best=None):
+    def log_update(
+        self,
+        iterations=None,
+        evaluations=None,
+        x=None,
+        x_best=None,
+        cost=None,
+        cost_best=None,
+        x0=None,
+    ):
         """
         Update the log with new values.
 
         Parameters
         ----------
+        iterations : list or array-like, optional
+            Iteration indices to log (default: None).
+        evaluations: list or array-like, optional
+            Evaluation indices to log (default: None).
         x : list or array-like, optional
             Parameter values (default: None).
         x_best : list or array-like, optional
             Parameter values corresponding to the best cost yet (default: None).
         cost : list, optional
             Cost values corresponding to x (default: None).
-        cost_best
+        cost_best : list, optional
             Cost values corresponding to x_best (default: None).
[…]
