### 2025-03-28 7e4d250
Rename plot.quick to plot.problem (#684)
* Rename plot.quick to plot.problem

* Update CHANGELOG.md

* Update missed examples

* Align new examples
