+++ examples/scripts/getting_started/simple_dfn.py
@@ -70,7 +70,7 @@ optim = pybop.IRPropPlus(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
