+++ examples/scripts/getting_started/weighted_cost.py
@@ -63,7 +63,7 @@ for cost in [weighted_cost, cost1, cost2]:
     print("True parameters:", parameters.true_value())
 
     # Plot the timeseries output
-    pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+    pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
     # Plot convergence
     pybop.plot.convergence(optim)
