+++ examples/scripts/comparison_examples/simulated_annealing.py
@@ -55,7 +55,7 @@ optim.optimiser.cooling_rate = 0.8
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
