+++ examples/scripts/comparison_examples/scipy_minimize.py
@@ -50,7 +50,7 @@ optim = pybop.SciPyMinimize(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)
