+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pybop"
-version = "25.1"
+version = "25.3"
 authors = [
   {name = "The PyBOP Team"},
 ]
