### 2025-04-01 5cb5c7a
refactor: draft out Problem and PythonProblem classes
