### 2025-04-02 f14b06f
refactor: add pybamm pipeline class for easy rebuild and solve, fill out pybamm problem class
