### 2025-04-02 382246f
refactor(PybammProblem): rebuild on set_params
