### 2025-04-02 cccde1c
refactor(PythonProblem): fix check and store function name
