### 2025-04-02 4fbe4b7
refactor: add Problem, PythonProblem, PybammProblem & PybammPipeline (#702)
* refactor: draft out Problem and PythonProblem classes

* refactor: add pybamm pipeline class for easy rebuild and solve, fill out pybamm problem class

* style: pre-commit fixes

* refactor(PybammProblem): rebuild on set_params

* refactor(PythonProblem): fix check and store function name

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
