+++ pybop/problems/base_problem.py
@@ -1,209 +1,76 @@
-from typing import Optional
-
 import numpy as np
-from pybamm import IDAKLUSolver
-
-from pybop import BaseModel, Dataset, Parameter, Parameters
-from pybop.parameters.parameter import Inputs
 
 
-class BaseProblem:
+class Problem:
     """
-    Base class for defining a problem within the PyBOP framework, compatible with PINTS.
-
-    Parameters
-    ----------
-    parameters : pybop.Parameter or pybop.Parameters
-        An object or list of the parameters for the problem.
-    model : object, optional
-        The model to be used for the problem (default: None).
-    check_model : bool, optional
-        Flag to indicate if the model should be checked (default: True).
-    signal: list[str], optional
-        A list of variables to analyse (default: ["Voltage [V]"]).
-    domain : str, optional
-        The name of the domain (default: "Time [s]").
-    additional_variables : list[str], optional
-        Additional variables to observe and store in the solution (default: []).
-    initial_state : dict, optional
-        A valid initial state (default: None).
+    Defines the a callable function `f(x)` that returns the evaluation
+    of a cost function. The input `x` is a set of parameters that are
+    passed via the `set_params` method. The cost function is evaluated
+    using the `run` method.
     """
 
-    def __init__(
+    def __init__(self, param_names: list[str] = None):
+        if param_names is None:
+            param_names = []
+        self._params = None
+        self._param_names = param_names
+
+    def check_and_store_params(self, p: np.ndarray) -> None:
+        """
+        Checks if the parameters are valid.
+        """
+        if len(p) != len(self._param_names):
+            raise ValueError(
+                f"Expected {len(self._param_names)} parameters, but got {len(p)}."
+            )
+        if not isinstance(p, np.ndarray):
+            raise TypeError("Parameters must be a numpy array.")
+        if not np.issubdtype(p.dtype, np.number):
+            raise TypeError("Parameters must be a numeric numpy array.")
+        self._params = p
+
+    def check_set_params_called(self) -> None:
+        """
+        Checks if the parameters have been set.
+        """
+        if self._params is None:
+            raise ValueError(
+                "Parameters have not been set. Call `set_params` before running the simulation."
+            )
+
+    @property
+    def params(self) -> np.ndarray:
+        """
+        Returns the parameters set for the simulation and cost function.
+        """
+        return self._params
+
+    @property
+    def param_names(self) -> list[str]:
+        """
+        Returns the names of the parameters set for the simulation and cost function.
+        """
+        return self._param_names
+
+    def run(self) -> float:
+        """
+        Evaluates the underlying simulation and cost function using the
+        parameters set in the previous call to `set_params`.
+        """
+        raise NotImplementedError
+
+    def run_with_sensitivities(
         self,
-        parameters: Parameters,
-        model: Optional[BaseModel] = None,
-        check_model: bool = True,
-        signal: Optional[list[str]] = None,
-        domain: Optional[str] = None,
-        additional_variables: Optional[list[str]] = None,
-        initial_state: Optional[dict] = None,
-    ):
-        signal = signal or ["Voltage [V]"]
-        if isinstance(signal, str):
-            signal = [signal]
-        elif not all(isinstance(item, str) for item in signal):
-            raise ValueError("Signal should be either a string or list of strings.")
-
-        # Check if parameters is a list of pybop.Parameter objects
-        if isinstance(parameters, list):
-            if all(isinstance(param, Parameter) for param in parameters):
-                parameters = Parameters(*parameters)
-            else:
-                raise TypeError(
-                    "All elements in the list must be pybop.Parameter objects."
-                )
-        # Check if parameters is a single pybop.Parameter object
[…]
