### 2025-04-02 5692a4a
feat: initial builders, updates to problem
