### 2025-04-03 f551082
Fix grouped conductivities
