### 2025-04-03 c8b79fd
Fix grouped conductivities (#704)
* Fix grouped conductivities

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
