### 2025-04-03 bd5a03c
Update plotting in new examples
