### 2025-04-04 be8643b
feat: bugfix fitting_problem output for calc_grad under fail conditions, adds fixed numpy random seed for pytest framework, updates to IDAKLU where beneficial
