+++ tests/integration/test_thevenin_parameterisation.py
@@ -32,7 +32,8 @@ class TestTheveninParameterisation:
                 "R1 [Ohm]": self.ground_truth[1],
             }
         )
-        return pybop.empirical.Thevenin(parameter_set=parameter_set)
+        solver = IDAKLUSolver()
+        return pybop.empirical.Thevenin(parameter_set=parameter_set, solver=solver)
 
     @pytest.fixture
     def parameters(self):
@@ -78,11 +79,10 @@ class TestTheveninParameterisation:
             (pybop.PSO, ""),
         ],
     )
-    def test_optimisers_on_simple_model(
+    def test_optimisers_on_thevenin_model(
         self, model, parameters, dataset, cost_class, optimiser, method
     ):
         # Define the cost to optimise
-        model.solver = IDAKLUSolver()
         problem = pybop.FittingProblem(model, parameters, dataset)
         cost = cost_class(problem)
 
@@ -127,5 +127,5 @@ class TestTheveninParameterisation:
                 ),
             ]
         )
-        sim = model.predict(experiment=experiment)
-        return sim
+        sol = model.predict(experiment=experiment)
+        return sol
