+++ tests/integration/test_model_experiment_changes.py
@@ -110,17 +110,19 @@ class TestModelAndExperimentChanges:
 
     def test_multi_fitting_problem(self):
         parameter_set = pybop.ParameterSet("Chen2020")
-        parameters = pybop.Parameter(
-            "Negative electrode active material volume fraction",
-            prior=pybop.Gaussian(0.68, 0.05),
-            true_value=parameter_set[
-                "Negative electrode active material volume fraction"
-            ],
+        parameters = pybop.Parameters(
+            pybop.Parameter(
+                "Negative electrode active material volume fraction",
+                prior=pybop.Gaussian(0.68, 0.05),
+                true_value=parameter_set[
+                    "Negative electrode active material volume fraction"
+                ],
+            )
         )
 
         model_1 = pybop.lithium_ion.SPM(parameter_set=parameter_set)
         experiment_1 = pybop.Experiment(
-            ["Discharge at 1C until 3 V (4 seconds period)"]
+            ["Discharge at 0.5C for 5 minutes (10 seconds period)"]
         )
         solution_1 = model_1.predict(experiment=experiment_1)
         dataset_1 = pybop.Dataset(
@@ -133,7 +135,7 @@ class TestModelAndExperimentChanges:
 
         model_2 = pybop.lithium_ion.SPMe(parameter_set=parameter_set.copy())
         experiment_2 = pybop.Experiment(
-            ["Discharge at 3C until 3 V (4 seconds period)"]
+            ["Discharge at 1C for 3 minutes (10 seconds period)"]
         )
         solution_2 = model_2.predict(experiment=experiment_2)
         dataset_2 = pybop.Dataset(
@@ -152,7 +154,9 @@ class TestModelAndExperimentChanges:
 
         # Test with a gradient and non-gradient-based optimiser
         for optimiser in [pybop.SNES, pybop.IRPropMin]:
-            optim = optimiser(cost)
+            optim = optimiser(
+                cost, sigma0=0.05, max_iterations=100, max_unchanged_iterations=30
+            )
             results = optim.run()
-            np.testing.assert_allclose(results.x, parameters.true_value, atol=2e-5)
-            np.testing.assert_allclose(results.final_cost, 0, atol=2e-5)
+            np.testing.assert_allclose(results.x, parameters.true_value(), atol=2e-5)
+            np.testing.assert_allclose(results.final_cost, 0, atol=3e-5)
