### 2025-04-04 fa151c3
CI: removes macos-13 examples, duplicate macos-14 integration
