### 2025-04-04 5f13c46
refactor: fitting_problem.py updates, add changelog entry, removes macos-13 integration from CI
