+++ .github/workflows/test_on_pull_request.yaml
@@ -37,7 +37,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-13]
+        os: [ubuntu-latest, windows-latest]
         python-version: ["3.12"]
 
     name: Integration tests (${{ matrix.os }} / Python ${{ matrix.python-version }})
