### 2025-04-04 de4a5c9
i703: fixes for pybamm v25.4 (#705)
* feat: bugfix fitting_problem output for calc_grad under fail conditions, adds fixed numpy random seed for pytest framework, updates to IDAKLU where beneficial

* CI: removes macos-13 examples, duplicate macos-14 integration

* refactor: fitting_problem.py updates, add changelog entry, removes macos-13 integration from CI
