+++ tests/integration/test_transformation.py
@@ -1,6 +1,7 @@
 import itertools
 
 import numpy as np
+import pybamm
 import pytest
 
 import pybop
@@ -46,7 +47,8 @@ class TestTransformation:
                 "R1 [Ohm]": self.ground_truth[1],
             }
         )
-        return pybop.empirical.Thevenin(parameter_set=parameter_set)
+        solver = pybamm.IDAKLUSolver()
+        return pybop.empirical.Thevenin(parameter_set=parameter_set, solver=solver)
 
     @pytest.fixture
     def parameters(self, transformation_r0, transformation_r1):
