### 2025-04-07 68d602a
feat: base builder as ABC, updates for problem API, adds pybamm_cost
