### 2025-04-07 c46d71a
feat: adds concurrent simulations to problem builder, corresponding updates to pybamm_problem.py
