### 2025-04-08 59b5c17
tests, examples: performance improvements
