+++ examples/scripts/getting_started/mcmc_example.py
@@ -17,7 +17,7 @@ parameter_set.update(
         "Positive electrode active material volume fraction": 0.71,
     }
 )
-synth_model = pybop.lithium_ion.DFN(parameter_set=parameter_set, solver=solver)
+synth_model = pybop.lithium_ion.SPMe(parameter_set=parameter_set, solver=solver)
 
 # Fitting parameters
 parameters = pybop.Parameters(
@@ -38,7 +38,7 @@ init_soc = 0.5
 sigma = 0.005
 experiment = pybop.Experiment(
     [
-        ("Discharge at 0.5C for 6 minutes (5 second period)",),
+        ("Discharge at 0.5C for 3 minutes (5 second period)",),
     ]
 )
 values = synth_model.predict(
@@ -56,14 +56,11 @@ dataset = pybop.Dataset(
         "Time [s]": values["Time [s]"].data,
         "Current function [A]": values["Current [A]"].data,
         "Voltage [V]": noisy(values["Voltage [V]"].data, sigma),
-        "Bulk open-circuit voltage [V]": noisy(
-            values["Bulk open-circuit voltage [V]"].data, sigma
-        ),
     }
 )
 
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=pybamm.IDAKLUSolver())
-signal = ["Voltage [V]", "Bulk open-circuit voltage [V]"]
+signal = ["Voltage [V]"]
 
 # Generate problem, likelihood, and sampler
 problem = pybop.FittingProblem(
@@ -75,7 +72,7 @@ posterior = pybop.LogPosterior(likelihood)
 sampler = pybop.DifferentialEvolutionMCMC(
     posterior,
     chains=3,
-    max_iterations=300,
+    max_iterations=250,  # Reduced for CI, increase for improved posteriors
     warm_up=100,
     verbose=True,
     parallel=parallel,  # (macOS/WSL/Linux only)
