### 2025-04-11 a2bb84a
Reduce integration test and example runtimes (#709)
* tests, examples: performance improvements

* notebooks: performance updates
