### 2025-04-23 74e4d6d
fix: removes default solver Casadi args, adds catch for EIS and applies Casadi solver.
