### 2025-04-23 70baa83
examples: removes IDAKLU explicit definition, adds Casadi for unscented_kalman_filter.py
