+++ examples/scripts/comparison_examples/scipy_minimize.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pybamm
 
 import pybop
 
@@ -35,7 +34,6 @@ dataset = pybop.Dataset(
 )
 
 # Generate problem, cost function, and optimisation class
-model.solver = pybamm.IDAKLUSolver()
 problem = pybop.FittingProblem(model, parameters, dataset)
 cost = pybop.SumSquaredError(problem)
 optim = pybop.SciPyMinimize(
