### 2025-04-23 2d4dcb1
tests: removes general explicit IDAKLU definition, adds increased tolerances where required, removes redundant non-convergence test
