+++ tests/integration/test_monte_carlo_thevenin.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pybamm
 import pytest
 
 import pybop
@@ -52,8 +51,7 @@ class TestSamplingThevenin:
                 "R1 [Ohm]": self.ground_truth[1],
             }
         )
-        solver = pybamm.IDAKLUSolver()
-        return pybop.empirical.Thevenin(parameter_set=parameter_set, solver=solver)
+        return pybop.empirical.Thevenin(parameter_set=parameter_set)
 
     @pytest.fixture
     def parameters(self):
