+++ tests/unit/test_models.py
@@ -442,36 +442,6 @@ class TestModels:
             ):
                 model.check_params(inputs=["unexpected_string"])
 
-    def test_non_converged_solution(self):
-        model = pybop.lithium_ion.DFN()
-        parameters = pybop.Parameters(
-            pybop.Parameter(
-                "Negative electrode active material volume fraction",
-                prior=pybop.Gaussian(0.2, 0.01),
-            ),
-            pybop.Parameter(
-                "Positive electrode active material volume fraction",
-                prior=pybop.Gaussian(0.2, 0.01),
-            ),
-        )
-        dataset = pybop.Dataset(
-            {
-                "Time [s]": np.linspace(0, 100, 100),
-                "Current function [A]": np.zeros(100),
-                "Voltage [V]": np.zeros(100),
-            }
-        )
-        problem = pybop.FittingProblem(model, parameters=parameters, dataset=dataset)
-
-        # Simulate the DFN with active material values of 0.
-        # The solution elements will not change as the solver will not converge.
-        output = problem.evaluate([0, 0])
-        output_S1, _ = problem.evaluateS1([0, 0])
-
-        for key in problem.signal:
-            assert np.allclose(output.get(key, [])[0], output.get(key, []))
-            assert np.allclose(output_S1.get(key, [])[0], output_S1.get(key, []))
-
     def test_set_initial_state(self):
         t_eval = np.linspace(0, 10, 100)
 
