### 2025-04-23 08913a9
coverage: adds explicit solver arg for ExponentialDecay
