### 2025-04-24 3bc298d
infra: update nox sessions to upgrade dependencies by default
