+++ noxfile.py
@@ -14,7 +14,7 @@ PYBAMM_VERSION = os.environ.get("PYBAMM_VERSION", None)
 
 @nox.session
 def unit(session):
-    session.install("-e", ".[all,dev]", silent=False)
+    session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     if PYBOP_SCHEDULED:
         session.run("pip", "install", f"pybamm=={PYBAMM_VERSION}", silent=False)
     session.run("pytest", "--unit")
@@ -22,7 +22,7 @@ def unit(session):
 
 @nox.session
 def coverage(session):
-    session.install("-e", ".[all,dev]", silent=False)
+    session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     session.install("pip")
     if PYBOP_SCHEDULED:
         session.run("pip", "install", f"pybamm=={PYBAMM_VERSION}", silent=False)
@@ -42,7 +42,7 @@ def coverage(session):
 @nox.session
 def plots(session):
     """Run the tests that generate plots."""
-    session.install("-e", ".[plot,dev]", silent=False)
+    session.install("-e", ".[plot,dev]", "--upgrade", silent=False)
     session.install("pip")
     session.run("pytest", "--plots", "-n", "0")
 
@@ -50,14 +50,14 @@ def plots(session):
 @nox.session
 def integration(session):
     """Run the integration tests."""
-    session.install("-e", ".[all,dev]", silent=False)
+    session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     session.run("pytest", "--integration")
 
 
 @nox.session
 def examples(session):
     """Run the examples and notebooks"""
-    session.install("-e", ".[all,dev]", silent=False)
+    session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     session.run("pytest", "--examples")
     notebooks(session)
 
@@ -66,7 +66,7 @@ def examples(session):
 def notebooks(session):
     """Run the Jupyter notebooks."""
     session.install("openpyxl", "ipywidgets")
-    session.install("-e", ".[all,dev]", silent=False)
+    session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     if PYBOP_SCHEDULED:
         session.run("pip", "install", f"pybamm=={PYBAMM_VERSION}", silent=False)
     session.run(
@@ -82,7 +82,7 @@ def notebooks(session):
 def notebooks_overwrite(session):
     """Run the Jupyter notebooks."""
     session.install("openpyxl", "ipywidgets")
-    session.install("-e", ".[all,dev]", silent=False)
+    session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     if PYBOP_SCHEDULED:
         session.run("pip", "install", f"pybamm=={PYBAMM_VERSION}", silent=False)
     session.run(
@@ -98,7 +98,7 @@ def notebooks_overwrite(session):
 @nox.session(name="tests")
 def run_tests(session):
     """Run all or a user-defined set of tests."""
-    session.install("-e", ".[all,dev]", silent=False)
+    session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     if PYBOP_SCHEDULED:
         session.run("pip", "install", f"pybamm=={PYBAMM_VERSION}", silent=False)
     specific_tests = session.posargs if session.posargs else []
@@ -120,7 +120,7 @@ def run_doc_tests(session):
     Checks if the documentation can be built, runs any doctests (currently not
     used).
     """
-    session.install("-e", ".[plot,docs,dev]", silent=False)
+    session.install("-e", ".[plot,docs,dev]", "--upgrade", silent=False)
     session.run("pytest", "--docs", "-n", "0")
 
 
[…]
