### 2025-04-24 1be15e8
Default solver from Casadi to IDAKLU (#546)
* fix: removes default solver Casadi args, adds catch for EIS and applies Casadi solver.

* examples: removes IDAKLU explicit definition, adds Casadi for unscented_kalman_filter.py

* tests: removes general explicit IDAKLU definition, adds increased tolerances where required, removes redundant non-convergence test

* coverage: adds explicit solver arg for ExponentialDecay

* increment pybamm version, add changlog entry

* infra: update nox sessions to upgrade dependencies by default
