### 2025-04-24 aae442b
feat: initial functional builder
