+++ pybop/problems/pybamm_problem.py
@@ -1,5 +1,8 @@
+from typing import Union
+
 import numpy as np
 
+from pybop import Inputs
 from pybop._pybamm_pipeline import PybammPipeline
 from pybop.problems.base_problem import Problem
 
@@ -14,10 +17,14 @@ class PybammProblem(Problem):
         pybamm_pipeline: list[PybammPipeline],
         param_names: list[str] = None,
         cost_names: list[str] = None,
+        simulation_weights: Union[list, np.array] = None,
+        cost_weights: Union[list, np.array] = None,
     ):
         super().__init__(param_names=param_names)
         self._pipeline = pybamm_pipeline
         self._cost_names = cost_names
+        self._simulation_weights = simulation_weights
+        self._cost_weights = cost_weights
 
     def set_params(self, p: np.ndarray) -> None:
         """
@@ -27,24 +34,26 @@ class PybammProblem(Problem):
 
         # rebuild the pipeline (if needed)
         for pipe in self._pipeline:
-            pipe.build(self._params)
+            pipe.rebuild(self._params)
 
-    def run(self) -> float:
+    def run(self, inputs: Inputs = None) -> float:
         """
         Evaluates the underlying simulation and cost function using the
         parameters set in the previous call to `set_params`.
         """
-        cost = np.zeros_like(self._pipeline)
+        cost = []
         for pipe in self._pipeline:
             self.check_set_params_called()
 
             # run simulation
-            sol = pipe.solve()
+            sol = pipe.solve(inputs=self._params.as_dict())
 
             # extract and sum cost function values. These are assumed to all be scalar values
             # (not to self: test this is true in tests....)
-            cost += sum([sol[n].values[0] for n in self._cost_names])
-        return cost
+            cost.append(
+                np.dot(self._cost_weights, [sol[n].values[0] for n in self._cost_names])
+            )
+        return np.dot(self._simulation_weights, np.array(cost))
 
     def run_with_sensitivities(
         self,
@@ -53,8 +62,8 @@ class PybammProblem(Problem):
         Evaluates the underlying simulation and cost function using the
         parameters set in the previous call to `set_params`.
         """
-        cost = np.zeros_like(self._pipeline)
-        cost_sens = np.zeros_like(self._pipeline)
+        cost = []
+        cost_sens = []
 
         for pipe in self._pipeline:
             self.check_set_params_called()
@@ -63,20 +72,27 @@ class PybammProblem(Problem):
             sol = pipe.solve(calculate_sensitivities=True)
 
             # extract cost function values. These are assumed to all be scalar values
-            # (not to self: test this is true in tests....)
-            cost += sum([sol[n].values[0] for n in self._cost_names])
+            # (note to self: test this is true in tests....)
+            cost.append(
+                np.dot(self._cost_weights, [sol[n].values[0] for n in self._cost_names])
+            )
 
             # sensitivities will all be 1D arrays of length n_params, sum over the different
             # cost functions to get the total sensitivity
-            cost_sens += np.array(
-                [
-                    sum(
-                        [
-                            sol[cost_n].sensitivities[param_n]
-                            for cost_n in self._cost_names
-                        ]
-                    )
-                    for param_n in self._params
-                ]
+            cost_sens.append(
+                np.array(
+                    [
+                        np.dot(
+                            self._cost_weights,
+                            [
+                                sol[cost_n].sensitivities[param_n]
+                                for cost_n in self._cost_names
+                            ],
+                        )
+                        for param_n in self._params
+                    ]
+                )
             )
-        return cost, cost_sens
+        return np.dot(self._simulation_weights, np.array(cost)), np.dot(
+            self._simulation_weights, np.array(cost_sens)
+        )
