+++ pybop/builders/__init__.py
@@ -1,2 +1,3 @@
 
 from .base import BaseBuilder
+from .pybamm import Pybamm
