+++ pybop/__init__.py
@@ -123,6 +123,12 @@ from .costs._likelihoods import (
     LogPosterior,
 )
 from .costs._weighted_cost import WeightedCost
+from .costs.pybamm_cost import PybammCost, SumSquaredError
+
+#
+# Problem Builder
+#
+from . import builders
 
 #
 # Experimental
