### 2025-04-24 e8fd105
feat: add init_build() to pipeline, params -> pybop.Parameters object, remove builder require_build method
