### 2025-04-25 5384ee8
tests: adds multisim builder test
