+++ tests/unit/test_problem.py
@@ -17,6 +17,10 @@ class TestProblem:
         return pybamm.lithium_ion.SPM()
 
     @pytest.fixture
+    def second_model(self):
+        return pybamm.lithium_ion.SPMe()
+
+    @pytest.fixture
     def parameters(self):
         return pybop.Parameters(
             pybop.Parameter(
@@ -95,10 +99,30 @@ class TestProblem:
         problem.set_params(np.array([0.6, 0.6]))
         assert problem.run() is not None
 
-        # Assertion to add
-        # Multi-simulation building
-        # Parameters
-        # Rebuild required construction
+    def test_multisim_builder(
+        self, first_model, second_model, parameter_values, experiment, dataset
+    ):
+        builder = pybop.builders.Pybamm()
+        builder.set_dataset(dataset)
+        builder.add_simulation(first_model, parameter_values=parameter_values)
+        builder.add_simulation(second_model, parameter_values=parameter_values)
+
+        builder.add_parameter(
+            pybop.Parameter(
+                "Negative electrode active material volume fraction", initial_value=0.6
+            )
+        )
+        builder.add_parameter(
+            pybop.Parameter(
+                "Positive electrode active material volume fraction", initial_value=0.6
+            )
+        )
+        # builder.add_cost(pybop.costs.SumSquaredError())
+        problem = builder.build()
+
+        assert problem is not None
+        problem.set_params(np.array([0.6, 0.6]))
+        assert problem.run() is not None
 
     def test_builder_with_rebuild_params(
         self, first_model, parameter_values, experiment, dataset
