### 2025-04-25 f9cc05c
fix: update SumSquaredCost init
