### 2025-04-30 d43557f
fix: link change to wayback machine snapshot with tinyurl replacement
