+++ pybop/__init__.py
@@ -87,6 +87,8 @@ from .models.base_model import Inputs
 #
 # Problem classes
 #
+from .problems.pybamm_problem import PybammProblem
+from .problems.python_problem import PythonProblem
 from .problems.base_problem_old import BaseProblem
 from .problems.fitting_problem import FittingProblem
 from .problems.multi_fitting_problem import MultiFittingProblem
