+++ pybop/builders/__init__.py
@@ -1,3 +1,4 @@
 
 from .base import BaseBuilder
 from .pybamm import Pybamm
+from .python import Python

[…]
