### 2025-05-02 4d57649
refactor: better separation of pipeline and problem, adds sensitivity tests
