### 2025-05-02 2174548
feat: adds python builder class and example sum squared cost (#707)
* feat: initial builders, updates to problem

* feat: base builder as ABC, updates for problem API, adds pybamm_cost

* feat: adds concurrent simulations to problem builder, corresponding updates to pybamm_problem.py

* feat: initial functional builder

* feat: add init_build() to pipeline, params -> pybop.Parameters object, remove builder require_build method

* tests: adds multisim builder test

* fix: update SumSquaredCost init

* refactor: Pybamm builder & problem now single instance, initial python builder

* refactor: be
[…]
