+++ tests/unit/test_problem_old.py
@@ -1,0 +1,319 @@
+import numpy as np
+import pytest
+from numpy.testing import assert_allclose, assert_array_equal
+
+import pybop
+
+
+class TestProblem:
+    """
+    A class to test the problem class.
+    """
+
+    pytestmark = pytest.mark.unit
+
+    @pytest.fixture
+    def model(self):
+        return pybop.lithium_ion.SPM()
+
+    @pytest.fixture
+    def parameters(self):
+        return pybop.Parameters(
+            pybop.Parameter(
+                "Negative particle radius [m]",
+                prior=pybop.Gaussian(2e-05, 0.1e-5),
+                bounds=[1e-6, 5e-5],
+            ),
+            pybop.Parameter(
+                "Positive particle radius [m]",
+                prior=pybop.Gaussian(0.5e-05, 0.1e-5),
+                bounds=[1e-6, 5e-5],
+            ),
+        )
+
+    @pytest.fixture
+    def experiment(self):
+        return pybop.Experiment(
+            [
+                (
+                    "Discharge at 1C for 5 minutes (1 second period)",
+                    "Rest for 2 minutes (1 second period)",
+                    "Charge at 1C for 5 minutes (1 second period)",
+                    "Rest for 2 minutes (1 second period)",
+                ),
+            ]
+            * 2
+        )
+
+    @pytest.fixture
+    def dataset(self, model, experiment):
+        x0 = np.array([2e-5, 0.5e-5])
+        model.parameter_set.update(
+            {
+                "Negative particle radius [m]": x0[0],
+                "Positive particle radius [m]": x0[1],
+            }
+        )
+        solution = model.predict(experiment=experiment)
+        return pybop.Dataset(
+            {
+                "Time [s]": solution["Time [s]"].data,
+                "Current function [A]": solution["Current [A]"].data,
+                "Voltage [V]": solution["Terminal voltage [V]"].data,
+            }
+        )
+
+    @pytest.fixture
+    def signal(self):
+        return "Voltage [V]"
+
+    def test_base_problem(self, parameters, model, dataset):
+        # Construct Problem
+        problem = pybop.BaseProblem(parameters, model=model)
+
+        with pytest.raises(NotImplementedError):
+            problem.evaluate([1e-5, 1e-5])
+        with pytest.raises(NotImplementedError):
+            problem.evaluateS1([1e-5, 1e-5])
+
+        with pytest.raises(ValueError):
+            pybop.BaseProblem(parameters, model=model, signal=[1e-5, 1e-5])
+
+        # Incorrect set target
+        with pytest.raises(ValueError, match="Dataset must be a pybop Dataset object."):
+            problem.set_target("This is not a dataset")
+
+        # No signal
+        problem.signal = None
+        with pytest.raises(ValueError, match="Signal must be defined to set target."):
+            problem.set_target(dataset)
+
+        # Different types of parameters
+        parameter_list = list(parameters.param.values())
+        problem = pybop.BaseProblem(parameters=parameter_list)
+        problem = pybop.BaseProblem(parameters=parameter_list[0])
+        with pytest.raises(
+            TypeError,
+            match="The input parameters must be a pybop Parameter, a list of pybop.Parameter objects, or a pybop Parameters object.",
+        ):
+            problem = pybop.BaseProblem(parameters="Invalid string")
+        with pytest.raises(
+            TypeError,
+            match="All elements in the list must be pybop.Parameter objects.",
+        ):
+            problem = pybop.BaseProblem(
+                parameters=[parameter_list[0], "Invalid string"]
+            )
+
+    def test_fitting_problem(self, parameters, dataset, model, signal):
+        with pytest.warns(UserWarning) as record:
+            problem = pybop.FittingProblem(
[…]
