+++ pybop/costs/__init__.py
@@ -1,0 +1,1 @@
+from .pybamm_cost import *
