### 2025-05-05 e283f49
Add return type
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
