### 2025-05-05 fd3b8c1
Init parameter grouping functions as static methods on the corresponding model
