+++ pybop/models/lithium_ion/basic_SPMe.py
@@ -562,145 +562,146 @@ class BaseGroupedSPMe(pybamm_lithium_ion.BaseModel):
             "positive particle": pybamm.FiniteVolume(),
         }
 
+    @staticmethod
+    def apply_parameter_grouping(parameter_set) -> dict:
+        """
+        A function to create a grouped SPMe parameter set from a standard
+        PyBaMM parameter set.
 
-def convert_physical_to_grouped_parameters(parameter_set):
-    """
-    A function to create a grouped SPMe parameter set from a standard
-    PyBaMM parameter set.
+        Parameters
+        ----------
+        parameter_set : Union[dict, pybop.ParameterSet, pybamm.ParameterValues]
+            A dict-like object containing the parameter values.
 
-    Parameters
-    ----------
-    parameter_set : Union[dict, pybop.ParameterSet, pybamm.ParameterValues]
-        A dict-like object containing the parameter values.
+        Returns
+        -------
+        dict
+            A dictionary of the grouped parameters.
+        """
+        parameter_set = ParameterSet.to_pybamm(parameter_set)
 
-    Returns
-    -------
-    dict
-        A dictionary of the grouped parameters.
-    """
-    parameter_set = ParameterSet.to_pybamm(parameter_set)
-
-    # Unpack physical parameters
-    F = parameter_set["Faraday constant [C.mol-1]"]
-    alpha_p = parameter_set["Positive electrode active material volume fraction"]
-    alpha_n = parameter_set["Negative electrode active material volume fraction"]
-    c_max_p = parameter_set["Maximum concentration in positive electrode [mol.m-3]"]
-    c_max_n = parameter_set["Maximum concentration in negative electrode [mol.m-3]"]
-    L_p = parameter_set["Positive electrode thickness [m]"]
-    L_n = parameter_set["Negative electrode thickness [m]"]
-    epsilon_p = parameter_set["Positive electrode porosity"]
-    epsilon_n = parameter_set["Negative electrode porosity"]
-    R_p = parameter_set["Positive particle radius [m]"]
-    R_n = parameter_set["Negative particle radius [m]"]
-    D_p = parameter_set["Positive particle diffusivity [m2.s-1]"]
-    D_n = parameter_set["Negative particle diffusivity [m2.s-1]"]
-    b_p = parameter_set["Positive electrode Bruggeman coefficient (electrolyte)"]
-    b_n = parameter_set["Negative electrode Bruggeman coefficient (electrolyte)"]
-    Cdl_p = parameter_set["Positive electrode double-layer capacity [F.m-2]"]
-    Cdl_n = parameter_set["Negative electrode double-layer capacity [F.m-2]"]
-    m_p = 3.42e-6  # (A/m2)(m3/mol)**1.5
-    m_n = 6.48e-7  # (A/m2)(m3/mol)**1.5
-    sigma_p = (
-        parameter_set["Positive electrode conductivity [S.m-1]"]
-        * alpha_p
-        ** parameter_set["Positive electrode Bruggeman coefficient (electrode)"]
-    )
-    sigma_n = (
-        parameter_set["Negative electrode conductivity [S.m-1]"]
-        * alpha_n
-        ** parameter_set["Negative electrode Bruggeman coefficient (electrode)"]
-    )
-
-    # Separator and electrolyte properties
-    ce0 = parameter_set["Initial concentration in electrolyte [mol.m-3]"]
-    De = parameter_set["Electrolyte diffusivity [m2.s-1]"]  # (ce0, T)
[…]
