+++ pybop/__init__.py
@@ -203,7 +203,7 @@ from ._classification import classify_using_hessian
 # Applications
 #
 from .applications.base_method import BaseApplication, Interpolant, InverseOCV
-from .applications.ocp_methods import OCPBlend, OCPAverage, OCPCapacityToStoichiometry
+from .applications.ocp_methods import OCPMerge, OCPAverage, OCPCapacityToStoichiometry
 from .applications.gitt_methods import GITTPulseFit, GITTFit
 
 #
