### 2025-05-05 4714f8b
Split init and call methods
