+++ pybop/applications/ocp_methods.py
@@ -29,44 +29,54 @@ class OCPMerge(BaseApplication):
         ocp_charge: pybop.Dataset,
         n_sto_points: int = 101,
     ):
+        self.ocp_discharge = ocp_discharge
+        self.ocp_charge = ocp_charge
+        self.n_sto_points = n_sto_points
+
+    def __call__(self) -> pybop.Dataset:
         # Use the discharge branch as the target to fit
         voltage_discharge = pybop.Interpolant(
-            ocp_discharge["Stoichiometry"], ocp_discharge["Voltage [V]"]
+            self.ocp_discharge["Stoichiometry"], self.ocp_discharge["Voltage [V]"]
         )
 
         # Use the charge branch as the model output
         voltage_charge = pybop.Interpolant(
-            ocp_charge["Stoichiometry"], ocp_charge["Voltage [V]"]
+            self.ocp_charge["Stoichiometry"], self.ocp_charge["Voltage [V]"]
         )
 
         if np.sign(
-            ocp_charge["Stoichiometry"][-1] - ocp_charge["Stoichiometry"][0]
-        ) == np.sign(ocp_charge["Voltage [V]"][-1] - ocp_charge["Voltage [V]"][0]):
+            self.ocp_charge["Stoichiometry"][-1] - self.ocp_charge["Stoichiometry"][0]
+        ) == np.sign(
+            self.ocp_charge["Voltage [V]"][-1] - self.ocp_charge["Voltage [V]"][0]
+        ):
             # Increasing stoichiometry corresponds to increasing voltage (full cell)
-            sto_min = np.min(ocp_charge["Stoichiometry"])
-            sto_max = np.max(ocp_discharge["Stoichiometry"])
+            sto_min = np.min(self.ocp_charge["Stoichiometry"])
+            sto_max = np.max(self.ocp_discharge["Stoichiometry"])
             low_sto_fit = voltage_charge
             high_sto_fit = voltage_discharge
         else:
             # Decreasing stoichiometry corresponds to increasing voltage (electrode)
-            sto_min = np.min(ocp_discharge["Stoichiometry"])
-            sto_max = np.max(ocp_charge["Stoichiometry"])
+            sto_min = np.min(self.ocp_discharge["Stoichiometry"])
+            sto_max = np.max(self.ocp_charge["Stoichiometry"])
             low_sto_fit = voltage_discharge
             high_sto_fit = voltage_charge
 
         # Generate evenly spaced data for dataset creation
-        sto_evenly_spaced = np.linspace(sto_min, sto_max, n_sto_points)
+        self.sto_evenly_spaced = np.linspace(sto_min, sto_max, self.n_sto_points)
 
         # Define a linear transition from the charge branch at low voltage
         # to the charge branch at high voltage
-        transition = np.linspace(0, 1, len(sto_evenly_spaced))
+        transition = np.linspace(0, 1, len(self.sto_evenly_spaced))
         voltage_merge = (1 - transition) * low_sto_fit(
-            sto_evenly_spaced
-        ) + transition * high_sto_fit(sto_evenly_spaced)
+            self.sto_evenly_spaced
+        ) + transition * high_sto_fit(self.sto_evenly_spaced)
 
         self.dataset = pybop.Dataset(
-            {"Stoichiometry": sto_evenly_spaced, "Voltage [V]": voltage_merge}
+            {"Stoichiometry": self.sto_evenly_spaced, "Voltage [V]": voltage_merge}
         )
+        self.check_monotonicity(voltage_merge)
+
+        return self.dataset
 
 
 class OCPAverage(BaseApplication):
@@ -106,34 +116,46 @@ class OCPAverage(BaseApplication):
[…]
