### 2025-05-06 dc30f5c
Create battery benchmark methods (#644)
* Create ocp_average

* Add stoichiometry_fitting

* Create test_applications.py

* Rename gitt to gitt_pulse

* Add ocp_blend

* Update Weppner & Huggins model

* Align model names

* Create SPDiffusion model

* Update examples

* Add SPDiffusion to model tests

* Update documentation

* Add gitt_methods

* Update ocp_average

* Add base and rename classes

* Create GITTFit with example

* Update stoichiometry_fitting.py

* Add titles

* Add test_gitt_pulse_fit

* Add Interpolant and InverseOCV

* Update OCP methods

* Update GITT methods

* Fix gitt_fi
[…]
