+++ pybop/applications/ocp_methods.py
@@ -1,0 +1,360 @@
+from typing import Callable, Optional
+
+import numpy as np
+
+import pybop
+from pybop import BaseApplication
+
+
+class OCPMerge(BaseApplication):
+    """
+    Generate a representative open-circuit potential (OCP) without hysteresis by
+    merging the provided charge and discharge branches of the OCP.
+
+    Parameters
+    ----------
+    ocp_discharge : pybop.Dataset
+        A dataset containing the "Stoichiometry" and "Voltage [V]" obtained from a
+        discharge measurement.
+    ocp_charge : pybop.Dataset
+        A dataset containing the "Stoichiometry" and "Voltage [V]" obtained from a
+        charge measurement.
+    n_sto_points : int, optional
+        The number of points in stoichiometry at which to calculate the voltage.
+    """
+
+    def __init__(
+        self,
+        ocp_discharge: pybop.Dataset,
+        ocp_charge: pybop.Dataset,
+        n_sto_points: int = 101,
+    ):
+        self.ocp_discharge = ocp_discharge
+        self.ocp_charge = ocp_charge
+        self.n_sto_points = n_sto_points
+
+    def __call__(self) -> pybop.Dataset:
+        # Use the discharge branch as the target to fit
+        voltage_discharge = pybop.Interpolant(
+            self.ocp_discharge["Stoichiometry"], self.ocp_discharge["Voltage [V]"]
+        )
+
+        # Use the charge branch as the model output
+        voltage_charge = pybop.Interpolant(
+            self.ocp_charge["Stoichiometry"], self.ocp_charge["Voltage [V]"]
+        )
+
+        if np.sign(
+            self.ocp_charge["Stoichiometry"][-1] - self.ocp_charge["Stoichiometry"][0]
+        ) == np.sign(
+            self.ocp_charge["Voltage [V]"][-1] - self.ocp_charge["Voltage [V]"][0]
+        ):
+            # Increasing stoichiometry corresponds to increasing voltage (full cell)
+            sto_min = np.min(self.ocp_charge["Stoichiometry"])
+            sto_max = np.max(self.ocp_discharge["Stoichiometry"])
+            low_sto_fit = voltage_charge
+            high_sto_fit = voltage_discharge
+        else:
+            # Decreasing stoichiometry corresponds to increasing voltage (electrode)
+            sto_min = np.min(self.ocp_discharge["Stoichiometry"])
+            sto_max = np.max(self.ocp_charge["Stoichiometry"])
+            low_sto_fit = voltage_discharge
+            high_sto_fit = voltage_charge
+
+        # Generate evenly spaced data for dataset creation
+        self.sto_evenly_spaced = np.linspace(sto_min, sto_max, self.n_sto_points)
+
+        # Define a linear transition from the charge branch at low voltage
+        # to the charge branch at high voltage
+        transition = np.linspace(0, 1, len(self.sto_evenly_spaced))
+        voltage_merge = (1 - transition) * low_sto_fit(
+            self.sto_evenly_spaced
+        ) + transition * high_sto_fit(self.sto_evenly_spaced)
+
+        self.dataset = pybop.Dataset(
+            {"Stoichiometry": self.sto_evenly_spaced, "Voltage [V]": voltage_merge}
+        )
+        self.check_monotonicity(voltage_merge)
+
+        return self.dataset
+
+
+class OCPAverage(BaseApplication):
+    """
+    Estimate the equlilibrium open-circuit potential (OCP) by averaging the charge
+    and discharge branches, using a method loosely based on method 4(a) proposed by
+    Lu et al. (2021) available at: https://doi.org/10.1149/1945-7111/ac11a5
+
+    Parameters
+    ----------
[…]
