### 2025-05-06 e2c6931
feat: initial EIS Builder
