+++ pybop/builders/__init__.py
@@ -2,3 +2,4 @@
 from .base import BaseBuilder
 from .pybamm import Pybamm
 from .python import Python
+from .pybamm_eis import PybammEIS
