+++ pybop/_pybamm_eis_pipeline.py
@@ -1,0 +1,190 @@
+import casadi
+import numpy as np
+import pybamm
+from scipy.sparse import csc_matrix
+from scipy.sparse.linalg import spsolve
+
+from pybop import Parameters, SymbolReplacer
+from pybop._pybamm_pipeline import PybammPipeline
+
+
+class PybammEISPipeline(PybammPipeline):
+    """
+    A class to build a PyBaMM pipeline for a given model and data, and run the resultant simulation.
+
+    There are two contexts in which this class can be used:
+    1. A pybamm model needs to be built once, and then run multiple times with different input parameters
+    2. A pybamm model needs to be built multiple times with different parameter values,
+        for the case where some of the parameters are geometric parameters which change the mesh
+
+    To enable 2., you can pass a list of parameter names to the constructor, these parameters will be set
+    before the model is built each time (using the `build` method).
+    To enable 1, you can just pass an empty list. The model will be built once and subsequent calls
+    to the `build` method will not change the model.
+    """
+
+    def __init__(
+        self,
+        model: pybamm.BaseModel,
+        parameter_values: pybamm.ParameterValues = None,
+        pybop_parameters: Parameters = None,
+        solver: pybamm.BaseSolver = None,
+        f_eval: list = None,
+        var_pts: dict = None,
+    ):
+        """
+        Arguments
+        ---------
+        model : pybamm.BaseModel
+            The PyBaMM model to be used.
+        parameter_values : pybamm.ParameterValues
+            The parameters to be used in the model.
+        solver : pybamm.BaseSolver
+            The solver to be used. If None, the idaklu solver will be used.
+        t_start : number
+            The start time of the simulation.
+        t_end : number
+            The end time of the simulation.
+        t_interp : np.ndarray
+            The time points at which to interpolate the solution. If None, no interpolation will be done.
+        rebuild_parameters : list[str]
+            The parameters that will be used to rebuild the model. If None, the model will not be rebuilt.
+        """
+        super().__init__(
+            model,
+            parameter_values=parameter_values,
+            pybop_parameters=pybop_parameters,
+            solver=solver,
+            var_pts=var_pts,
+        )
+
+        # Set-up model for EIS
+        self._f_eval = f_eval
+        self._model = self.set_up_for_eis(model.new_copy())
+        self._parameter_values["Current function [A]"] = 0
+
+        v_scale = getattr(model.variables["Voltage [V]"], "scale", 1)
+        i_scale = getattr(model.variables["Current [A]"], "scale", 1)
+        self.z_scale = self._parameter_values.evaluate(v_scale / i_scale)
+
+    def set_up_for_eis(self, model):
+        """
+        Set up the model for electrochemical impedance spectroscopy (EIS) simulations.
+        This method sets up the model for EIS simulations by adding the necessary
+        algebraic equations and variables to the model.
+        Originally developed by pybamm-eis: https://github.com/pybamm-team/pybamm-eis
+        Parameters
+        ----------
+        model : pybamm.Model
+            The PyBaMM model to be used for EIS simulations.
+        """
+        V_cell = pybamm.Variable("Voltage variable [V]")
+        model.variables["Voltage variable [V]"] = V_cell
+        V = model.variables["Voltage [V]"]
+
+        # Add algebraic equation for the voltage
+        model.algebraic[V_cell] = V_cell - V
+        model.initial_conditions[V_cell] = model.param.ocv_init
+
+        # Create the FunctionControl submodel and extract variables
+        external_circuit_variables = pybamm.external_circuit.FunctionControl(
+            model.param,
+            None,
+            model.options,
+            control="algebraic",
+        ).get_fundamental_variables()
+
[…]

[…]
