### 2025-05-08 f6a5f30
Update pybop/costs/pybamm_cost.py
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
