### 2025-05-08 a232abc
Update pybop/costs/pybamm_sum_squared_error.py
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
