### 2025-05-08 f468b03
add some tests
