### 2025-05-09 2b51c7d
bug: handle pybamm discrete time sum sensitivity bug
