### 2025-05-09 4dbd0ca
refactor: restructure pybam costs and test them along with the problem class (#720)
* feat: restructure pybamm costs

* style: pre-commit fixes

* fix imports

* Update pybop/costs/pybamm_cost.py

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

* Update pybop/costs/pybamm_sum_squared_error.py

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

* add some tests

* bug: handle pybamm discrete time sum sensitivity bug

* update Dict -> dict

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.
[…]
