+++ pybop/costs/__init__.py
@@ -1,1 +1,0 @@
-from .pybamm_cost import *

[…]
