+++ pybop/optimisers/base_optimiser_new.py
@@ -1,0 +1,238 @@
+import warnings
+from copy import deepcopy
+from dataclasses import dataclass, field
+
+import jax.numpy as jnp
+import numpy as np
+import pybop
+
+
+@dataclass
+class OptimisationLog:
+    """Stores optimisation progress data."""
+
+    iterations: list[int] = field(default_factory=list)
+    evaluations: list[int] = field(default_factory=list)
+    x: list[list[float]] = field(default_factory=list)
+    x_best: list[list[float]] = field(default_factory=list)
+    x_search: list[list[float]] = field(default_factory=list)
+    x0: list[list[float]] = field(default_factory=list)
+    cost: list[float] = field(default_factory=list)
+    cost_best: list[float] = field(default_factory=list)
+
+
+class OptimisationLogger:
+    def __init__(self, verbose: bool = False, verbose_print_rate: int = 50):
+        self._verbose = verbose
+        self._verbose_print_rate = verbose_print_rate
+        self._iter_count = 0
+
+    def log_update(
+        self,
+        iterations=None,
+        evaluations=None,
+        x_model=None,
+        x_search=None,
+        x_model_best=None,
+        x_seach_best=None,
+        cost=None,
+        cost_best=None,
+        x0=None,
+    ):
+        """
+        Update the log with new values.
+
+        Parameters
+        ----------
+        iterations : list or array-like, optional
+            Iteration indices to log (default: None).
+        evaluations: list or array-like, optional
+            Evaluation indices to log (default: None).
+        x_model : list or array-like, optional
+            Parameter values in model space (default: None).
+        x_search : list or array-like, optional
+            Search parameter values in search space (default: None).
+        x_best_model : list or array-like, optional
+            Parameter values in model space corresponding to the best cost yet (default: None).
+        x_best_search : list or array-like, optional
+            Parameter values in search space corresponding to the best cost yet (default: None).
+        cost : list, optional
+            Cost values corresponding to x (default: None).
+        cost_best : list, optional
+            Cost values corresponding to x_best (default: None).
+        x0 : list or array-like, optional
+            Initial parameter values (default: None).
+        """
+        # Update logs for each provided parameter
+        self._update_log_entry("iterations", iterations)
+        self._update_log_entry("evaluations", evaluations)
+        self._update_log_entry("x_model", x_model)
+        self._update_log_entry("x_search", x_search)
+        self._update_log_entry("x_model_best", x_model_best)
+        self._update_log_entry("x_search_best", x_seach_best)
+        self._update_log_entry("x0", x0)
+        self._update_log_entry("cost", cost)
+        self._update_log_entry("cost_best", cost_best)
+
+        # Verbose output
+        self._print_verbose_output()
+        self._iter_count += 1
+
+    def _update_log_entry(self, key, value):
+        """Update a log entry if the value is provided."""
+        if value is not None:
+            getattr(self.log, key).extend(self._to_list(value))
+
+    def _to_list(self, array_like):
+        """Convert input to a list."""
+        if isinstance(array_like, (list, tuple, np.ndarray, jnp.ndarray)):
+            return list(array_like)
+        return [array_like]
+
+    def _print_verbose_output(self):
+        """Print verbose optimization information if enabled."""
+        if not self._verbose:
+            return
+
+        latest_iter = (
+            self.log.iterations[-1] if self.log.iterations else self._iter_count
+        )
+
+        # Only print on first 10 iterations, then every Nth iteration
+        if latest_iter > 10 and latest_iter % self._verbose_print_rate != 0:
+            return
+
+        latest_eval = self.log.evaluations[-1] if self.log.evaluations else "N/A"
+        latest_x_best = self.log.x_best[-1] if self.log.x_best else "N/A"
[…]
