+++ pybop/optimisers/base_pints_optimiser_new.py
@@ -1,0 +1,525 @@
+from time import time
+
+import numpy as np
+from pints import PSO as PintsPSO
+from pints import NelderMead as PintsNelderMead
+from pints import Optimiser as PintsOptimiser
+from pints import ParallelEvaluator as PintsParallelEvaluator
+from pints import PopulationBasedOptimiser
+from pints import PopulationBasedOptimiser as PintsPopulationBasedOptimiser
+from pints import RectangularBoundaries as PintsRectangularBoundaries
+from pints import SequentialEvaluator as PintsSequentialEvaluator
+from pints import strfloat as PintsStrFloat
+
+from pybop import (
+    AdamWImpl,
+    BaseJaxCost,
+    BaseOptimiser,
+    GradientDescentImpl,
+    OptimisationResult,
+    SequentialJaxEvaluator,
+)
+
+
+class BasePintsOptimiser(BaseOptimiser):
+    """
+    A base class for defining optimisation methods from the PINTS library.
+
+    Parameters
+    ----------
+    cost : callable
+        The cost function to be minimized.
+    pints_optimiser : class
+        The PINTS optimiser class to be used.
+    max_iterations : int, optional
+        Maximum number of iterations for the optimisation.
+    min_iterations : int, optional (default=2)
+        Minimum number of iterations before termination.
+    max_unchanged_iterations : int, optional (default=15)
+        Maximum number of iterations without improvement before termination.
+    parallel : bool, optional (default=False)
+        Whether to run the optimisation in parallel.
+    **optimiser_kwargs : optional
+        Valid PINTS option keys and their values, for example:
+        x0 : array_like
+            Initial position from which optimization will start.
+        sigma0 : float
+            Initial step size or standard deviation depending on the optimiser.
+        bounds : dict
+            A dictionary with 'lower' and 'upper' keys containing arrays for lower and
+            upper bounds on the parameters.
+        use_f_guessed : bool
+            Whether to track guessed function values.
+        absolute_tolerance : float
+            Absolute tolerance for convergence checking.
+        relative_tolerance : float
+            Relative tolerance for convergence checking.
+        max_evaluations : int
+            Maximum number of function evaluations.
+        threshold : float
+            Threshold value for early termination.
+    """
+
+    def __init__(
+        self,
+        cost,
+        pints_optimiser,
+        max_iterations: int = None,
+        min_iterations: int = 2,
+        max_unchanged_iterations: int = 15,
+        multistart: int = 1,
+        parallel: bool = False,
+        **optimiser_kwargs,
+    ):
+        # First set attributes to default values
+        self._boundaries = None
+        self._needs_sensitivities = None
+        self._use_f_guessed = None
+        self._n_workers = 1
+        self._callback = None
+        self.set_parallel(parallel)
+        self.set_max_iterations(max_iterations)
+        self.set_min_iterations(min_iterations)
+        self._unchanged_max_iterations = max_unchanged_iterations
+        self._absolute_tolerance = 1e-5
+        self._relative_tolerance = 1e-2
+        self._max_evaluations = None
+        self._threshold = None
+        self._evaluations = None
+        self._iterations = None
+        self.option_methods = {
+            "use_f_guessed": self.set_f_guessed_tracking,
+            "max_evaluations": self.set_max_evaluations,
+            "threshold": self.set_threshold,
+        }
+
+        self._pints_optimiser = pints_optimiser
+        optimiser_kwargs["multistart"] = multistart
+        super().__init__(cost, **optimiser_kwargs)
+
+    def _set_up_optimiser(self):
+        """
+        Parse optimiser options and create an instance of the PINTS optimiser.
+        """
[…]

[…]
