### 2025-05-12 611b816
refactor EISBuilder, EISProblem, adds python costs
