+++ pybop/costs/error_measures.py
@@ -2,7 +2,6 @@ from typing import Optional, Union
 
 import numpy as np
 
-from pybop import FittingCost
 from pybop.costs.base_cost import BaseCost
 
 
@@ -13,23 +12,29 @@ class NewMeanSquaredError(BaseCost):
     Computes the mean square error between model predictions and the target
     data, providing a measure of the differences between predicted values and
     observed values.
+
+    Parameters
+    ----------
+    weighting : np.ndarray, optional
+        The type of weighting array to use when taking the sum or mean of the error
+        measure. Options: "equal"(default), "domain", or a custom numpy array.
     """
 
-    @staticmethod
     def __call__(
+        self,
         r: np.ndarray,
         dy: Optional[np.ndarray] = None,
     ) -> Union[float, tuple[float, np.ndarray]]:
-        e = np.mean(np.abs(r) ** 2)
+        e = np.sum(np.abs(r) ** 2 * self.weighting)
 
         if dy is not None:
-            de = 2 * np.mean((r) * dy, axis=(1, 2))
+            de = 2 * np.sum((r * self.weighting) * dy, axis=(1, 2))
             return e, de
 
         return e
 
 
-class RootMeanSquaredError(FittingCost):
+class RootMeanSquaredError(BaseCost):
     """
     Root mean square error (RMSE) cost function.
 
@@ -38,7 +43,7 @@ class RootMeanSquaredError(FittingCost):
     observed values.
     """
 
-    def _error_measure(
+    def __call__(
         self,
         r: np.ndarray,
         dy: Optional[np.ndarray] = None,
@@ -54,16 +59,22 @@ class RootMeanSquaredError(FittingCost):
         return e
 
 
-class MeanAbsoluteError(FittingCost):
+class MeanAbsoluteError(BaseCost):
     """
     Mean absolute error (MAE) cost function.
 
     Computes the mean absolute error (MAE) between model predictions
     and target data. The MAE is a measure of the average magnitude
     of errors in a set of predictions, without considering their direction.
+
+    Parameters
+    ----------
+    weighting : np.ndarray, optional
+        The type of weighting array to use when taking the sum or mean of the error
+        measure. Options: "equal"(default), "domain", or a custom numpy array.
     """
 
-    def _error_measure(
+    def __call__(
         self,
         r: np.ndarray,
         dy: Optional[np.ndarray] = None,
@@ -77,7 +88,7 @@ class MeanAbsoluteError(FittingCost):
         return e
 
 
-class SumSquaredError(FittingCost):
+class SumSquaredError(BaseCost):
     """
     Sum of squared error (SSE) cost function.
 
@@ -86,7 +97,7 @@ class SumSquaredError(FittingCost):
     predicted and observed values.
     """
 
-    def _error_measure(
+    def __call__(
         self,
         r: np.ndarray,
         dy: Optional[np.ndarray] = None,
@@ -100,7 +111,7 @@ class SumSquaredError(FittingCost):
         return e
 
 
-class Minkowski(FittingCost):
+class Minkowski(BaseCost):
     """
     The Minkowski distance is a generalisation of several distance metrics,
     including the Euclidean and Manhattan distances. It is defined as:
@@ -132,7 +143,7 @@ class Minkowski(FittingCost):
     def __init__(
         self, problem, p: float = 2.0, weighting: Union[str, np.ndarray] = None
     ):
-        super().__init__(problem, weighting=weighting)
+        super().__init__()
         if p < 0:
             raise ValueError(
                 "The order of the Minkowski distance must be greater than 0."
@@ -143,7 +154,7 @@ class Minkowski(FittingCost):
             )
         self.p = float(p)
 
-    def _error_measure(
+    def __call__(
[…]
