### 2025-05-12 fd3e259
docs: docstring updates
