+++ pybop/_pybamm_eis_pipeline.py
@@ -1,3 +1,5 @@
+import warnings
+
 import casadi
 import numpy as np
 import pybamm
@@ -10,7 +12,7 @@ from pybop._pybamm_pipeline import PybammPipeline
 
 class PybammEISPipeline(PybammPipeline):
     """
-    A class to build a PyBaMM pipeline for a given model and data, and run the resultant simulation.
+    A class to build an EIS PyBaMM pipeline for a given model and data, and run the resultant simulation.
 
     There are two contexts in which this class can be used:
     1. A pybamm model needs to be built once, and then run multiple times with different input parameters
@@ -26,14 +28,14 @@ class PybammEISPipeline(PybammPipeline):
     def __init__(
         self,
         model: pybamm.BaseModel,
+        f_eval: list,
         parameter_values: pybamm.ParameterValues = None,
         pybop_parameters: Parameters = None,
         solver: pybamm.BaseSolver = None,
-        f_eval: list = None,
         var_pts: dict = None,
     ):
         """
-        Arguments
+        Parameters
         ---------
         model : pybamm.BaseModel
             The PyBaMM model to be used.
@@ -41,14 +43,10 @@ class PybammEISPipeline(PybammPipeline):
             The parameters to be used in the model.
         solver : pybamm.BaseSolver
             The solver to be used. If None, the idaklu solver will be used.
-        t_start : number
-            The start time of the simulation.
-        t_end : number
-            The end time of the simulation.
-        t_interp : np.ndarray
-            The time points at which to interpolate the solution. If None, no interpolation will be done.
-        rebuild_parameters : list[str]
-            The parameters that will be used to rebuild the model. If None, the model will not be rebuilt.
+        f_eval : list
+            The frequencies at which to evaluate the impedance.
+        var_pts : dict
+            The number of points at which to discretise the model.
         """
         super().__init__(
             model,
@@ -63,21 +61,44 @@ class PybammEISPipeline(PybammPipeline):
         self._model = self.set_up_for_eis(model.new_copy())
         self._parameter_values["Current function [A]"] = 0
 
+        # Initialise
+        self.M = None
+        self._jac = None
+        self.b = None
+
         v_scale = getattr(model.variables["Voltage [V]"], "scale", 1)
         i_scale = getattr(model.variables["Current [A]"], "scale", 1)
         self.z_scale = self._parameter_values.evaluate(v_scale / i_scale)
 
-    def set_up_for_eis(self, model):
+    def set_up_for_eis(self, model: pybamm.BaseModel) -> pybamm.BaseModel:
         """
         Set up the model for electrochemical impedance spectroscopy (EIS) simulations.
-        This method sets up the model for EIS simulations by adding the necessary
-        algebraic equations and variables to the model.
+        This method adds the necessary algebraic equations and variables to the model.
         Originally developed by pybamm-eis: https://github.com/pybamm-team/pybamm-eis
+
         Parameters
         ----------
-        model : pybamm.Model
+        model : pybamm.BaseModel
             The PyBaMM model to be used for EIS simulations.
+
+        Returns
+        -------
+        pybamm.BaseModel
+            The modified model ready for EIS simulations.
+
+        Raises
+        ------
+        ValueError
+            If the model is missing required variables.
         """
+        # Verify model has required variables
+        required_vars = ["Voltage [V]", "Current [A]"]
+        for var in required_vars:
+            if var not in model.variables:
+                raise ValueError(
+                    f"Model must contain variable '{var}' for EIS simulation"
+                )
+
         V_cell = pybamm.Variable("Voltage variable [V]")
         model.variables["Voltage variable [V]"] = V_cell
         V = model.variables["Voltage [V]"]
@@ -123,7 +144,12 @@ class PybammEISPipeline(PybammPipeline):
         """
[…]
