+++ pybop/optimisers/pints_optimisers.py
@@ -1,3 +1,4 @@
+from typing import Optional
 from pints import CMAES as PintsCMAES
 from pints import PSO as PintsPSO
 from pints import SNES as PintsSNES
@@ -5,6 +6,8 @@ from pints import XNES as PintsXNES
 from pints import IRPropMin as PintsIRPropMin
 from pints import NelderMead as PintsNelderMead
 
+from pybop.problems.base_problem import Problem
+import pybop
 from pybop import (
     AdamWImpl,
     BasePintsOptimiser,
@@ -27,38 +30,10 @@ class GradientDescent(BasePintsOptimiser):
 
     Parameters
     ----------
-    cost : callable
+    problem: pybop.Problem
         The cost function to be minimised.
-    max_iterations : int, optional
-        Maximum number of iterations for the optimisation.
-    min_iterations : int, optional (default=2)
-        Minimum number of iterations before termination.
-    max_unchanged_iterations : int, optional (default=15)
-        Maximum number of iterations without improvement before termination.
-    multistart : int, optional (default=1)
-        Number of optimiser restarts from randomly sample position. These positions
-        are sampled from the priors.
-    parallel : bool, optional (default=False)
-        Whether to run the optimisation in parallel.
-    **optimiser_kwargs : optional
-        Valid PINTS option keys and their values, for example:
-        x0 : array_like
-            Initial position from which optimisation will start.
-        sigma0 : float
-            Initial step size or standard deviation depending on the optimiser.
-        bounds : dict
-            A dictionary with 'lower' and 'upper' keys containing arrays for lower and
-            upper bounds on the parameters.
-        use_f_guessed : bool
-            Whether to return the guessed function values.
-        absolute_tolerance : float
-            Absolute tolerance for convergence checking.
-        relative_tolerance : float
-            Relative tolerance for convergence checking.
-        max_evaluations : int
-            Maximum number of function evaluations.
-        threshold : float
-            Threshold value for early termination.
+    options: pybop.PintsOptions
+        optimisation options
 
     See Also
     --------
@@ -67,23 +42,13 @@ class GradientDescent(BasePintsOptimiser):
 
     def __init__(
         self,
-        cost,
-        max_iterations: int = None,
-        min_iterations: int = 2,
-        max_unchanged_iterations: int = 15,
-        multistart: int = 1,
-        parallel: bool = False,
-        **optimiser_kwargs,
+        problem: Problem,
+        options: Optional[pybop.PintsOptions],
     ):
         super().__init__(
-            cost,
+            problem,
             GradientDescentImpl,
-            max_iterations,
-            min_iterations,
-            max_unchanged_iterations,
-            multistart,
-            parallel,
-            **optimiser_kwargs,
+            options,
         )
 
 
@@ -100,38 +65,10 @@ class AdamW(BasePintsOptimiser):
 
     Parameters
     ----------
-    cost : callable
+    problem: pybop.Problem
         The cost function to be minimised.
-    max_iterations : int, optional
-        Maximum number of iterations for the optimisation.
-    min_iterations : int, optional (default=2)
-        Minimum number of iterations before termination.
-    max_unchanged_iterations : int, optional (default=15)
-        Maximum number of iterations without improvement before termination.
-    multistart : int, optional (default=1)
-        Number of optimiser restarts from randomly sample position. These positions
-        are sampled from the priors.
-    parallel : bool, optional (default=False)
-        Whether to run the optimisation in parallel.
-    **optimiser_kwargs : optional
-        Valid PINTS option keys and their values, for example:
-        x0 : array_like
-            Initial position from which optimisation will start.
-        sigma0 : float
-            Initial step size or standard deviation depending on the optimiser.
-        bounds : dict
[…]
