### 2025-05-14 0e3c8e1
refactor: EIS pipeline composes pybamm pipeline
