### 2025-05-14 4c96b36
Add EIS Builder (#717)
