+++ pybop/_pybamm_eis_pipeline.py
@@ -1,0 +1,233 @@
+import warnings
+
+import casadi
+import numpy as np
+import pybamm
+from scipy.sparse import csc_matrix
+from scipy.sparse.linalg import spsolve
+
+from pybop import Parameters, SymbolReplacer
+from pybop._pybamm_pipeline import PybammPipeline
+
+
+class PybammEISPipeline:
+    """
+    A class to build an EIS PyBaMM pipeline for a given model and data, and run the resultant simulation.
+
+    There are two contexts in which this class can be used:
+    1. A pybamm model needs to be built once, and then run multiple times with different input parameters
+    2. A pybamm model needs to be built multiple times with different parameter values,
+        for the case where some of the parameters are geometric parameters which change the mesh
+
+    To enable 2., you can pass a list of parameter names to the constructor, these parameters will be set
+    before the model is built each time (using the `build` method).
+    To enable 1, you can just pass an empty list. The model will be built once and subsequent calls
+    to the `build` method will not change the model.
+    """
+
+    def __init__(
+        self,
+        model: pybamm.BaseModel,
+        f_eval: list,
+        parameter_values: pybamm.ParameterValues = None,
+        pybop_parameters: Parameters = None,
+        solver: pybamm.BaseSolver = None,
+        var_pts: dict = None,
+    ):
+        """
+        Parameters
+        ---------
+        model : pybamm.BaseModel
+            The PyBaMM model to be used.
+        parameter_values : pybamm.ParameterValues
+            The parameters to be used in the model.
+        solver : pybamm.BaseSolver
+            The solver to be used. If None, the idaklu solver will be used.
+        f_eval : list
+            The frequencies at which to evaluate the impedance.
+        var_pts : dict
+            The number of points at which to discretise the model.
+        """
+
+        self._pybamm_pipeline = PybammPipeline(
+            model,
+            parameter_values=parameter_values,
+            pybop_parameters=pybop_parameters,
+            solver=solver,
+            var_pts=var_pts,
+        )
+
+        # Set-up model for EIS
+        self._f_eval = f_eval
+        self.set_up_for_eis(self._pybamm_pipeline.model)
+        self._pybamm_pipeline.set_parameter_value("Current function [A]", 0)
+
+        # Initialise
+        self.M = None
+        self._jac = None
+        self.b = None
+
+        v_scale = getattr(model.variables["Voltage [V]"], "scale", 1)
+        i_scale = getattr(model.variables["Current [A]"], "scale", 1)
+        self.z_scale = self._pybamm_pipeline.parameter_values.evaluate(
+            v_scale / i_scale
+        )
+
+    def set_up_for_eis(self, model: pybamm.BaseModel) -> None:
+        """
+        Set up the model for electrochemical impedance spectroscopy (EIS) simulations.
+        This method adds the necessary algebraic equations and variables to the model.
+        Originally developed by pybamm-eis: https://github.com/pybamm-team/pybamm-eis
+
+        Parameters
+        ----------
+        model : pybamm.BaseModel
+            The PyBaMM model to be used for EIS simulations.
+
+        Returns
+        -------
+        pybamm.BaseModel
+            The modified model ready for EIS simulations.
+
+        Raises
+        ------
+        ValueError
+            If the model is missing required variables.
+        """
+        # Verify model has required variables
+        required_vars = ["Voltage [V]", "Current [A]"]
+        for var in required_vars:
+            if var not in model.variables:
+                raise ValueError(
+                    f"Model must contain variable '{var}' for EIS simulation"
+                )
+
+        V_cell = pybamm.Variable("Voltage variable [V]")
+        model.variables["Voltage variable [V]"] = V_cell
+        V = model.variables["Voltage [V]"]
[…]
