### 2025-05-14 d8ca7dd
refactor: W&H, GroupedSPMe restructure for builder, new problem
