+++ pybop/models/lithium_ion/echem.py
@@ -1,8 +1,6 @@
 from pybamm import lithium_ion as pybamm_lithium_ion
 
 from pybop.models.lithium_ion.base_echem import EChemBaseModel
-from pybop.models.lithium_ion.basic_SPMe import BaseGroupedSPMe
-from pybop.models.lithium_ion.weppner_huggins import BaseWeppnerHuggins
 
 
 class SPM(EChemBaseModel):
@@ -244,97 +242,97 @@ class MSMR(EChemBaseModel):
         )
 
 
-class WeppnerHuggins(EChemBaseModel):
-    """
-    Represents the Weppner & Huggins model to fit diffusion coefficients to GITT data.
+# class WeppnerHuggins(EChemBaseModel):
+#     """
+#     Represents the Weppner & Huggins model to fit diffusion coefficients to GITT data.
+#
+#     Parameters
+#     ----------
+#     name: str, optional
+#         A name for the model instance, defaults to "Weppner & Huggins model".
+#     eis : bool, optional
+#         A flag to build the forward model for EIS predictions. Defaults to False.
+#     **model_kwargs : optional
+#         Valid PyBaMM model option keys and their values, for example:
+#         parameter_set : pybamm.ParameterValues or dict, optional
+#             The parameters for the model. If None, default parameters provided by PyBaMM are used.
+#         geometry : dict, optional
+#             The geometry definitions for the model. If None, default geometry from PyBaMM is used.
+#         submesh_types : dict, optional
+#             The types of submeshes to use. If None, default submesh types from PyBaMM are used.
+#         var_pts : dict, optional
+#             The discretization points for each variable in the model. If None, default points from PyBaMM are used.
+#         spatial_methods : dict, optional
+#             The spatial methods used for discretization. If None, default spatial methods from PyBaMM are used.
+#         solver : pybamm.Solver, optional
+#             The solver to use for simulating the model. If None, the default solver from PyBaMM is used.
+#     """
+#
+#     def __init__(
+#         self,
+#         name="Weppner & Huggins model",
+#         eis: bool = False,
+#         **model_kwargs,
+#     ):
+#         super().__init__(
+#             pybamm_model=BaseWeppnerHuggins, name=name, eis=eis, **model_kwargs
+#         )
 
-    Parameters
-    ----------
-    name: str, optional
-        A name for the model instance, defaults to "Weppner & Huggins model".
-    eis : bool, optional
-        A flag to build the forward model for EIS predictions. Defaults to False.
-    **model_kwargs : optional
-        Valid PyBaMM model option keys and their values, for example:
-        parameter_set : pybamm.ParameterValues or dict, optional
-            The parameters for the model. If None, default parameters provided by PyBaMM are used.
-        geometry : dict, optional
-            The geometry definitions for the model. If None, default geometry from PyBaMM is used.
-        submesh_types : dict, optional
-            The types of submeshes to use. If None, default submesh types from PyBaMM are used.
-        var_pts : dict, optional
-            The discretization points for each variable in the model. If None, default points from PyBaMM are used.
-        spatial_methods : dict, optional
-            The spatial methods used for discretization. If None, default spatial methods from PyBaMM are used.
-        solver : pybamm.Solver, optional
-            The solver to use for simulating the model. If None, the default solver from PyBaMM is used.
-    """
 
-    def __init__(
-        self,
-        name="Weppner & Huggins model",
-        eis: bool = False,
-        **model_kwargs,
-    ):
-        super().__init__(
-            pybamm_model=BaseWeppnerHuggins, name=name, eis=eis, **model_kwargs
-        )
-
-
-class GroupedSPMe(EChemBaseModel):
-    """
-    Represents the grouped-parameter version of the SPMe.
-
-    Parameters
-    ----------
-    name: str, optional
[…]
