+++ pybop/models/lithium_ion/__init__.py
@@ -2,4 +2,6 @@
 # Import lithium ion based models
 #
 from .base_echem import EChemBaseModel
-from .echem import SPM, SPMe, DFN, MPM, MSMR, WeppnerHuggins, GroupedSPMe
+# from .echem import SPM, SPMe, DFN, MPM, MSMR, WeppnerHuggins, GroupedSPMe
+from .basic_SPMe import GroupedSPMe
+from .weppner_huggins import WeppnerHuggins
