+++ pybop/models/lithium_ion/weppner_huggins.py
@@ -6,11 +6,9 @@ from pybamm import lithium_ion as pybamm_lithium_ion
 from pybamm import t as pybamm_t
 
 
-class BaseWeppnerHuggins(pybamm_lithium_ion.BaseModel):
+class WeppnerHuggins(pybamm_lithium_ion.BaseModel):
     """
-    WeppnerHuggins Model for GITT. Credit: pybamm-param team.
-
-    This model can be used with PyBOP through the `pybop.WeppnerHuggins` class.
+    Represents the Weppner & Huggins model to fit diffusion coefficients to GITT data.
 
     Parameters
     ----------
