### 2025-05-14 72b94c0
update docstring on pints optimiser
