### 2025-05-14 76504be
refactor scipy optimisers
