+++ pybop/optimisers/scipy_optimisers.py
@@ -6,6 +6,7 @@ import numpy as np
 from scipy.optimize import Bounds, OptimizeResult, differential_evolution, minimize
 
 from pybop import BaseOptimiser, OptimisationResult, SciPyEvaluator
+from pybop.problems.base_problem import Problem
 
 
 class BaseSciPyOptimiser(BaseOptimiser):
@@ -22,50 +23,26 @@ class BaseSciPyOptimiser(BaseOptimiser):
         Valid SciPy option keys and their values.
     """
 
-    def __init__(self, cost, **optimiser_kwargs):
-        self.num_resamples = 40
-        self.key_mapping = {"max_iterations": "maxiter", "population_size": "popsize"}
-        super().__init__(cost, **optimiser_kwargs)
+    def __init__(
+        self,
+        problem: Problem,
+        multistart: int = 1,
+    ):
+        super().__init__(problem, multistart=multistart)
 
-    def _sanitise_inputs(self):
-        """
-        Check and remove any duplicate optimiser options.
-        """
-        # Unpack values from any nested options dictionary
-        if "options" in self.unset_options.keys():
-            key_list = list(self.unset_options["options"].keys())
-            for key in key_list:
-                if key not in self.unset_options.keys():
-                    self.unset_options[key] = self.unset_options["options"].pop(key)
-                else:
-                    raise Exception(
-                        f"A duplicate {key} option was found in the options dictionary."
-                    )
-            self.unset_options.pop("options")
-
-        # Convert PyBOP keys to SciPy
-        for pybop_key, scipy_key in self.key_mapping.items():
-            if pybop_key in self.unset_options:
-                if scipy_key in self.unset_options:
-                    raise Exception(
-                        f"The alternative {pybop_key} option was passed in addition to the expected {scipy_key} option."
-                    )
-                # Rename the key
-                self.unset_options[scipy_key] = self.unset_options.pop(pybop_key)
-
+    def scipy_bounds(self) -> Bounds:
+        bounds = self.problem.params.get_bounds()
         # Convert bounds to SciPy format
-        if isinstance(self.bounds, dict):
-            self._scipy_bounds = Bounds(
-                self.bounds["lower"], self.bounds["upper"], True
-            )
-        elif isinstance(self.bounds, Bounds) or self.bounds is None:
-            self._scipy_bounds = self.bounds
+        if isinstance(bounds, dict):
+            return Bounds(bounds["lower"], bounds["upper"], True)
+        elif isinstance(bounds, Bounds) or bounds is None:
+            return bounds
         else:
             raise TypeError(
                 "Bounds provided must be either type dict or SciPy.optimize.bounds object."
             )
 
-    def _run(self):
+    def evaluator(self) -> SciPyEvaluator:
         """
         Internal method to run the optimisation using a PyBOP optimiser.
 
@@ -82,26 +59,7 @@ class BaseSciPyOptimiser(BaseOptimiser):
             )
 
         # Create evaluator object
-        self.evaluator = SciPyEvaluator(fun)
-
-        # Run with timing
-        start_time = time()
-        result = self._run_optimiser()
-        total_time = time() - start_time
-
-        try:
-            nit = result.nit
-        except AttributeError:
-            nit = -1
-
-        return OptimisationResult(
-            optim=self,
-            x=result.x,
-            n_iterations=nit,
-            scipy_result=result,
-            time=total_time,
-            message=result.message,
-        )
+        return SciPyEvaluator(fun)
 
 
 class SciPyMinimize(BaseSciPyOptimiser):
@@ -113,6 +71,10 @@ class SciPyMinimize(BaseSciPyOptimiser):
 
     Parameters
     ----------
+    problem : pybop.Problem
[…]
