### 2025-05-15 a95ccf1
feat: updates to python builder, problem, adds BaseCost args for explicit model vars and data columns
