+++ pybop/builders/python.py
@@ -1,49 +1,119 @@
-import pybop
+from typing import Callable, Union
+
+import numpy as np
+
 from pybop import PythonProblem
 from pybop.builders import BaseBuilder
-from pybop.problems.base_problem import Problem
+from pybop.costs.base_cost import BaseCost
 
 
 class Python(BaseBuilder):
     """
-    dataset : pybop.Dataset or dict, optional
-        The dataset to be used in the simulation construction.
+    Builder for Python-based optimisation problems.
+
+    This builder allows for the creation of optimisation problems using custom
+    Python functions instead of specialised simulation frameworks.
+
     """
 
     def __init__(self):
-        self._model = None
-        self._params = None
-        self._costs = None
-        self._cost_weights = []
+        super().__init__()
+        self._model_func = None
+        self._model_func_with_sens = None
 
-    def add_func(self, func):
-        self._func = func
+    def set_simulation(
+        self,
+        model_func: Callable,
+        model_func_with_sens: Callable = None,
+    ) -> None:
+        """
+        Set the simulation functions for the problem.
 
-    def add_cost(self, cost: pybop.BaseCost, weight: float = 1.0) -> None:
+        Parameters
+        ----------
+        model_func : Callable
+            Function that takes parameters and dataset, returns simulation results
+            Expected signature: func(params: dict, dataset: dict) -> dict
+        model_func_with_sens : Callable, optional
+            Function that returns both results and sensitivities
+            Expected signature: func(params: dict, dataset: dict) -> tuple[dict, np.ndarray]
         """
-        Adds cost to the problem.
-        ToDo: BaseCost needs to be updated to support array/list-like execution
+        if not callable(model_func):
+            raise TypeError("model_func must be a callable")
+
+        self._model_func = model_func
+        self._model_func_with_sens = model_func_with_sens
+
+    def add_cost(self, cost: Union[BaseCost, Callable], weight: float = 1.0) -> None:
         """
+        Add a cost component to the problem.
+
+        Parameters
+        ----------
+        cost : Union[BaseCost, Callable]
+            Cost function
+        weight : float, optional
+            Weight for this cost component, by default 1.0
+        """
+        if isinstance(cost, BaseCost):
+            if cost.weighting is None or cost.weighting == "equal":
+                cost.weighting = 1.0
+            elif cost.weighting == "domain" and self.domain is not None:
+                self._set_cost_domain_weighting(cost)
+
         self._costs.append(cost)
         self._cost_weights.append(weight)
 
-    def build(self) -> Problem:
-        """Build a pure python problem."""
+    def _set_cost_domain_weighting(self, cost):
+        """Calculate domain-based weighting."""
+        domain_data = self._dataset[self.domain]
+        domain_spacing = domain_data[1:] - domain_data[:-1]
+        mean_spacing = np.mean(domain_spacing)
 
-        if not len(self._cost_weights) == len(self._costs):
-            raise ValueError(
-                "Number of cost weights and the number of costs do not match"
+        # Create a domain weighting array in one operation
+        cost.weighting = np.concatenate(
+            (
+                [(mean_spacing + domain_spacing[0]) / 2],
+                (domain_spacing[1:] + domain_spacing[:-1]) / 2,
+                [(domain_spacing[-1] + mean_spacing) / 2],
             )
+        ) * ((len(domain_data) - 1) / (domain_data[-1] - domain_data[0]))
 
-        if self._func is None:
-            raise ValueError("A Pybamm model needs to be provided before building.")
+    def build(self) -> PythonProblem:
+        """
+        Build the Python problem with the configured components.
 
-        if self._costs is None:
-            raise ValueError("A cost must be provided before building.")
+        Returns
+        -------
+        PythonProblem
+            The constructed optimisation problem
+
+        Raises
+        ------
+        ValueError
[…]
