+++ pybop/costs/base_cost.py
@@ -14,8 +14,15 @@ class BaseCost:
         measure. Options: "equal"(default), "domain", or a custom numpy array.
     """
 
-    def __init__(self, weighting: Union[str, np.ndarray] = None):
+    def __init__(
+        self,
+        model_variable: str,
+        data_column: str,
+        weighting: Union[str, np.ndarray] = None,
+    ):
         self.weighting = weighting
+        self._model_variable = [model_variable]
+        self._data_column = [data_column]
 
     def __call__(
         self,
@@ -41,3 +48,11 @@ class BaseCost:
             gradient with dimension (len(parameters)), otherwise returns only the cost.
         """
         raise NotImplementedError
+
+    @property
+    def model_variable(self):
+        return self._model_variable
+
+    @property
+    def data_column(self):
+        return self._data_column
