### 2025-05-15 f0a372d
tests: removing model unit tests
