+++ tests/unit/test_models.py
@@ -17,39 +17,6 @@ class TestModels:
 
     pytestmark = pytest.mark.unit
 
-    @pytest.mark.parametrize(
-        "model_class, expected_name, options",
-        [
-            (pybop.lithium_ion.SPM, "Single Particle Model", None),
-            (pybop.lithium_ion.SPMe, "Single Particle Model with Electrolyte", None),
-            (pybop.lithium_ion.DFN, "Doyle-Fuller-Newman", None),
-            (pybop.lithium_ion.MPM, "Many Particle Model", None),
-            (
-                pybop.lithium_ion.MSMR,
-                "Multi Species Multi Reactions Model",
-                {"number of MSMR reactions": ("6", "4")},
-            ),
-            (pybop.lithium_ion.WeppnerHuggins, "Weppner & Huggins model", None),
-            (
-                pybop.lithium_ion.GroupedSPMe,
-                "Grouped Single Particle Model with Electrolyte",
-                None,
-            ),
-            (
-                pybop.lithium_ion.GroupedSPMe,
-                "Grouped Single Particle Model with Electrolyte",
-                {"surface form": "differential"},
-            ),
-            (pybop.empirical.Thevenin, "Equivalent Circuit Thevenin Model", None),
-        ],
-    )
-    def test_model_classes(self, model_class, expected_name, options):
-        parameter_set = pybop.ParameterSet({"Nominal cell capacity [A.h]": 5.12})
-        model = model_class(options=options, parameter_set=parameter_set)
-        assert model.pybamm_model is not None
-        assert model.name == expected_name
-        assert model.parameter_set["Nominal cell capacity [A.h]"] == 5.12
-
     @pytest.fixture(
         params=[
             pybop.lithium_ion.SPM(),
@@ -66,99 +33,6 @@ class TestModels:
     def model(self, request):
         model = request.param
         return model.copy()
-
-    def test_non_default_solver(self):
-        solver = pybamm.CasadiSolver(
-            mode="fast",
-            atol=1e-6,
-            rtol=1e-6,
-        )
-        model = pybop.lithium_ion.SPM(solver=solver)
-        assert model.solver.mode == "fast"
-        assert model.solver.atol == 1e-6
-        assert model.solver.rtol == 1e-6
-
-    def test_predict_without_pybamm(self, model):
-        model.pybamm_model = None
-
-        with pytest.raises(
-            ValueError,
-            match="The predict method currently only supports PyBaMM models.",
-        ):
-            model.predict(None, None)
-
-        # Test new_copy() without pybamm_model
-        if not isinstance(
-            model, (pybop.lithium_ion.MSMR, pybop.lithium_ion.GroupedSPMe)
-        ):
-            new_model = model.new_copy()
-            assert new_model.pybamm_model is not None
-            assert new_model.parameter_set is not None
-
-    def test_predict_with_inputs(self, model):
-        # Define inputs
-        t_eval = np.linspace(0, 10, 100)
-        if isinstance(model, (pybop.lithium_ion.GroupedSPMe)):
-            inputs = {
-                "Negative electrode relative porosity": 0.52,
-                "Positive electrode relative porosity": 0.63,
-            }
-        elif isinstance(model, (pybop.lithium_ion.EChemBaseModel)):
-            if model.pybamm_model.options["working electrode"] == "positive":
-                inputs = {
-                    "Positive electrode active material volume fraction": 0.63,
-                }
-            else:
-                inputs = {
-                    "Negative electrode active material volume fraction": 0.52,
-                    "Positive electrode active material volume fraction": 0.63,
-                }
-        elif isinstance(model, (pybop.empirical.Thevenin)):
-            inputs = {
-                "R0 [Ohm]": 0.0002,
[…]
