### 2025-05-15 afdbac2
Create msmr_fitting.py
