+++ examples/scripts/battery_parameterisation/msmr_fitting.py
@@ -14,201 +14,202 @@ from scipy import constants
 
 import pybop
 
-# Generate some synthetic OCP data for testing
 parameter_set = pybop.ParameterSet("Chen2020")
-cell_type = "positive electrode"  # "negative electrode" or "positive electrode"
-if cell_type == "positive electrode":
-    ocp_function = parameter_set["Positive electrode OCP [V]"]
-    soc_min, soc_max = 0.266145163492257, 0.7  # 0.905926128940627
-else:
-    ocp_function = parameter_set["Negative electrode OCP [V]"]
-    soc_min, soc_max = 0.0312962309919435, 0.7  # 0.901446800739041
 
-soc = np.linspace(soc_min, soc_max, 101)
-ocp = ocp_function(soc)
-weighting = 1e-3
-ocp_dataset = pybop.Dataset(
-    {
-        "SoC": soc,
-        "Voltage [V]": ocp,
-        "Weighted dSoC/dV": weighting * np.gradient(soc, ocp),
-    },
-    domain="SoC",
-)
+for cell_type in ["negative electrode", "positive electrode"]:
+    # Generate some synthetic OCP data for testing
+    if cell_type == "positive electrode":
+        ocp_function = parameter_set["Positive electrode OCP [V]"]
+        soc_min, soc_max = 0.266145163492257, 0.7  # 0.905926128940627
+    else:
+        ocp_function = parameter_set["Negative electrode OCP [V]"]
+        soc_min, soc_max = 0.0312962309919435, 0.7  # 0.901446800739041
 
-# Define the MSMR model parameters using a set number of electrode reactions
-n_reactions = 4
-T = 298.15
-F_RT = constants.physical_constants["Faraday constant"][0] / (constants.R * T)
-if cell_type == "positive electrode":
-    msmr_params = {
-        "X_shift": 0.340,
-        "U0_0": 3.751,
-        "X_0": 0.311,
-        "w_0": 1.647,
-        "U0_1": 4.002,
-        "X_1": 0.382,
-        "w_1": 3.011,
-        "U0_2": 4.185,
-        "X_2": 0.087,
-        "w_2": 0.297,
-        "U0_3": 4.283,
-        "X_3": 0.219,
-        "w_3": 4.361,
-    }
-else:
-    msmr_params = {
-        "X_shift": 0.020,
-        "U0_0": 0.090,
-        "X_0": 0.622,
-        "w_0": 0.042,
-        "U0_1": 0.134,
-        "X_1": 0.183,
-        "w_1": 0.108,
-        "U0_2": 0.215,
-        "X_2": 0.105,
-        "w_2": 0.780,
-        "U0_3": 0.393,
-        "X_3": 0.090,
-        "w_3": 5.360,
-    }
-
-
-def sub(s: str, j: int):
-    return s + "_" + str(j)
-
-
-class FunctionFitting(pybop.FittingProblem):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.parameter_values = msmr_params
-
-        # Define the state of charge
-        def soc(U):
-            X = self.parameter_values["X_shift"]
-            for j in range(n_reactions):
-                X += self.p("X", j) / (
-                    1 + np.exp(F_RT * (U - self.p("U0", j)) / self.p("w", j))
-                )
-            return X
-
-        self.soc = soc
-
-    def p(self, s: str, j: int):
-        return self.parameter_values[sub(s, j)]
-
-    def evaluate(self, inputs=None):
-        inputs = self.parameters.verify(inputs)
-        self.parameter_values.update(inputs)
-
-        soc = []
-        for voltage in self.domain_data:
-            try:
-                soc.append(self.soc(voltage))
-            except (ZeroDivisionError, RuntimeError, ValueError) as e:
[…]
