### 2025-05-16 0aeaa13
optimisers running, many tests still to go
