+++ pybop/__init__.py
@@ -171,12 +171,18 @@ from .optimisers._simulated_annealing import SimulatedAnnealingImpl
 from .optimisers._irprop_plus import IRPropPlusImpl
 from .optimisers._cost_interface import CostInterface
 from .optimisers._result import OptimisationResult
-from .optimisers.base_optimiser import BaseOptimiser, OptimisationLogger
+from .optimisers.base_optimiser import (
+    BaseOptimiser,
+    OptimisationLogger,
+    OptimiserOptions,
+)
 from .optimisers.base_pints_optimiser import BasePintsOptimiser, PintsOptions
 from .optimisers.scipy_optimisers import (
     BaseSciPyOptimiser,
     SciPyMinimize,
+    ScipyMinimizeOptions,
     SciPyDifferentialEvolution,
+    SciPyDifferentialEvolutionOptions,
 )
 from .optimisers.pints_optimisers import (
     GradientDescent,
@@ -192,7 +198,6 @@ from .optimisers.pints_optimisers import (
     AdamW,
     SimulatedAnnealing,
 )
-from .optimisers.optimisation import Optimisation
 
 #
 # Monte Carlo classes
