+++ pybop/optimisers/base_pints_optimiser.py
@@ -1,4 +1,6 @@
+from dataclasses import dataclass
 from time import time
+from typing import Optional
 
 import numpy as np
 import pints
@@ -24,24 +26,22 @@ from pybop import (
 )
 
 
-class PintsOptions:
+@dataclass
+class PintsOptions(pybop.OptimiserOptions):
     """
     A class to hold PINTS options for the optimisation process.
     """
 
     default_max_iterations = 1000
-
-    def __init__(self):
-        self.max_iterations: int = self.default_max_iterations
-        self.min_iterations: int = 2
-        self.max_unchanged_iterations: int = 15
-        self.multistart: int = 1
-        self.parallel: bool = False
-        self.use_f_guessed: bool = False
-        self.absolute_tolerance: float = 1e-5
-        self.relative_tolerance: float = 1e-2
-        self.max_evaluations: int = None
-        self.threshold: float = None
+    max_iterations: int = default_max_iterations
+    min_iterations: int = 2
+    max_unchanged_iterations: int = 15
+    parallel: bool = False
+    use_f_guessed: bool = False
+    absolute_tolerance: float = 1e-5
+    relative_tolerance: float = 1e-2
+    max_evaluations: Optional[int] = None
+    threshold: Optional[float] = None
 
 
 class BasePintsOptimiser(pybop.BaseOptimiser):
@@ -62,7 +62,7 @@ class BasePintsOptimiser(pybop.BaseOptimiser):
         self,
         problem: Problem,
         pints_optimiser,
-        options: PintsOptions = None,
+        options: Optional[PintsOptions] = None,
     ):
         # First set attributes to default values
         self._boundaries = None
@@ -86,10 +86,24 @@ class BasePintsOptimiser(pybop.BaseOptimiser):
         self._threshold = options.threshold
         self._pints_optimiser = pints_optimiser
         self._optimizer = None
+        logger = pybop.OptimisationLogger()
         super().__init__(
             problem,
-            multistart=options.multistart,
+            options=options,
+            logger=logger,
         )
+
+    @staticmethod
+    def default_options() -> PintsOptions:
+        """
+        Returns the default options for the PINTS optimiser.
+
+        Returns
+        -------
+        PintsOptions
+            The default options for the PINTS optimiser.
+        """
+        return PintsOptions()
 
     @property
     def max_iterations(self):
@@ -98,7 +112,7 @@ class BasePintsOptimiser(pybop.BaseOptimiser):
         """
         return self._max_iterations
 
-    def set_max_iterations(self, iterations="default"):
+    def set_max_iterations(self, iterations: str | int | None = "default"):
         """
         Set the maximum number of iterations as a stopping criterion.
         Credit: PINTS
@@ -135,7 +149,7 @@ class BasePintsOptimiser(pybop.BaseOptimiser):
             if issubclass(self._pints_optimiser, PintsPSO):
                 if not all(
                     np.isfinite(value)
-                    for sublist in self.bounds.values()
+                    for sublist in bounds.values()
                     for value in sublist
                 ):
                     raise ValueError(
@@ -200,32 +214,28 @@ class BasePintsOptimiser(pybop.BaseOptimiser):
         if self._needs_sensitivities:
 
             def fun(x):
-                model_x = self.problem.params.to_model(x)
-                self.problem.set_params(model_x)
-                self.problem.run_with_sensitivities()
+                self.problem.set_params(x)
+                return self.problem.run_with_sensitivities()
 
         else:
 
             def fun(x):
-                model_x = self.problem.params.to_model(x)
-                self.problem.set_params(model_x)
-                self.problem.run()
+                self.problem.set_params(x)
+                return self.problem.run()
 
[…]
