+++ pybop/plot/voronoi.py
@@ -1,5 +1,3 @@
-from typing import Union
-
 import numpy as np
 from scipy.spatial import Voronoi, cKDTree
 
