### 2025-05-19 009bf78
feat: adds remaining costs to pybamm costs, enables cost hyper identification
