+++ pybop/costs/pybamm/__init__.py
@@ -1,0 +1,7 @@
+from .pybamm_cost import (
+    BaseCost,
+    PybammExpressionMetadata,
+    PybammParameterMetadata,
+    SumSquaredError,
+    MeanAbsoluteError
+)
