### 2025-05-19 825068f
bug: fixed optimiser tests (and a few others)
