+++ pybop/builders/__init__.py
@@ -1,0 +1,4 @@
+from .base import BaseBuilder
+from .python import Python
+from .pybamm_eis import PybammEIS
+from .pybamm import Pybamm
