### 2025-05-19 f2a1987
style: pre-commit fixes
