+++ pybop/builders/python.py
@@ -1,6 +1,7 @@
 import pybop
 from pybop import PythonProblem
 from pybop.problems.base_problem import Problem
+
 from .base import BaseBuilder
 
 
