+++ pybop/problems/base_problem.py
@@ -1,4 +1,5 @@
 from typing import Optional
+
 import numpy as np
 
 from pybop import Parameters
