### 2025-05-20 7cb6b23
feat: add NegativeGaussianLogLikelihood, sensitivities shape
