+++ pybop/costs/pybamm/pybamm_cost.py
@@ -1,11 +1,12 @@
 from dataclasses import dataclass
-from typing import Optional
+from typing import Optional, Union
 from uuid import uuid4
 
 import numpy as np
 import pybamm
 
-import pybop
+from pybop import Dataset
+from pybop import Parameter as PybopParameter
 
 
 @dataclass
@@ -59,7 +60,7 @@ class BaseCost:
     def variable_expression(
         self,
         model: pybamm.BaseModel,
-        dataset: Optional[pybop.Dataset] = None,
+        dataset: Optional[Dataset] = None,
     ) -> PybammExpressionMetadata:
         """
         Defines the variable expression for the cost function, returning a
@@ -73,7 +74,7 @@ class BaseCost:
         self,
         model: pybamm.BaseModel,
         param: pybamm.ParameterValues,
-        dataset: Optional[pybop.Dataset] = None,
+        dataset: Optional[Dataset] = None,
     ):
         # if dataset is provided, must contain time data
         if dataset is not None and "Time [s]" not in dataset:
@@ -86,6 +87,16 @@ class BaseCost:
                 check_already_exists=False,
             )
 
+    def _construct_discrete_time_node(self, dataset, model, name):
+        """
+        Constructs the pybamm DiscreteTimeData node in the expression tree
+        """
+        times = dataset["Time [s]"]
+        values = dataset[self._data_name]
+        data = pybamm.DiscreteTimeData(times, values, f"{name}_data")
+        var = model.variables[self._variable_name]
+        return data, var
+
     def _check_state(self, dataset, model, name) -> None:
         # dataset must be provided and contain the data
         if dataset is None:
@@ -96,6 +107,26 @@ class BaseCost:
         if self._variable_name not in model.variables:
             raise ValueError(f"Model must contain {self._variable_name} for {name}.")
 
+    def _get_sigma_parameter(self, cost_name, parameters) -> pybamm.Parameter:
+        """
+        Returns the sigma node, either fixed or as an estimated parameter.
+        Updates metadata if sigma is estimated.
+        """
+
+        if isinstance(self._sigma, PybopParameter) or self._sigma is None:
+            sigma_name = self._sigma.name if self._sigma else f"sigma_{cost_name}"
+            sigma = pybamm.Parameter(sigma_name)
+            parameters[sigma_name] = PybammParameterMetadata(
+                parameter=sigma,
+                default_value=self._sigma.value if self._sigma else 1.0,
+            )
+        elif isinstance(self._sigma, (float, int)):
+            sigma = pybamm.Scalar(self._sigma)
+        else:
+            sigma = self._sigma  # Assume pybamm.Parameter
+
+        return sigma
+
 
 class SumSquaredError(BaseCost):
     """
@@ -103,7 +134,10 @@ class SumSquaredError(BaseCost):
     """
 
     def __init__(
-        self, variable_name: str, data_name: str, sigma: Optional[float] = None
+        self,
+        variable_name: str,
+        data_name: str,
+        sigma: Optional[Union[float, PybopParameter]] = None,
     ):
         super().__init__()
         self._sigma = sigma
@@ -113,29 +147,15 @@ class SumSquaredError(BaseCost):
     def variable_expression(
         self,
         model: pybamm.BaseModel,
-        dataset: Optional[pybop.Dataset] = None,
+        dataset: Optional[Dataset] = None,
     ) -> PybammExpressionMetadata:
         # Check args
         name = SumSquaredError.make_unique_cost_name()
         self._check_state(dataset, model, name)
-
-        # Construct cost
-        times = dataset["Time [s]"]
-        values = dataset[self._data_name]
-        data = pybamm.DiscreteTimeData(times, values, f"{name}_data")
[…]
