### 2025-05-21 a247133
Update pybop/optimisers/_adamw.py
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
