### 2025-05-21 28e35ee
Update pybop/optimisers/_irprop_plus.py
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
