### 2025-05-21 23b56b7
Update pybop/optimisers/base_optimiser.py
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
