### 2025-05-21 002b831
Update pybop/optimisers/base_pints_optimiser.py
Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>
