### 2025-05-22 bca1650
fix and ignore more linter errors
