### 2025-05-22 ab0bb8b
refactor: optimisers (#729)
* drafting out changes

* refactor pints optimisers

* update docstring on pints optimiser

* refactor scipy optimisers

* optimisers running, many tests still to go

* style: pre-commit fixes

* bug: fixed optimiser tests (and a few others)

* style: pre-commit fixes

* Update pybop/optimisers/_adamw.py

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

* Update pybop/optimisers/_irprop_plus.py

Co-authored-by: Brady Planden <55357039+BradyPlanden@users.noreply.github.com>

* Update pybop/optimisers/base_optimiser.py

Co-authored-by: B
[…]
