+++ pybop/problems/pybamm_eis_problem.py
@@ -1,4 +1,4 @@
-from typing import Union
+from typing import Optional, Union
 
 import numpy as np
 
@@ -16,16 +16,16 @@ class PybammEISProblem(Problem):
     def __init__(
         self,
         eis_pipeline: PybammEISPipeline,
-        pybop_params: Parameters = None,
-        costs: list = None,
-        cost_weights: Union[list, np.array] = None,
-        fitting_data: list = None,
+        pybop_params: Optional[Parameters] = None,
+        costs: Optional[list] = None,
+        cost_weights: Union[list, np.ndarray, None] = None,
+        fitting_data: Optional[list] = None,
     ):
-        super().__init__(pybop_params=pybop_params)
         self._pipeline = eis_pipeline
         self._costs = costs
         self._cost_weights = cost_weights
         self._fitting_data = fitting_data
+        super().__init__(pybop_params=pybop_params)
 
     def set_params(self, p: np.ndarray) -> None:
         """
