+++ pybop/builders/__init__.py
@@ -1,5 +1,4 @@
-
 from .base import BaseBuilder
-from .pybamm import Pybamm
 from .python import Python
 from .pybamm_eis import PybammEIS
+from .pybamm import Pybamm
