+++ pybop/optimisers/_result.py
@@ -1,13 +1,8 @@
-import warnings
-from typing import TYPE_CHECKING, Optional, Union
+from typing import Optional
 
 import numpy as np
-from pybamm import Solution
 
-from pybop import BaseCost, BaseLikelihood, Inputs
-
-if TYPE_CHECKING:
-    from pybop import BaseOptimiser
+from pybop.problems.base_problem import Problem
 
 
 class OptimisationResult:
@@ -16,10 +11,10 @@ class OptimisationResult:
 
     Attributes
     ----------
-    optim : pybop.BaseOptimiser
+    problem: pybop.Problem
         The optimisation object used to generate the results.
     x : ndarray
-        The solution of the optimisation.
+        The solution of the optimisation (in model space)
     final_cost : float
         The cost associated with the solution x.
     n_iterations : int
@@ -28,133 +23,94 @@ class OptimisationResult:
         The reason for stopping given by the optimiser.
     scipy_result : scipy.optimize.OptimizeResult, optional
         The result obtained from a SciPy optimiser.
-    pybamm_solution: pybamm.Solution or list[pybamm.Solution], optional
-        The best solution object(s) obtained from the optimisation.
     """
 
     def __init__(
         self,
-        optim: "BaseOptimiser",
-        x: Union[Inputs, np.ndarray] = None,
-        final_cost: Optional[float] = None,
-        sensitivities: Optional[dict] = None,
-        n_iterations: Optional[int] = None,
-        n_evaluations: Optional[int] = None,
-        time: Optional[float] = None,
+        problem: Problem,
+        x: np.ndarray,
+        final_cost: float,
+        n_iterations: int,
+        n_evaluations: int,
+        time: float,
         message: Optional[str] = None,
-        scipy_result=None,
     ):
-        self.optim = optim
-        self.cost = self.optim.cost
-        self.minimising = not self.optim.invert_cost
-        self._transformation = self.optim.transformation
+        self._problem = problem
         self.n_runs = 0
         self._best_run = None
-        self._x = []
-        self._final_cost = []
-        self._sensitivities = None
-        self._fisher = []
-        self._n_iterations = []
-        self._n_evaluations = []
-        self._message = []
-        self._scipy_result = []
-        self._time = []
-        self._x0 = []
-        self._pybamm_solution = []
+        self._x = [x]
+        self._final_cost = [final_cost]
+        self._n_iterations = [n_iterations]
+        self._n_evaluations = [n_evaluations]
+        self._message = [message]
+        self._time = [time]
 
-        if x is not None:
-            # Transform the parameter values and update the sign of any final cost
-            # coming directly from an optimiser
-            x = self._transformation.to_model(x) if self._transformation else x
-            final_cost = (
-                final_cost * (1 if self.minimising else -1)
-                if final_cost is not None
-                else self.cost(x)
-            )
-            x0 = self.optim.parameters.initial_value()
+        x0 = self._problem.params.initial_value()
+        self._x0 = [x0]
 
-            # Evaluate the problem once more to update the solution
-            try:
-                self.cost(x)
-                pybamm_solution = self.cost.pybamm_solution
-            except Exception:
-                warnings.warn(
-                    "Failed to evaluate the model with best fit parameters.",
-                    UserWarning,
-                    stacklevel=2,
-                )
-                pybamm_solution = None
+        # Calculate Fisher Information if available
+        try:
+            fisher = self._problem.observed_fisher(x)
+            diag_fish = np.diag(fisher) if fisher is not None else None
+        except NotImplementedError:
+            diag_fish = None
+        self._fisher = [diag_fish]
 
-            # Calculate Fisher Information if Likelihood
[…]
