+++ pybop/problems/python_problem.py
@@ -3,6 +3,7 @@ from typing import Callable
 import numpy as np
 
 from pybop import Parameters
+from pybop.costs.base_cost import BaseCost
 from pybop.problems.base_problem import Problem
 
 
@@ -15,9 +16,9 @@ class PythonProblem(Problem):
 
     Parameters
     ----------
-    model_func : Callable
+    model : Callable
         The function that performs the simulation
-    model_func_with_sens : Callable, optional
+    model_with_sens : Callable, optional
         Function that performs simulation and returns sensitivities
     pybop_params : Parameters
         Container for optimisation parameters
@@ -29,21 +30,16 @@ class PythonProblem(Problem):
 
     def __init__(
         self,
-        model_func: Callable,
-        model_func_with_sens: Callable = None,
+        model: Callable,
+        model_with_sens: Callable = None,
         pybop_params: Parameters = None,
-        costs: list[str] = None,
+        costs: list[BaseCost] = None,
         cost_weights: list[float] = None,
         dataset: dict[str, np.ndarray] = None,
     ):
         super().__init__(pybop_params=pybop_params)
-        if not callable(model_func):
-            raise TypeError("model_func must be a callable function.")
-        if model_func_with_sens is not None and not callable(model_func_with_sens):
-            raise TypeError("model_func_with_sens must be a callable function.")
-
-        self._model_func = model_func
-        self._model_func_with_sens = model_func_with_sens
+        self._model = model
+        self._model_with_sens = model_with_sens
         self._costs = costs
         self._cost_weights = cost_weights
         self._dataset = dataset
@@ -60,14 +56,32 @@ class PythonProblem(Problem):
         self.check_set_params_called()
 
         # Run the model with current parameters
-        results = self._model_func(self.params, self._dataset)
+        results = self._model(self.params.current_value())
 
-        costs = [
-            cost(results[signal] - self._dataset[column])
-            for cost in self._costs
-            for column, signal in zip(cost.model_variable, cost.data_column)
-        ]
-        return np.dot(costs, self._cost_weights)
+        total_cost: float = 0.0
+
+        # Calculate weighted costs and gradients
+        for i, cost_func in enumerate(self._costs):
+            if hasattr(cost_func, "model_variable") and hasattr(
+                cost_func, "data_column"
+            ):
+                for column, signal in zip(
+                    cost_func.data_column, cost_func.model_variable
+                ):
+                    if signal not in results or column not in self._dataset:
+                        raise KeyError(
+                            f"Signal '{signal}' or column '{column}' not found"
+                        )
+
+                    residual = results[signal] - self._dataset[column]
+                    val = cost_func(residual)
+                    total_cost += val * self._cost_weights[i]
+            else:
+                raise RuntimeError(
+                    "Cost function must have model_variable and data_column"
+                )
+
+        return total_cost
 
     def run_with_sensitivities(self) -> tuple[float, np.ndarray]:
         """
@@ -79,22 +93,43 @@ class PythonProblem(Problem):
             The cost value and array of parameter gradients
         """
         self.check_set_params_called()
-
-        if self._model_func_with_sens is None:
+        if self._model_with_sens is None:
             raise RuntimeError("No sensitivity function provided")
 
         # Run the model with current parameters
-        results, sens = self._model_func_with_sens(
-            self.params, self._dataset
-        )  # ToDo: sense needs a better format (combined dict.sens?, or just dict)
+        results, sens = self._model_with_sens(self.params.current_value())
 
-        costs = [
-            cost(results[signal] - self._dataset[column], dy=sens)
-            for cost in self._costs
-            for column, signal in zip(cost.model_variable, cost.data_column)
-        ]
+        if results is None or sens is None:
[…]
