### 2025-05-22 7800828
docs: update model signature
