### 2025-05-22 db8af74
docs: update base builder docstring
