### 2025-05-22 1e22537
feat: adds posterior functionality
