### 2025-05-22 1bcf8f1
fix: likelihoods are negative, update for priors
