### 2025-05-22 a5b3939
feat: adds checks for posterior w/o priors
