### 2025-05-23 db325af
feat: initial cost tests, adds salib sensitivities to BaseProblem, rename parameters.get_bounds_for_plotly to `bounds_as_numpy`
