+++ tests/unit/test_cost.py
@@ -1,5 +1,3 @@
-from copy import copy
-
 import numpy as np
 import pybamm
 import pytest
@@ -15,136 +13,109 @@ class TestCosts:
     pytestmark = pytest.mark.unit
 
     @pytest.fixture
-    def model(self, ground_truth):
-        model = pybop.lithium_ion.SPM()
-        model.parameter_set["Negative electrode active material volume fraction"] = (
-            ground_truth
-        )
-        return model
+    def model(self):
+        return pybamm.lithium_ion.SPM()
 
     @pytest.fixture
-    def ground_truth(self):
-        return 0.52
-
-    @pytest.fixture
-    def parameters(self):
-        return pybop.Parameters(
-            pybop.Parameter(
-                "Negative electrode active material volume fraction",
-                prior=pybop.Gaussian(0.5, 0.01),
-                bounds=[0.375, 0.625],
-            )
-        )
-
-    @pytest.fixture
-    def experiment(self):
-        return pybop.Experiment(["Discharge at 1C for 10 minutes (20 second period)"])
-
-    @pytest.fixture
-    def dataset(self, model, experiment):
-        solution = model.predict(experiment=experiment)
+    def dataset(self):
         return pybop.Dataset(
             {
-                "Time [s]": solution["Time [s]"].data,
-                "Current function [A]": solution["Current [A]"].data,
-                "Voltage [V]": solution["Terminal voltage [V]"].data,
+                "Time [s]": np.linspace(0, 360, 10),
+                "Current function [A]": 1e-2 * np.ones(10),
+                "Voltage [V]": np.ones(10),
             }
         )
 
-    def test_base(self, model, parameters, dataset):
-        problem = pybop.FittingProblem(model, parameters, dataset)
-        for cost_class in [pybop.BaseCost, pybop.FittingCost, pybop.DesignCost]:
-            base_cost = cost_class(problem)
-            assert base_cost.problem == problem
-            with pytest.raises(NotImplementedError):
-                base_cost([0.5])
-
-    @pytest.fixture(params=[2.5, 3.777])
-    def problem(self, model, parameters, dataset, request):
-        cut_off = request.param
-        model.parameter_set.update({"Lower voltage cut-off [V]": cut_off})
-        return pybop.FittingProblem(model, parameters, dataset)
-
-    @pytest.fixture(
-        params=[
-            pybop.MeanAbsoluteError,
-            pybop.ObserverCost,
-            pybop.LogPosterior,
-        ]
-    )
-    def fitting_cost(self, problem, request):
-        cls = request.param
-        if cls is pybop.MeanAbsoluteError:
-            return cls(problem)
-        elif cls is pybop.LogPosterior:
-            return cls(pybop.GaussianLogLikelihoodKnownSigma(problem, sigma0=0.002))
-        elif cls is pybop.ObserverCost:
-            inputs = problem.parameters.initial_value()
-            state = problem.model.reinit(inputs)
-            n = len(state)
-            sigma_diag = [0.0] * n
-            sigma_diag[0] = 1e-4
-            sigma_diag[1] = 1e-4
-            process_diag = [0.0] * n
-            process_diag[0] = 1e-4
-            process_diag[1] = 1e-4
-            sigma0 = np.diag(sigma_diag)
-            process = np.diag(process_diag)
-            dataset = pybop.Dataset(data_dictionary=problem.dataset)
-            return cls(
-                pybop.UnscentedKalmanFilterObserver(
-                    problem.parameters,
-                    problem.model,
-                    sigma0=sigma0,
-                    process=process,
[…]
