+++ .pre-commit-config.yaml
@@ -4,7 +4,7 @@ ci:
 
 repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.11.10"
+    rev: "v0.11.11"
     hooks:
       - id: ruff
         args: [--fix, --show-fixes]
