### 2025-05-27 fefe1e1
working on error measure updates, pybamm cost non discretetimesum usage
