+++ pybop/costs/pybamm/pybamm_cost.py
@@ -316,10 +316,11 @@ class Minkowski(BaseCost):
 
         # Create Expression
         diff = var - data
+        p = pybamm.Scalar(self._p)
         abs_diff = pybamm.AbsoluteValue(diff)
-        powered_diff = abs_diff**self._p
+        powered_diff = pybamm.Power(abs_diff, p)
         sum_powered = pybamm.DiscreteTimeSum(powered_diff)
-        expression = sum_powered ** (1 / self._p)
+        expression = pybamm.Power(sum_powered, 1 / p)
 
         return PybammExpressionMetadata(
             variable_name=name,
