### 2025-06-04 a63db0e
feat: add experiment functionality to pybamm pipeline
