+++ pybop/_pybamm_pipeline.py
@@ -1,9 +1,19 @@
+import warnings
 from copy import copy, deepcopy
+from functools import lru_cache
+from typing import Union
 
 import numpy as np
 import pybamm
 
 from pybop import Inputs, Parameters
+
+
+class OperatingMode:
+    """Enum-like class for operating modes."""
+
+    WITHOUT_EXPERIMENT = "without experiment"
+    WITH_EXPERIMENT = "with experiment"
 
 
 class PybammPipeline:
@@ -27,6 +37,7 @@ class PybammPipeline:
         parameter_values: pybamm.ParameterValues | None = None,
         pybop_parameters: Parameters | None = None,
         solver: pybamm.BaseSolver | None = None,
+        experiment: pybamm.Experiment = None,
         t_start: np.number = 0.0,
         t_end: np.number = 1.0,
         t_interp: np.ndarray | None = None,
@@ -64,15 +75,15 @@ class PybammPipeline:
         self._pybop_parameters = pybop_parameters or Parameters([])
         self._parameter_names = self.pybop_parameters.keys()
         self._geometry = model.default_geometry
-        self._methods = model.default_spatial_methods
-        self._solver = pybamm.IDAKLUSolver() if solver is None else solver
+        self._var_pts = var_pts or model.default_var_pts
+        self._spatial_methods = model.default_spatial_methods  # allow user input
+        self._solver = model.default_solver if solver is None else solver
         self._t_start = np.float64(t_start)
         self._t_end = np.float64(t_end)
         self._t_interp = t_interp
         self._initial_state = initial_state
         self._built_initial_soc = None
-        self._var_pts = var_pts or model.default_var_pts
-        self._submesh_types = model.default_submesh_types
+        self._submesh_types = model.default_submesh_types  # allow user input
         self._built_model = self._model
         self.requires_rebuild = (
             build_on_eval
@@ -81,6 +92,43 @@ class PybammPipeline:
             if initial_state is not None
             else self._determine_rebuild()
         )
+        self._setup_operating_mode(experiment)
+
+        self._save_at_cycles = None
+        self.callbacks = None
+        self._starting_solution = None
+        self._calc_esoh = False
+        self.steps_to_built_models = None
+        self.steps_to_built_solvers = None
+        self.experiment_unique_steps_to_model = None
+        self.get_esoh_solver = lru_cache()(self._get_esoh_solver)
+
+    def _get_esoh_solver(self, calc_esoh):
+        if calc_esoh is False:
+            return None
+
+        return pybamm.lithium_ion.ElectrodeSOHSolver(
+            self._parameter_values, self._model.param, options=self._model.options
+        )
+
+    def _setup_operating_mode(self, experiment) -> None:
+        if experiment is not None:
+            self.operating_mode = OperatingMode.WITH_EXPERIMENT
+            self.experiment = self._process_experiment(experiment)
+        else:
+            self.operating_mode = OperatingMode.WITHOUT_EXPERIMENT
+            self._model.events = []  # Turn off events for non-experiment optimisation
+            self.experiment = None
+
+    def _process_experiment(self, experiment) -> pybamm.Experiment:
+        if isinstance(experiment, (str, pybamm.step.BaseStep)):
+            return pybamm.Experiment([experiment]).copy()
+        elif isinstance(experiment, list):
+            return pybamm.Experiment(experiment).copy()
+        elif isinstance(experiment, pybamm.Experiment):
+            return experiment.copy()
+        else:
+            raise TypeError("Invalid experiment type")
 
     def _determine_rebuild(self) -> bool:
         """
@@ -92,8 +140,8 @@ class PybammPipeline:
         geometry = deepcopy(self._geometry)
 
         # Apply "[input]"
[…]
