### 2025-06-04 922fcfc
refactor: python problem/builder for generalised usage
