+++ pybop/problems/python_problem.py
@@ -1,9 +1,9 @@
-from typing import Callable
+from collections.abc import Sequence
+from typing import Callable, Optional
 
 import numpy as np
 
 from pybop import Parameters
-from pybop.costs.base_cost import BaseCost
 from pybop.problems.base_problem import Problem
 
 
@@ -11,125 +11,150 @@ class PythonProblem(Problem):
     """
     Defines a problem that uses Python callables for simulation and cost evaluation.
 
-    This problem type allows for direct integration with custom Python functions
-    instead of relying on external simulation frameworks like PyBaMM.
+    Provides a flexible interface for custom Python functions instead of relying
+    on external simulation frameworks like PyBaMM. Designed to work with the Python
+    builder pattern for progressive model composition.
+
+    The problem supports either standard models OR models with sensitivities, but not both
+    simultaneously.
 
     Parameters
     ----------
-    model : Callable
-        The function that performs the simulation
-    model_with_sens : Callable, optional
-        Function that performs simulation and returns sensitivities
-    pybop_params : Parameters
-        Container for optimisation parameters
-    costs : list
-        Cost objects either Callable | BaseCost
-    cost_weights : list
-        Weights for each cost component
+    model : Sequence[Callable], optional
+        Sequence of functions that perform simulation. Each function should accept
+        parameter values and return a numeric result or dict of results.
+    model_with_sens : Sequence[Callable], optional
+        Sequence of functions that perform simulation and return sensitivities.
+        Each function should accept parameter values and return (results, gradients).
+    pybop_params : Parameters, optional
+        Container for optimisation parameters defining the parameter space.
+    weights : Sequence[float], optional
+        Weights for each callable component. Length must match the number of models.
+
+    Raises
+    ------
+    ValueError
+        If both model types are provided, if neither is provided, or if weights
+        length doesn't match model count.
+
+    Examples
+    --------
+    >>> def quadratic_model(params):
+    ...     return np.sum(params**2)
+    >>>
+    >>> problem = PythonProblem(
+    ...     model=[quadratic_model],
+    ...     pybop_params=my_params,
+    ...     weights=[1.0]
+    ... )
+    >>> result = problem.run()
+
+    Notes
+    -----
+    This class is typically instantiated via the Python builder class rather
+    than directly, which provides a more convenient fluent interface.
     """
 
     def __init__(
         self,
-        model: Callable,
-        model_with_sens: Callable = None,
-        pybop_params: Parameters = None,
-        costs: list[BaseCost] = None,
-        cost_weights: list[float] = None,
-        dataset: dict[str, np.ndarray] = None,
+        model: Optional[Sequence[Callable]] = None,
+        model_with_sens: Optional[Sequence[Callable]] = None,
+        pybop_params: Optional[Parameters] = None,
+        weights: Optional[Sequence[float]] = None,
     ):
         super().__init__(pybop_params=pybop_params)
-        self._model = model
-        self._model_with_sens = model_with_sens
-        self._costs = costs
-        self._cost_weights = cost_weights
-        self._dataset = dataset
+        self._model = tuple(model) if model is not None else None
+        self._model_with_sens = (
+            tuple(model_with_sens) if model_with_sens is not None else None
+        )
+        self._weights = weights
 
     def run(self) -> float:
         """
-        Evaluates the model and computes the weighted cost.
+        Execute all standard models with current parameters and return weighted sum.
+
+        This method evaluates each model function with the current parameter values,
+        applies the corresponding weights, and returns the sum.
 
         Returns
         -------
         float
-            The weighted sum of all cost components
+            Weighted sum of all model results
+
+        Raises
+        ------
+        RuntimeError
+            If no standard models are available (i.e., only sensitivity models exist)
         """
-        self.check_set_params_called()
 
[…]
