+++ pybop/builders/python.py
@@ -1,127 +1,127 @@
-from typing import Callable, Union
-
-import numpy as np
+from typing import Callable
 
 from pybop import PythonProblem
 from pybop.builders.base import BaseBuilder
-from pybop.costs.base_cost import BaseCost
 
 
 class Python(BaseBuilder):
     """
-    Builder for Python-based optimisation problems.
+    Builder for Python-based problems.
 
-    This builder allows for the creation of optimisation problems using custom
-    Python functions instead of specialised simulation frameworks.
+    This builder creates problems using custom Python functions instead of
+    specialised simulation frameworks. It supports both standard models and
+    models with sensitivity analysis.
 
+    If this problem is used by Pybop Optimisation or Sampling classes, the
+    functions will be minimised.
+
+    Examples
+    --------
+    >>> builder = Python()
+    >>> builder.add_func(my_model_function, weight=1.5)
+    >>> problem = builder.build()
     """
 
     def __init__(self):
         super().__init__()
-        self._model = None
-        self._model_with_sens = None
+        self._models: list[Callable] = []
+        self._models_with_sens: list[Callable] = []
+        self._weights: list[float] = []
 
-    def set_simulation(
-        self,
-        model: Callable = None,
-        model_with_sens: Callable = None,
-    ) -> None:
+    def add_func(self, model: Callable, weight: float = 1.0) -> "Python":
         """
-        Set the simulation functions for the problem.
+        Add a simulation function to the problem.
 
         Parameters
         ----------
         model : Callable
-            Function that takes parameters and dataset, returns simulation results
-            Expected signature: func(params: list[float] or float) -> dict
-        model_with_sens : Callable, optional
-            Function that returns both results and sensitivities
-            Expected signature: func(params: list[float] or float) -> tuple[dict, dict]
+            Function with signature: func(params: Union[List[float], float]) -> dict
+            Returns dict where keys match signal names and values are predictions.
+        weight : float, default=1.0
+            Weight for this model in multi-objective optimisation.
 
-        where the keys in the dicts match the signal names and the values are
-        list-like objects containing the predicted values.
+        Returns
+        -------
+        Python
+            Self for method chaining.
+
+        Raises
+        ------
+        TypeError
+            If model is not callable.
         """
-        if model is not None and not callable(model):
-            raise TypeError("The model must be a callable obj")
+        if not callable(model):
+            raise TypeError("Model must be callable")
 
-        if model_with_sens is not None and not callable(model_with_sens):
-            raise TypeError("The model must be a callable obj")
+        self._models.append(model)
+        self._weights.append(weight)
+        return self
 
-        self._model = model
-        self._model_with_sens = model_with_sens
-
-    def add_cost(self, cost: Union[BaseCost, Callable], weight: float = 1.0) -> None:
+    def add_func_with_sens(
+        self, model_with_sens: Callable, weight: float = 1.0
+    ) -> "Python":
         """
-        Add a cost component to the problem.
+        Add a simulation function with sensitivity analysis to the problem.
 
         Parameters
         ----------
-        cost : Union[BaseCost, Callable]
-            Cost function
-        weight : float, optional
-            Weight for this cost component, by default 1.0
+        model_with_sens : Callable
+            Function with signature: func(params: Union[List[float], float]) -> Tuple[dict, dict]
+            Returns (results, sensitivities) where both dicts have keys matching
+            signal names.
+        weight : float, default=1.0
+            Weight for this model in multi-objective optimisation.
+
+        Returns
+        -------
+        Python
+            Self for method chaining.
+
+        Raises
+        ------
+        TypeError
+            If model_with_sens is not callable.
         """
-        if isinstance(cost, BaseCost):
[…]
