### 2025-06-05 a0ab313
docstring update pybop/builders/base.py
Co-authored-by: Martin Robinson <martinjrobins@gmail.com>
