+++ pybop/samplers/base_pints_sampler.py
@@ -1,15 +1,12 @@
+from dataclasses import dataclass
 import logging
 import time
 from functools import partial
-from typing import Optional, Union
+from typing import Optional, Type, Union
 
 import numpy as np
-from pints import (
-    MultiSequentialEvaluator,
-    ParallelEvaluator,
-    SequentialEvaluator,
-    SingleChainMCMC,
-)
+import pints
+
 
 from pybop import (
     BaseCost,
@@ -18,6 +15,43 @@ from pybop import (
     MultiChainProcessor,
     SingleChainProcessor,
 )
+from pybop.problems.base_problem import Problem
+from pybop.samplers.base_sampler import SamplerOptions
+
+
+@dataclass
+class PintsSamplerOptions(SamplerOptions):
+    max_iterations: int = 500
+    n_workers: int = 1
+    chains_in_memory: bool = True
+    log_to_screen: bool = True
+    log_filename: Optional[str] = None
+    initial_phase_iterations: int = 250
+    parallel: bool = False
+    verbose: bool = False
+    warm_up: bool = True
+    chain_files: Optional[list[str]] = None
+    evaluation_files: Optional[list[str]] = None
+
+    def validate(self):
+        """
+        Validate the options.
+
+        Raises
+        ------
+        ValueError
+            If the options are invalid.
+        """
+        if self.warm_up < 0:
+            raise ValueError("Number of warm-up steps must be non-negative.")
+        if self.max_iterations < 1:
+            raise ValueError("Maximum number of iterations must be greater than 0.")
+        if self.n_workers < 1:
+            raise ValueError("Number of workers must be greater than 0.")
+        if self.initial_phase_iterations < 1:
+            raise ValueError(
+                "Number of initial phase iterations must be greater than 0."
+            )
 
 
 class BasePintsSampler(BaseSampler):
@@ -30,79 +64,47 @@ class BasePintsSampler(BaseSampler):
 
     Parameters
     ----------
-    log_pdf : pybop.LogPosterior or List[pybop.LogPosterior]
-        An object to be sampled, currently supports the pybop.LogPosterior class.
-    sampler : pybop.Sampler
-        The sampling algorithm to use.
-    chains : int
-        Number of chains to run concurrently. Each chain contains separate markov samples from
-        the log_pdf.
-    x0 : numpy.ndarray
-        Initial values of the parameters for the optimisation.
-    cov0 : list-like
-        Initial covariance for the chains in the parameters. Either a scalar value
-        (same for all coordinates) or an array with one entry per dimension.
-    kwargs
-        Additional keyword arguments.
+    problem: pybop.Problem
+        The problem representing the negative unnormalised posterior distribution.
+    sampler: pints.MCMCSampler
+        The PINTS sampler to be used for sampling.
+    options: Optional[PintsSamplerOptions]
+        Options for the sampler, by default None.
     """
 
     def __init__(
         self,
-        log_pdf: Union[LogPosterior, list[LogPosterior]],
-        sampler,
-        chains: int = 1,
-        warm_up=None,
-        x0=None,
-        cov0=0.1,
-        **kwargs,
+        problem: Problem,
+        sampler: Type[pints.SingleChainMCMC | pints.MultiChainMCMC],
+        options: Optional[PintsSamplerOptions] = None,
     ):
-        super().__init__(log_pdf, x0, chains, cov0)
-
-        # Set kwargs
-        self._max_iterations = kwargs.get("max_iterations", 500)
-        self._log_to_screen = kwargs.get("log_to_screen", True)
-        self._log_filename = kwargs.get("log_filename", None)
-        self._initial_phase_iterations = kwargs.get("initial_phase_iterations", 250)
-        self._chains_in_memory = kwargs.get("chains_in_memory", True)
-        self._chain_files = kwargs.get("chain_files", None)
[…]
