+++ pybop/samplers/base_sampler.py
@@ -1,76 +1,79 @@
-from typing import Union
+from dataclasses import dataclass
+from typing import Optional, Union
 
 import numpy as np
 from pints import ParallelEvaluator
 
-from pybop import CostInterface, LogPosterior
+from pybop.problems.base_problem import Problem
 
 
-class BaseSampler(CostInterface):
+@dataclass
+class SamplerOptions:
+    n_chains: int = 1
+    n_workers: int = 1
+    parallel: bool = False
+
+    def validate(self):
+        """
+        Validate the options.
+
+        Raises
+        ------
+        ValueError
+            If the options are invalid.
+        """
+        if self.n_chains < 1:
+            raise ValueError("Number of chains must be greater than 0.")
+        if self.n_workers < 1:
+            raise ValueError("Number of workers must be greater than 0.")
+
+
+class BaseSampler:
     """
     Base class for Monte Carlo samplers.
 
     Parameters
     ----------
-    log_pdf : pybop.LogPosterior or List[pybop.LogPosterior]
-        The posterior or PDF to be sampled.
-    chains : int
-        Number of chains to be used.
-    x0
-        List-like initial values of the parameters for Monte Carlo sampling.
-    cov0
-        The covariance matrix to be sampled.
-
-    Note: Samplers perform maximisation of the Posterior by default.
+    problem : Problem
+        The problem representing the negative unnormalised posterior distribution.
+    options : SamplerOptions, optional
+        Options for the sampler, by default SamplerOptions().
     """
 
     def __init__(
         self,
-        log_pdf: Union[LogPosterior, list[LogPosterior]],
-        x0,
-        chains: int,
-        cov0: Union[np.ndarray, float],
+        problem: Problem,
+        options: Optional[SamplerOptions] = None,
     ):
-        self._log_pdf = log_pdf
-        self._cov0 = cov0
+        self._problem = problem
+        self._options = options or SamplerOptions()
+        self._options.validate()
 
-        # Number of chains
-        self._n_chains = chains
-        if self._n_chains < 1:
-            raise ValueError("Number of chains must be greater than 0")
+        # Set parallelisation
+        self.set_parallel(self.options.parallel)
 
-        # Set up parameters based on log_pdf
-        if isinstance(log_pdf, LogPosterior):
-            self.parameters = log_pdf.parameters
-            self.n_parameters = log_pdf.n_parameters
-        elif isinstance(log_pdf, (list, np.ndarray)) and isinstance(
-            log_pdf[0], LogPosterior
-        ):
-            self.parameters = log_pdf[0].parameters
-            self.n_parameters = log_pdf[0].n_parameters
-        else:
-            raise ValueError(
-                "log_pdf must be a LogPosterior or List[LogPosterior]"
-            )  # TODO: Update for more general sampling
+        # Get initial conditions
+        x0 = problem.params.rvs(n_samples=self._options.n_chains, apply_transform=True)
+        if x0 is None:
+            raise ValueError("Initial parameter values could not be sampled.")
+        self._x0 = x0
+        self._cov0 = np.diag(problem.params.get_sigma0())
 
-        transformation = self.parameters.construct_transformation()
-        super().__init__(transformation=transformation)
+    @property
+    def x0(self) -> np.ndarray:
+        return self._x0
 
-        # Check initial conditions
-        if x0 is not None:
-            if len(x0) != self.n_parameters:
-                raise ValueError(
-                    "x0 must have the same number of parameters as log_pdf"
-                )
-            self.parameters.update(initial_values=x0)
+    @property
+    def cov0(self) -> np.ndarray:
+        return self._cov0
 
-        # Update x0 w/ transformation if applicable - reshape to align with chains
[…]
