### 2025-06-06 84ff298
refactor: updates to pybamm pipeline, builder
