### 2025-06-06 196bf46
CasadiSolver with gradients cases addressed
