+++ pybop/models/base_model.py
@@ -1,4 +1,5 @@
 import copy
+import warnings
 from dataclasses import dataclass
 from typing import Callable, Optional, Union
 
@@ -649,13 +650,22 @@ class BaseModel:
         ):
             raise ValueError("These parameter values are infeasible.")
 
+        # Check solver and adjust sensitivity calculation flag
+        if isinstance(self._solver, pybamm.CasadiSolver):
+            warnings.warn(
+                "Casadi solver does not support sensitivity analysis. Sensitivities will be disabled."
+                )
+            cal_sensitivity = False
+        else:
+            cal_sensitivity = True
+
         self._pybamm_solution = self._solver.solve(
             self._built_model,
             inputs=inputs,
             t_eval=[t_eval[0], t_eval[-1]]
             if isinstance(self._solver, IDAKLUSolver)
             else t_eval,
-            calculate_sensitivities=True,
+            calculate_sensitivities=cal_sensitivity,
             t_interp=t_eval if self._solver.supports_interp else None,
         )
 
