### 2025-06-07 3680b2c
feat: updates for design costs, sensitivities for pybamm v25.6, events
