### 2025-06-09 3ff1d51
fix: parameter types for BaseProblem, ErrorMeasure dimensionality, updates for tests
