+++ tests/unit/test_cost.py
@@ -30,7 +30,7 @@ class TestCosts:
     def one_parameter(self):
         return pybop.Parameter(
             "Negative electrode active material volume fraction",
-            prior=pybop.Gaussian(0.5, 0.01),
+            initial_value=0.6,
             bounds=[0.375, 0.625],
         )
 
@@ -89,10 +89,32 @@ class TestCosts:
 
         assert higher_cost > lower_cost
 
+    @pytest.mark.parametrize(
+        "pybamm_costs",
+        [
+            pybop.costs.pybamm.SumSquaredError,
+            pybop.costs.pybamm.NegativeGaussianLogLikelihood,
+        ],
+    )
+    def test_pybamm_costs_with_sigma(self, pybamm_costs, model, dataset, one_parameter):
+        builder = pybop.Pybamm()
+        builder.set_simulation(model)
+        builder.set_dataset(dataset)
+        builder.add_parameter(one_parameter)
+        builder.add_cost(pybamm_costs("Voltage [V]", "Voltage [V]"))
+        problem = builder.build()
+
+        problem.set_params(np.array([0.55, 0.01]))
+        higher_cost = problem.run()
+        problem.set_params(np.array([0.52, 0.01]))
+        lower_cost = problem.run()
+
+        assert higher_cost > lower_cost
+
     def test_multi_cost_weighting(self, model, dataset, one_parameter):
         # ToDo: Add test for design cost as well
         def problem(weights):
-            builder = pybop.Pybamm()
+            builder = pybop.builders.Pybamm()
             builder.set_simulation(model)
             builder.set_dataset(dataset)
             builder.add_parameter(one_parameter)
@@ -181,14 +203,18 @@ class TestCosts:
         ):
             pybop.costs.pybamm.Minkowski("Voltage [V]", "Voltage [V]", p=np.inf)
 
-    def test_sumofpower(self, problem):
+    def test_sumofpower(self):
         # Incorrect order
         with pytest.raises(
-            ValueError, match="The order of 'p' must be greater than 0."
+            ValueError,
+            match="The order of the Minkowski distance must be greater than 0.",
         ):
             pybop.costs.pybamm.Minkowski("Voltage [V]", "Voltage [V]", p=-1)
 
-        with pytest.raises(ValueError, match="p = np.inf is not yet supported."):
+        with pytest.raises(
+            ValueError,
+            match="For p = infinity, an implementation of the Chebyshev distance is required.",
+        ):
             pybop.costs.pybamm.Minkowski("Voltage [V]", "Voltage [V]", p=np.inf)
 
     @pytest.fixture
@@ -205,131 +231,131 @@ class TestCosts:
             }
         )
 
-    @pytest.mark.parametrize(
-        "cost_class",
-        [
-            pybop.MeanSquaredError,
-            pybop.RootMeanSquaredError,
-            pybop.MeanAbsoluteError,
-            pybop.SumSquaredError,
-            pybop.Minkowski,
-            pybop.SumOfPower,
-        ],
-    )
-    def test_error_weighting(
-        self, model, parameters, dataset, randomly_spaced_dataset, cost_class
-    ):
-        problem = pybop.FittingProblem(model, parameters, dataset)
-        cost = cost_class(problem, weighting=1.0)
-        x = [0.5]
-        e, de = cost(x, calculate_grad=True)
-
-        # Test that the equal weighting is the same as weighting by one
-        costE = cost_class(problem, weighting="equal")
-        eE, deE = costE(x, calculate_grad=True)
-        np.testing.assert_allclose(e, eE)
-        np.testing.assert_allclose(de, deE)
-
-        # Test that domain-based weighting also matches for evenly spaced data
[…]
