### 2025-06-09 b652adf
style: pre-commit fixes
