### 2025-06-09 7bdb4e7
fix: sensitivity analysis dimensions, updates tests
