+++ pyproject.toml
@@ -27,7 +27,7 @@ classifiers = [
 requires-python = ">=3.9, <3.13"
 dependencies = [
 #    "pybamm[jax]>=25.1.0",
-    "pybamm[jax] @ git+https://github.com/pybamm-team/PyBaMM@i5039-dts-generalise",
+    "pybamm[jax] @ git+https://github.com/pybamm-team/PyBaMM@develop",
     "numpy>=1.16, <2.0",
     "scipy>=1.3",
     "pints>=0.5",
