### 2025-06-10 6d57bbf
feat: updates to design costs + tests, fix for ScaledCost + test
