+++ tests/unit/test_cost.py
@@ -90,6 +90,51 @@ class TestCosts:
     @pytest.mark.parametrize(
         "pybamm_costs",
         [
+            pybop.costs.pybamm.GravimetricEnergyDensity,
+            pybop.costs.pybamm.GravimetricPowerDensity,
+            pybop.costs.pybamm.VolumetricEnergyDensity,
+            pybop.costs.pybamm.VolumetricPowerDensity,
+        ],
+    )
+    def test_pybamm_design_costs(self, pybamm_costs, model, dataset, one_parameter):
+        parameter_values = pybamm.ParameterValues("Chen2020")
+        parameter_values.update(
+            {
+                "Electrolyte density [kg.m-3]": pybamm.Parameter(
+                    "Separator density [kg.m-3]"
+                ),
+                "Negative electrode active material density [kg.m-3]": pybamm.Parameter(
+                    "Negative electrode density [kg.m-3]"
+                ),
+                "Negative electrode carbon-binder density [kg.m-3]": pybamm.Parameter(
+                    "Negative electrode density [kg.m-3]"
+                ),
+                "Positive electrode active material density [kg.m-3]": pybamm.Parameter(
+                    "Positive electrode density [kg.m-3]"
+                ),
+                "Positive electrode carbon-binder density [kg.m-3]": pybamm.Parameter(
+                    "Positive electrode density [kg.m-3]"
+                ),
+            },
+            check_already_exists=False,
+        )
+        builder = pybop.Pybamm()
+        builder.set_simulation(model, parameter_values=parameter_values)
+        builder.set_dataset(dataset)
+        builder.add_parameter(one_parameter)
+        builder.add_cost(pybamm_costs())
+        problem = builder.build()
+
+        problem.set_params(np.array([0.55]))
+        higher_cost = problem.run()
+        problem.set_params(np.array([0.52]))
+        lower_cost = problem.run()
+
+        assert higher_cost > lower_cost
+
+    @pytest.mark.parametrize(
+        "pybamm_costs",
+        [
             pybop.costs.pybamm.SumSquaredError,
             pybop.costs.pybamm.NegativeGaussianLogLikelihood,
         ],
@@ -109,8 +154,23 @@ class TestCosts:
 
         assert higher_cost > lower_cost
 
+    def test_pybamm_scaled_cost(self, model, dataset, one_parameter):
+        builder = pybop.Pybamm()
+        builder.set_simulation(model)
+        builder.set_dataset(dataset)
+        builder.add_parameter(one_parameter)
+        cost = pybop.costs.pybamm.SumOfPower("Voltage [V]", "Voltage [V]")
+        builder.add_cost(pybop.costs.pybamm.ScaledCost(cost))
+        problem = builder.build()
+
+        problem.set_params(np.array([0.55]))
+        higher_cost = problem.run()
+        problem.set_params(np.array([0.52]))
+        lower_cost = problem.run()
+
+        assert higher_cost > lower_cost
+
     def test_multi_cost_weighting(self, model, dataset, one_parameter):
-        # ToDo: Add test for design cost as well
         def problem(weights):
             builder = pybop.builders.Pybamm()
             builder.set_simulation(model)
@@ -162,34 +222,12 @@ class TestCosts:
         assert val > 0
         assert grad > 0
 
-        # # Test cost direction
-        # higher_cost = cost([0.55])
-        # lower_cost = cost([0.52])
-        # assert higher_cost > lower_cost or (
-        #     higher_cost == lower_cost and not np.isfinite(higher_cost)
-        # )
-        #
-        # e, de = cost([0.5], calculate_grad=True)
-        #
[…]
