### 2025-06-10 5078f63
docs: update cell_mass static method
