+++ tests/unit/test_problem.py
@@ -14,12 +14,8 @@ class TestProblem:
     pytestmark = pytest.mark.unit
 
     @pytest.fixture
-    def first_model(self):
+    def model(self):
         return pybamm.lithium_ion.SPM()
-
-    @pytest.fixture
-    def second_model(self):
-        return pybamm.lithium_ion.SPMe()
 
     @pytest.fixture
     def parameters(self):
@@ -55,10 +51,8 @@ class TestProblem:
         )
 
     @pytest.fixture
-    def dataset(self, first_model, parameter_values, experiment):
-        sim = pybamm.Simulation(
-            first_model, experiment, parameter_values=parameter_values
-        )
+    def dataset(self, model, parameter_values, experiment):
+        sim = pybamm.Simulation(model, experiment, parameter_values=parameter_values)
         sol = sim.solve()
         return pybop.Dataset(
             {
@@ -82,11 +76,11 @@ class TestProblem:
     def signal(self):
         return "Voltage [V]"
 
-    def test_builder(self, first_model, parameter_values, experiment, dataset):
+    def test_builder(self, model, parameter_values, experiment, dataset):
         builder = pybop.builders.Pybamm()
         builder.set_dataset(dataset)
         builder.set_simulation(
-            first_model,
+            model,
             parameter_values=parameter_values,
             solver=IDAKLUSolver(atol=1e-6, rtol=1e-6),
         )
@@ -100,7 +94,12 @@ class TestProblem:
                 "Positive electrode active material volume fraction", initial_value=0.6
             )
         )
-        builder.add_cost(pybop.PybammSumSquaredError("Voltage [V]", "Voltage [V]", 1.0))
+        builder.add_cost(
+            pybop.costs.pybamm.SumSquaredError("Voltage [V]", "Voltage [V]", 1.0)
+        )
+        builder.add_cost(
+            pybop.costs.pybamm.MeanAbsoluteError("Voltage [V]", "Voltage [V]")
+        )
         problem = builder.build()
 
         assert problem is not None
@@ -111,43 +110,25 @@ class TestProblem:
         assert (value1 - value2) / value1 > 1e-5
         problem.set_params(np.array([0.6, 0.6]))
         value1s, grad1s = problem.run_with_sensitivities()
-        assert grad1s.shape == (2, 1)
+        assert grad1s.shape == (2,)
         problem.set_params(np.array([0.7, 0.7]))
         value2s, grad2s = problem.run_with_sensitivities()
-        np.testing.assert_allclose(value1s, value1)
-        np.testing.assert_allclose(value2s, value2)
+        np.testing.assert_allclose(value1s, value1, atol=1e-5)
+        np.testing.assert_allclose(value2s, value2, atol=1e-5)
 
-    def test_builder_with_rebuild_params(
-        self, first_model, parameter_values, experiment, dataset
-    ):
+        # Test building twice
+        problem2 = builder.build()
+        assert problem2 is not None
+        assert problem2 != problem
+
+    def test_builder_likelihoods(self, model, parameter_values, experiment, dataset):
         builder = pybop.builders.Pybamm()
         builder.set_dataset(dataset)
         builder.set_simulation(
-            first_model,
+            model,
             parameter_values=parameter_values,
+            solver=IDAKLUSolver(atol=1e-7, rtol=1e-7),
         )
-        builder.add_parameter(
-            pybop.Parameter("Negative electrode thickness [m]", initial_value=1e-6)
-        )
-        builder.add_parameter(
-            pybop.Parameter("Positive particle radius [m]", initial_value=1e-5)
-        )
-        builder.add_cost(pybop.PybammSumSquaredError("Voltage [V]", "Voltage [V]"))
-        problem = builder.build()
-
-        assert problem is not None
-
[…]
