### 2025-06-10 fa763eb
docs: update builder function signatures
