### 2025-06-10 a4fecff
fix: pipeline property
