+++ pybop/problems/pybamm_problem.py
@@ -1,7 +1,6 @@
 from typing import Union
 
 import numpy as np
-import pybamm
 from pybamm import Solution
 
 from pybop import JointLogPrior, Parameters
@@ -131,5 +130,5 @@ class PybammProblem(Problem):
         return cost, weighted_sensitivity
 
     @property
-    def pipeline(self) -> pybamm.Pipeline:
+    def pipeline(self):
         return self._pipeline
