### 2025-06-23 #7874 @yuefan98
Hi @jgostick and @BradyPlanden. Here is my review of this work.

Overall, this is a nice work about battery parameterization, and I really like the design optimization perspective of this work, which can be seminal in the field. The following are some of my comments:

Comments need to be addressed to finish the review (pybop version 25.3 through pip install):
1. In the documentation, it is recommended to directly install the package through `pip install pybop`. It will be worth adding an `environment.yml` file and a workflow to create a new virtual environment to avoid potential Python version conflicts and provide better reproducibility in calculations
2. As a Mac user with an ARM-based CPU, I encountered an issue with importing `pybop`, which shows the following error. I have to reinstall jaxlib based on [this suggestion](https://stackoverflow.com/questions/68327863/importing-jax-fails-on-mac-with-m1-chip) to make it work. This could be an active investigation or please include a note in the documentation.
`RuntimeError: This version of jaxlib was built using AVX instructions, which your CPU and/or operating system do not support. You may be able work around this issue by building jaxlib from source.`
3. I encountered the following warning when running  `creating_a_model.ipynb`. The results are not reproducible with the default solver. Switching to `pybamm.CasadiSolver()` can resolve the issue. That said, I recommend that the authors rerun all examples in a new virtual environment, ensure everything works properly, and resolve any potential version conflicts.
`The default solver changed to IDAKLUSolver after the v25.4.0. release. You can swap back to the previous default by using `pybamm.CasadiSolver()` instead.`
4. In general, I found it hard to reproduce the exact result shown in the example notebook (i.e., completely different minimization landscape for `energy_based_electrode_design.ipynb`), though the optimization results are still close (it tends to have fewer function calls on my side, which could be a good thing in general). I wish this could be quantified with better environment control or clearer documentation of version dependencies.

Comments can be improved over time:
1. It will be great to render the examples in the documentation so that users can easily navigate through different examples.
2. I also recommend including some demonstrations of key functionalities in the get-started page. New users can easily get lost without a clear introduction to core functionalities.
