### 2025-06-24 #7874 @yuefan98
Hi @agriyakhetarpal,
I have tested it with M1 chip iMac (conda version 4.12.0) and M1 Pro chip MacBook Pro (conda version 23.1.0). Both of them have the same issue with Pybop. And I haven't been experiencing any issues with PyBaMM previously.
The environment is created using the following command.
`conda create --name pybop python=3.11`

pybop and other necessary packages are installed later with `pip install`.
