### 2025-08-29 #7874 @BradyPlanden
Hi @jgostick, @yuefan98,

Thank you for the thorough review and organisation — your feedback is greatly appreciated.

Over the past few months, we have undertaken a [large restructure](https://github.com/pybop-team/PyBOP/pull/792) of the PyBOP package to improve both maintainability and user functionality (including installation). This restructuring updates the package architecture significantly from what is currently presented in the JOSS article. I will update the article to reflect these changes within the coming week.

Regarding your specific comments:

**1. Virtual environment installation guidance**
Thank you for this suggestion. To avoid being overly prescriptive, we have intentionally chosen not to dictate a specific virtual environment package (e.g., `conda`, `uv`, `poetry`, `venv`). We expect most users will have a preferred method for managing virtual environments. That being said, we do [reference](https://github.com/pybop-team/PyBOP?tab=readme-ov-file#installation) the extended installation instructions for the PyBaMM package, which documents a simple virtual environment creation.

**2. JAX support**
In the newly restructured PyBOP, JAX support has been dropped. The updated architecture achieves our desired functionality without requiring JAX integration. I will remove this section from the article accordingly.

**3. Default solver warning**
This warning originates from the underlying PyBaMM package itself. PyBOP aims to surface warnings and errors from its dependencies where appropriate, so this warning is passed through. However, the warning does not affect performance or functionality, as the new default `IDAKLU` solver is thoroughly tested within PyBOP. Future PyBaMM releases will deprecate this warning.

**4. Reproducibility**
This variation could stem from two potential sources:
   - **Parameter initialisation**: When a `prior` is provided without an `initial_guess`, initial parameter candidates are drawn from a random distribution, naturally resulting in different optimisation trajectories.
   - **Outdated notebook figures**: The figures in the notebook may be out-of-date since they are not automatically regenerated for each release. An automated `nox` session to refresh figures during releases would be beneficial—for now, all notebooks will be updated in the next release due to the restructuring.

Thank you again for your valuable review and suggestions!

cc: @martinjrobins @nicolacourtier
