### 2025-11-25 #7874 @jgostick
Hey Team, an EiC will come by this repo shortly to make the acceptance official, but I'm satisfied.  Congrats!
